% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_icgc_metadata}
\alias{add_icgc_metadata}
\title{Layer on ICGC metadata from a supplemental table to fill in missing COO}
\usage{
add_icgc_metadata(incoming_metadata)
}
\arguments{
\item{incoming_metadata}{A metadata table (probably output from get_gambl_metadata)}
}
\value{

}
\description{
Layer on ICGC metadata from a supplemental table to fill in missing COO
}
