% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{annotate_hotspots}
\alias{annotate_hotspots}
\title{Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations}
\usage{
annotate_hotspots(
  mutation_maf,
  recurrence_min = 5,
  analysis_base = c("FL--DLBCL", "BL--DLBCL"),
  p_thresh = 0.05
)
}
\arguments{
\item{analysis_base}{}
}
\value{

}
\description{
Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations
}
\examples{
hot_ssms = annotate_hotspots(all_ssm)
hot_maf = read.maf(hot_ssms)
oncoplot(hot_maf,genes=c("MEF2B","TP53","MYD88"),additionalFeature = c("hot_spot",TRUE))
}
