% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_maf_triplet.R
\name{annotate_maf_triplet}
\alias{annotate_maf_triplet}
\title{Annotate MAF with triplet context}
\usage{
annotate_maf_triplet(
  maf,
  all_SNVs = TRUE,
  ref,
  alt,
  projection = "grch37",
  fastaPath
)
}
\arguments{
\item{maf}{MAF file (required columns: Reference_Allele, Tumor_Seq_Allele2, STRAND_VEP)}

\item{all_SNVs}{To give us all the triplet sequences of SNVs and not specifying any specific ref and alt alleles (default is TRUE)}

\item{ref}{Reference allele}

\item{alt}{Alternative allele}

\item{projection}{The genome build projection for the variants you are working with (default is grch37)}

\item{fastaPath}{Can be a path to a FASTA file}
}
\value{
A data frame with an extra column for triple sequence
}
\description{
Give triple sequence of mutated base with its adjacent bases (-1 and +1) based on the STRAND_VEP column of MAF
}
\details{
It gives the reference and alternative alleles and looks for the rows of the data frame based on these values for + strand genes and their
complement alleles rows for - strand genes, then it can look for the adjacent bases in that mutation position. Also, it can look for all the SNVs in
the MAF data frame and provide triple sequences for them.
}
\examples{
annotate_maf_triplet(maf, all_SNVs = FALSE, "C", "T")

}
