% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_ssm_blacklist}
\alias{annotate_ssm_blacklist}
\title{Annotate and auto-drop a MAF data frame with existing blacklists to remove variants that would be dropped during the merge process}
\usage{
annotate_ssm_blacklist(
  mutations_df,
  unix_group = "gambl",
  tool_name = "slms_3",
  tool_version = "1.0",
  annotator_name = "vcf2maf",
  annotator_version = "1.2",
  flavour = "",
  genome_build = "grch37",
  drop_threshold = 4,
  return_blacklist = FALSE
)
}
\arguments{
\item{flavour}{Set to "clustered" if you want to use the blacklist for the new and improved SLMS-3 outputs (otherwise leave empty)}

\item{genome_build}{The genome build projection for the variants you want (grch37 is the only one currently supported)}

\item{drop_threshold}{The minimum count from one of the blacklists to drop a variant}
}
\value{
A MAF format data frame with two new columns indicating the number of occurrences of each variant in the two blacklists
}
\description{
Annotate and auto-drop a MAF data frame with existing blacklists to remove variants that would be dropped during the merge process
}
\examples{
deblacklisted_maf_df = annotate_ssm_blacklist(original_maf_df)
}
