% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{append_to_table}
\alias{append_to_table}
\title{Housekeeping function to add results to a table}
\usage{
append_to_table(table_name, data_df)
}
\arguments{
\item{file_path}{}
}
\value{

}
\description{
Housekeeping function to add results to a table
}
