% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ashm_multi_rainbow_plot}
\alias{ashm_multi_rainbow_plot}
\title{Title}
\usage{
ashm_multi_rainbow_plot(
  regions_bed,
  regions_to_display,
  exclude_classifications,
  metadata,
  custom_colours,
  classification_column = "lymphgen",
  db = "gambl_test",
  table_name = "maf_slms3_hg19_icgc"
)
}
\arguments{
\item{regions_bed}{Bed file with chromosome coordinates, should contain columns chr, start, end, name (with these exact names)}

\item{regions_to_display}{Optional vector of names from default regions_bed to use}

\item{metadata}{A metadata file already subsetted and arranged on the order you want the samples vertically displayed}

\item{table_name}{}
}
\value{

}
\description{
Title
}
