% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ashm_rainbow_plot}
\alias{ashm_rainbow_plot}
\title{Make a rainbow plot of all mutations in a region, ordered and coloured by metadata}
\usage{
ashm_rainbow_plot(
  db = "gambl_test",
  table_name = "maf_slms3_hg19_icgc",
  mutations_maf,
  metadata,
  exclude_classifications,
  drop_unmutated = FALSE,
  classification_column,
  bed,
  region,
  custom_colours,
  hide_ids = TRUE
)
}
\arguments{
\item{db}{The GAMBL database name}

\item{table_name}{The table we are querying}

\item{mutations_maf}{A data frame containing mutations (MAF format) within a region of interest (i.e. use the get_ssm_by_region)}

\item{metadata}{should be a data frame with sample_id as a column that should match Tumor_Sample_Barcode in the database}

\item{classification_column}{The name of the metadata column to use for ordering and colouring samples}

\item{bed}{Optional data frame specifying the regions to annotate (required columns: start, end, name)}
}
\value{
ggplot2 object
}
\description{
Make a rainbow plot of all mutations in a region, ordered and coloured by metadata
}
\examples{
# basic usage
ashm_rainbow_plot(mutations_maf=get_ssm_by_region(region="chr8:128806578-128806992"),metadata=get_gambl_metadata())
# advanced usages
mybed = data.frame(start=c(128806578,128805652,128748315), end=c(128806992,128809822,128748880), name=c("TSS","enhancer","MYC-e1"))
ashm_rainbow_plot(mutations_maf=my_mutations,metadata=my_metadata,bed=mybed)
}
