% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assign_cn_to_ssm}
\alias{assign_cn_to_ssm}
\title{Annotate mutations with their copy number information}
\usage{
assign_cn_to_ssm(
  this_sample,
  tool_name = "battenberg",
  database_name = "gambl_test",
  table_name = "maf_slms3_hg19_icgc",
  coding_only = FALSE
)
}
\arguments{
\item{this_sample}{Sample ID of the sample you want to annotate}

\item{tool_name}{Copy number caller to use for annotation}

\item{coding_only}{}
}
\value{
A data frame (MAF-like format) with two extra columns:
log.ratio is the log ratio from the seg file (NA when no overlap was found)
CN is the rounded absolute copy number estimate of the region based on log.ratio (NA when no overlap was found)
}
\description{
Annotate mutations with their copy number information
}
