% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_mutation_frequency_sliding_windows}
\alias{calc_mutation_frequency_sliding_windows}
\title{Count the number of mutations in a sliding window across a region for all samples}
\usage{
calc_mutation_frequency_sliding_windows(
  this_region,
  chromosome,
  start_pos,
  end_pos,
  metadata,
  slide_by = 100,
  window_size = 1000,
  plot_type = "none",
  return_format = "long-simple",
  min_count_per_bin = 3,
  return_count = FALSE,
  drop_unmutated = FALSE,
  classification_column = "lymphgen"
)
}
\arguments{
\item{metadata}{}
}
\value{

}
\description{
Count the number of mutations in a sliding window across a region for all samples
}
