% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{cbioportal_create}
\alias{cbioportal_create}
\title{Create cBioPortal Study.}
\usage{
cbioportal_create(
  this_seqtype = "genome",
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  gambl_maf = "maf_slms3_hg19",
  gambl_icgc_maf = "maf_slms3_hg19_icgc",
  cancer_type = "mixed",
  overwrite = TRUE,
  check_study = TRUE,
  out_dir
)
}
\arguments{
\item{this_seqtype}{The seq type you want to generate a study for. Default is "genome".}

\item{short_name}{A concise name for your portal project. Default is "GAMBL".}

\item{human_friendly_name}{A slightly more verbose name for your project. Default is "GAMBL data".}

\item{project_name}{Unique ID for your project. Default is "gambl_all".}

\item{description}{A verbose description of your data set. This is what the study will be named when accessing it through cBioPortal. Default is "GAMBL data from genome".}

\item{gambl_maf}{MAF origin.}

\item{gambl_icgc_maf}{ICGC MAF origin.}

\item{cancer_type}{Cancer types included in study, default is "mixed".}

\item{overwrite}{Flag to specify that files should be overwritten if they exist. Default is TRUE.}

\item{check_study}{Boolean parameter that controls if the generated study should be checked for sample IDs in case lists, that are not described in the clinical file. Default is TRUE.}

\item{out_dir}{The full path to the base directory where the files are being created.}
}
\value{
Nothing. Rather, this function generates all files necessary for successfully importing a study into an active cBioPortal instance.
}
\description{
Wrapper function for creating a import-ready cBioPortal study.
}
\details{
This function internally calls \link{setup_study}, \link{setup_fusions}, \link{finalize_study} and \link{study_check} to generate all necessary files for importing a study into cBioPortal.
This function was developed to streamline this step and at the same time ensure that the study information and selected data type is consistent throughout the individual steps of generating a study.
In addition, the user can also control if the generated study should be checked for sample IDs in case lists that are not described in the clinical file.
This potentially will prevent an annoying error that prevents the study to be imported into the active cBioPortal instance, default is TRUE.
Fusions are also handled based on the selected seq type (\code{this_seqtype}).
}
\examples{
\dontrun{
#generate cBioPortal study for all GAMBL genome samples:
cbioportal_create()

#generate a cBioPortal study for all GAMVL capture samples:
cbioportal_create(this_seqtype = "capture", description = "GAMBL data from exomes")
}

}
