% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_config_value}
\alias{check_config_value}
\title{INTERNAL FUNCTION for checking the existance of a config value, not meant for out-of-package usage.}
\usage{
check_config_value(config_key)
}
\arguments{
\item{config_key}{key from config, prefixed with config::get()}
}
\value{
A string with the path to a file specified in the config or nothing (if config key is NULL).
}
\description{
INTERNAL FUNCTION for checking the existance of a config value, not meant for out-of-package usage.
}
