% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_for_missing_features}
\alias{check_for_missing_features}
\title{Check matrix against missing features.}
\usage{
check_for_missing_features(incoming_matrix, feature_set)
}
\arguments{
\item{incoming_matrix}{The incoming matrix to be checked for any missing features.}

\item{feature_set}{The feature set.}
}
\value{
matrix where column names correspond to all features from the provided set
}
\description{
Operate on the matrix supplied by user and check it for any features missing compared to the provided set.
}
