% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_gamblr_config}
\alias{check_gamblr_config}
\title{Check GAMBLR Config.}
\usage{
check_gamblr_config(
  compare_timestamps = FALSE,
  ssh_session,
  archive_mode = FALSE,
  force_backup = FALSE
)
}
\arguments{
\item{compare_timestamps}{Whether the function will compare timestamps on your local files to the remote copy. Only relevant if you are working remotely.}

\item{ssh_session}{The ssh_session object see get_ssh_session() for more information. Only relevant if you are working remotely.}

\item{archive_mode}{This is not currently working but the idea here is to keep a GSC archive of GAMBL in sync with the actively updated outputs}

\item{force_backup}{Boolean parameter set to FALSE per default.}
}
\description{
Check that the GAMBLR config you have loaded will work in your setup.
}
\details{
This function is mostly for remote GAMBLRs to ensure they keep their local mirror of the GAMBL data up-to-date.
}
\examples{
check_gamblr_config()

}
