% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_remote_configuration}
\alias{check_remote_configuration}
\title{Check for a remote session and automagically confirm setup will work properly}
\usage{
check_remote_configuration(auto_connect = FALSE)
}
\arguments{
\item{auto_connect}{Set to TRUE to ensure an ssh_session is created if absent}
}
\description{
This function determines if a user is working in GAMBLR remotely and, if so, will
check if their config is loaded properly and ssh_session is available.
}
