% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromosome_arms_grch37}
\alias{chromosome_arms_grch37}
\title{Chromosome Arms grch37.}
\format{
\subsection{\code{chromosome_arms_grch37}}{

A data frame with 48 rows and 4 columns.
\describe{
\item{chromosome}{Chromosomes without chr-prefix, 1:22, X and Y.}
\item{start}{Start coordinates for the specified chromosome arm.}
\item{end}{End coordinates for the specified chromosome arm.}
\item{arm}{Chromosome arm, either p or q.}
}
}
}
\usage{
chromosome_arms_grch37
}
\description{
A data frame with the chromosome arm coordinates in respect to grch37.
}
\keyword{datasets}
