% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromosome_arms_hg38}
\alias{chromosome_arms_hg38}
\title{Chromosome Arms hg38.}
\format{
\subsection{\code{chromosome_arms_hg38}}{

A data frame with 48 rows and 4 columns.
\describe{
\item{chromosome}{Chromosomes with chr-prefix, 1:22, X and Y.}
\item{start}{Start coordinates for the specified chromosome arm.}
\item{end}{End coordinates for the specified chromosome arm.}
\item{arm{Chromosome arm, either p or q.}
}
}
}
\usage{
chromosome_arms_hg38
}
\description{
A data frame with the chromosome arm coordinates in respect to hg38.
}
\keyword{datasets}
