% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckelly.R
\name{ckelly}
\alias{ckelly}
\title{Arbitage-like Kelly Criterion}
\usage{
ckelly(homeodds, awayodds, chomeodds, cawayodds,
  oddstype = default.GamblR.odds())
}
\arguments{
\item{homeodds}{A vector of odds for the first team for book A}

\item{awayodds}{A vector of odds for the second team for book A}

\item{chomeodds}{A vector of odds for the first team for book B}

\item{cawayodds}{A vector of odds for the second team for book B}

\item{oddstype}{Format of input odds (e.g. decimal, see Convert odds)}
}
\value{
A vector of bankroll fractions (for each individual bet)
}
\description{
Compare odds from multiple books with the Kelly Criterion
}
\details{
Uses the odds from book A to determine the win probabilites of team A or B and then gives the Kelly optimal bet for those same teams on book B. Useful for betting when trusted book updates quickly leaving other books with favourable odds.
}
