% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{classify_dlbcl}
\alias{classify_dlbcl}
\title{Classify DLBCLs according to genetic subgroups.}
\usage{
classify_dlbcl(
  these_samples_metadata,
  maf_data,
  seg_data,
  sv_data,
  this_seq_type = "genome",
  projection = "grch37",
  output = "both",
  method = "chapuy",
  adjust_ploidy = TRUE
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains sample_id column with ids for the samples to be classified.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. At least must contain the first 45 columns of standard MAF format.}

\item{seg_data}{The SEG data frame to be used for matrix assembling. Must be of standard SEG formatting, for example, as returned by get_sample_cn_segments.}

\item{sv_data}{The SV data frame to be used for matrix assembling. Must be of standard BEDPE formatting, for example, as returned by get_combined_sv.}

\item{this_seq_type}{The seq_type of the samples. Only used to retrerive data through GAMBLR when it is not provided. Defaults to genome.}

\item{projection}{The projection of the samples. Only used to retrerive data through GAMBLR when it is not provided. Defaults to grch37.}

\item{output}{The output to be returned after prediction is done. Can be one of predictions, matrix, or both. Defaults to both.}

\item{method}{Classification method. One of chapuy (used as default), lacy, or hmrn.}

\item{adjust_ploidy}{Whether to perform ploidy adjustment for the CNV data. Defaults to TRUE (recommended).}
}
\value{
data frame with classification, binary matrix used in classification, or both
}
\description{
Using the user-provided or GAMBLR-retrieved data, this function will assemble the matrix according to the approach of
Chapuy et al (2018) or Lacy et al (2020) classifiers. Since neither of this classifiers is publicly released, we have implemented a solution
that closely (> 92\% accuracy) recapitulates each of these systems. For the classifier of Chapuy et al, the constructed matrix will be
used to calculate class probability using the bundled feature weights obtained from our reproduction of the classifier. For the Lacy et al
classifier, the matrix will be used for prediction of random forest model, which is supplied with the GAMBLR package. Following the modification
of Lacy classifier described in Runge et al (PMID 33010029), specifying the method of this function as hmrn will also consider
truncating mutations in NOTCH1 for the separate N1 subgroup.
}
\examples{
test_meta <- get_gambl_metadata(case_set = "DLBCL-unembargoed")
predictions_chapuy <- classify_dlbcl(these_samples_metadata = test_meta, output = "predictions")
predictions_lacy <- classify_dlbcl(these_samples_metadata = test_meta, method = "lacy")
predictions_hmrn <- classify_dlbcl(these_samples_metadata = test_meta, method = "hmrn", output = "predictions")
matrix_and_predictions <- classify_dlbcl(these_samples_metadata = test_meta)

}
