% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{classify_dlbcl_chapuy}
\alias{classify_dlbcl_chapuy}
\title{Classify DLBCLs according to genetic subgroups of Chapuy et al.}
\usage{
classify_dlbcl_chapuy(
  these_samples_metadata,
  maf_data,
  seg_data,
  sv_data,
  projection = "grch37",
  output = "both"
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains sample_id column with ids for the samples to be classified.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. At least must contain the first 45 columns of standard MAF format.}

\item{seg_data}{The SEG data frame to be used for matrix assembling. Must be of standard SEG formatting, for example, as returned by get_sample_cn_segments.}

\item{sv_data}{The SV data frame to be used for matrix assembling. Must be of standard BEDPE formatting, for example, as returned by get_combined_sv.}

\item{projection}{The projection of the samples. Only used to retrerive data through GAMBLR when it is not provided. Defaults to grch37.}

\item{output}{The output to be returned after prediction is done. Can be one of predictions, matrix, or both. Defaults to both.}
}
\value{
data frame with classification, binary matrix used in classification, or both
}
\description{
Use the feature weights from NMF model to assemble the binary matrix and classify DLBCL tumors based on C0-C5 system of Chapuy et al
}
