% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{classify_dlbcl_lacy}
\alias{classify_dlbcl_lacy}
\title{Classify DLBCLs according to genetic subgroups of Lacy et al.}
\usage{
classify_dlbcl_lacy(
  these_samples_metadata,
  maf_data,
  seg_data,
  sv_data,
  projection = "grch37",
  output = "both",
  include_N1 = FALSE
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains sample_id column with ids for the samples to be classified.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. At least must contain the first 45 columns of standard MAF format.}

\item{seg_data}{The SEG data frame to be used for matrix assembling. Must be of standard SEG formatting, for example, as returned by get_sample_cn_segments.}

\item{sv_data}{The SV data frame to be used for matrix assembling. Must be of standard BEDPE formatting, for example, as returned by get_combined_sv.}

\item{projection}{The projection of the samples. Only used to retrerive data through GAMBLR when it is not provided. Defaults to grch37.}

\item{output}{The output to be returned after prediction is done. Can be one of predictions, matrix, or both. Defaults to both.}

\item{include_N1}{Whether to set samples with NOTCH1 truncating mutations to N1 group as described in Runge et al (2021). Defaults to FALSE.}
}
\value{
data frame with classification, binary matrix used in classification, or both
}
\description{
Use the random forest model to classify DLBCL tumors based on system of Lacy et al
}
