% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cnvKompare}
\alias{cnvKompare}
\title{This function will compare CNV data between samples with multiple time points. It can also handle same-sample comparison
between different CNV callers if sample ID is specified in unique fashion. For groups with more than 2 samples,
optionally the pairwise comparisons can be performed. The comparison is made based on the internally calculated score,
which reflects percentage of each cytoband covered by CNV (rounded to the nearest 5\%) and its absolute CN. Optionally,
the heatmap of cnvKompare scores can be returned. In addition, the function will return all concordant and discordant cytobands.
Finally, the time series plot of CNV log ratios will be returned for all lymphoma genes, with further functionality to subset
it to a panel of genes of interest.}
\usage{
cnvKompare(
  patient_id,
  sample_ids,
  this_seg,
  seg_path,
  genes_of_interest,
  projection = "grch37",
  ignore_cytoband_labels = c("acen", "gvar", "stalk"),
  max_overlap = 20,
  exclude_sex = FALSE,
  return_heatmap = TRUE,
  compare_pairwise = TRUE
)
}
\arguments{
\item{patient_id}{Specify patient_id to retrieve sample ids from GAMBL metadata.}

\item{sample_ids}{Optionally, specify sample ids for comparison.}

\item{this_seg}{Optional input data frame of seg file. Must adhere to seg format.}

\item{seg_path}{Optionally, specify the path to a local seg file. Must adhere to seg format.}

\item{genes_of_interest}{Provide specific genes to be displayed on the time-series plot.}

\item{projection}{Argument specifying the projection of seg file, which will determine coordinates of the cytobands. Default is grch37, but hg38 is also accepted.}

\item{ignore_cytoband_labels}{Cytobands to be ignored. By default, "acen", "gvar", "stalk" are excluded.}

\item{max_overlap}{For time-series plot, how many maximum overlapping points are allowed?}

\item{exclude_sex}{Boolean argument specifying whether to exclude sex chromosomes from calculation. Default is FALSE.}

\item{return_heatmap}{Boolean argument specifying whether to return a heatmap of cnvKompare scores. Default is TRUE.}

\item{compare_pairwise}{Boolean argument specifying whether to perform pairwise comparisons is there are more than 2 time points in the group. Default is TRUE.}

\item{pga_cutoff}{Minimum PGA for the sample to adjust ploidy. Default is 0.05 (5\%).}
}
\value{
A list of overall and pairwise percent concordance, concordant and discordant cytobands, comparison heatmap of cnvKompare socres, and time series ggplot object.
}
\description{
This function will compare CNV data between samples with multiple time points. It can also handle same-sample comparison
between different CNV callers if sample ID is specified in unique fashion. For groups with more than 2 samples,
optionally the pairwise comparisons can be performed. The comparison is made based on the internally calculated score,
which reflects percentage of each cytoband covered by CNV (rounded to the nearest 5\%) and its absolute CN. Optionally,
the heatmap of cnvKompare scores can be returned. In addition, the function will return all concordant and discordant cytobands.
Finally, the time series plot of CNV log ratios will be returned for all lymphoma genes, with further functionality to subset
it to a panel of genes of interest.
}
\examples{
cnvKompare(patient_id = "00-14595", genes_of_interest = c("EZH2", "TP53", "MYC", "CREBBP", "GNA13"))

}
