% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_ancestry}
\alias{collate_ancestry}
\title{INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.}
\usage{
collate_ancestry(sample_table, somalier_output)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{somalier_output}{Somalier ancestery.tsv}
}
\value{
Table.
}
\description{
INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
}
\examples{
table = collate_ancestry(sample_table = "my_sample_table.txt")

}
