% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_ashm_results}
\alias{collate_ashm_results}
\title{Determine the hypermutation status of a few genes}
\usage{
collate_ashm_results(sample_table)
}
\value{

}
\description{
Determine the hypermutation status of a few genes
}
