% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_curated_sv_results}
\alias{collate_curated_sv_results}
\title{Title}
\usage{
collate_curated_sv_results(sample_table)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column}

\item{path_to_files}{Full local (base) path to the home of GAMBL outputs}
}
\value{

}
\description{
Title
}
