% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_extra_metadata}
\alias{collate_extra_metadata}
\title{INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.}
\usage{
collate_extra_metadata(sample_table, file_path)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{file_path}{Path to extra emtadata.}
}
\value{
Table.
}
\description{
INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
}
\examples{
table = collate_extra_metadata(sample_table = "my_sample_table.txt")

}
