% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_lymphgen}
\alias{collate_lymphgen}
\title{Expand a sample_table (meta data) horizontally with different flavours of lymphgen data.}
\usage{
collate_lymphgen(sample_table)
}
\arguments{
\item{sample_table}{Input data frame with sample IDs in the firs column.}
}
\description{
Expand a sample_table (meta data) horizontally with different flavours of lymphgen data.
}
\examples{
metadata = get_gambl_metadata()
colalted_meta = collate_lymphgen(sample_table = metadata)

}
