% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_nfkbiz_results}
\alias{collate_nfkbiz_results}
\title{Determine which cases have NFKBIZ UTR mutations}
\usage{
collate_nfkbiz_results(sample_table)
}
\value{

}
\description{
Determine which cases have NFKBIZ UTR mutations
}
