% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_qc_results}
\alias{collate_qc_results}
\title{Collate qc metrics.}
\usage{
collate_qc_results(sample_table, seq_type_filter = "genome")
}
\arguments{
\item{sample_table}{Vector of sample IDs}

\item{seq_type_filter}{default is genome, capture is also available for unix_group icgc_dart.}
}
\value{
The sample table with additional columns.
}
\description{
Collate qc metrics.
}
\examples{
qc_metrics = collate_qc_results(sample_table = sample_table, seq_type_filter = "genome")

}
