% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_results}
\alias{collate_results}
\title{Title}
\usage{
collate_results(sample_table, write_to_file = FALSE)
}
\value{
A table keyed on biopsy_id that contains a bunch of per-sample results from GAMBL
}
\description{
Title
}
