% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_ssm_results}
\alias{collate_ssm_results}
\title{Compute some summary statistics based on SSM calls}
\usage{
collate_ssm_results(sample_table, from_flatfile = TRUE)
}
\arguments{
\item{sample_table}{}
}
\value{

}
\description{
Compute some summary statistics based on SSM calls
}
