% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_sv_results}
\alias{collate_sv_results}
\title{Determine and summarize which cases have specific oncogene SVs}
\usage{
collate_sv_results(
  sample_table,
  tool = "manta",
  oncogenes = c("MYC", "BCL2", "BCL6", "CCND1", "IRF4")
)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column}

\item{oncogenes}{}
}
\value{

}
\description{
Determine and summarize which cases have specific oncogene SVs
}
