% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{comp_report}
\alias{comp_report}
\title{Construct pdf with sample-level plots, using minimum of arguments}
\usage{
comp_report(this_sample, export_individual_plots = FALSE, out)
}
\arguments{
\item{this_sample}{Sample ID to be plotted in report.}

\item{export_individual_plots}{Boolean parameter, set to TRUE to export individual plots.}

\item{out}{Path to output folder.}
}
\value{
Nothing.
}
\description{
Construct pdf with sample-level plots, using minimum of arguments
}
\examples{
comp_report(this_sample = this_sample, out = "../../../newplots/", export_individual_plots = TRUE)

}
