% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_mutation_flavour}
\alias{compare_mutation_flavour}
\title{Get a MAF that is just the variants unique to one of two flavours of variant calls available}
\usage{
compare_mutation_flavour(
  these_sample_ids,
  flavour1 = "clustered",
  flavour2 = ""
)
}
\arguments{
\item{flavour2}{}
}
\value{

}
\description{
Get a MAF that is just the variants unique to one of two flavours of variant calls available
}
