% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{consolidate_lymphgen}
\alias{consolidate_lymphgen}
\title{Consolidate Lymphgen.}
\usage{
consolidate_lymphgen(sample_table, derived_data_path = "", verbose = TRUE)
}
\arguments{
\item{sample_table}{Input data frame with metadata.}

\item{derived_data_path}{Optional argument specifying the path to a folder with files following the pattern *lymphgen.txt.}

\item{verbose}{Default is TRUE.}
}
\value{
A data frame with a supplemented lymphGen column.
}
\description{
Replace the lymphgen column in the incoming metadata with classification for additional samples.
}
\details{
Supplement the "lymphgen" column of the metadata with classification for additional samples.
Expects at least to have columns "patient_id" to bind on, and "lymphgen" to supplement the data on.
}
\examples{
metadata = get_gambl_metadata()
GAMBLR::consolidate_lymphgen(metadata)

}
