% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertodds.R
\name{convertodds}
\alias{convertodds}
\title{Convert odds}
\usage{
convertodds(odds, from = default.GamblR.odds(), to = "decimal",
  round = TRUE)
}
\arguments{
\item{odds}{A vector of odds in the "from" format}

\item{from}{Odds to convert from}

\item{to}{Odds to convert to}

\item{round}{Round the result? (TRUE/FALSE)}
}
\value{
A vector of odds in the "to" format
}
\description{
Convert between odd types
}
\details{
Can convert between implied probability (improb), decimal, american, fractional, hongkong, indonesian, and malay odd types
}
