% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{count_ssm_by_region}
\alias{count_ssm_by_region}
\title{Count the variants in a region with a variety of filtering options}
\usage{
count_ssm_by_region(
  region,
  chromosome,
  start,
  end,
  these_samples_metadata,
  count_by,
  seq_type = "genome",
  ssh_session
)
}
\arguments{
\item{region}{}

\item{chromosome}{}

\item{start}{}

\item{end}{}

\item{these_samples_metadata}{}

\item{count_by}{Defaults to counting all variants. Specify 'sample_id' if you want to collapse and count only one per sample}

\item{seq_type}{}

\item{ssh_session}{}
}
\value{

}
\description{
Count the variants in a region with a variety of filtering options
}
