% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{custom_caselist}
\alias{custom_caselist}
\title{Custom cBioPortal case list.}
\usage{
custom_caselist(
  these_sample_ids,
  caselist_name,
  caselist_description,
  return_missing_samples = FALSE,
  dir
)
}
\arguments{
\item{these_sample_ids}{A vector of sample IDs to be subset into a case list. Required parameter.}

\item{caselist_name}{Name of the generated case list (name does not include the file format, this will be added automagically). This parameter is required.}

\item{caselist_description}{A verbose description of the created case list. Required.}

\item{return_missing_samples}{Boolean parameter. Set to TRUE to return all sample IDs that are in the desired case list, but not represented in the study specific clinical file. Default is FALSE.}

\item{dir}{The directory where all study specific files live.}
}
\description{
Create and a custom case list for easy data subset in cBioPortal.
}
\details{
Convenience function for specifying custom case lists that can be browsed on cBioPortal.
This function takes a set of sample IDs \code{these_sample_ids} and intersect the IDs with what's available in the study-specific clinical file.
This function also extracts the project name for the specified study, i.e the project name that is defined withing the folder specified under the \code{dir} parameter.
}
\examples{
\dontrun{
#get some sample IDs
my_samples = get_gambl_metadata() \%>\%
 dplyr::filter(pathology == "FL") \%>\%
 dplyr::filter(cohort == "FL_GenomeCanada") \%>\%
 pull(sample_id)

#create case list with selected sample IDs
custom_caselist(these_sample_ids = my_samples,
                caselist_name = "FL_Canada",
                caselist_description = "Follicular Lymphoma from the Genome Canada Study",
                dir = "../path/to/study_directory/")
}

}
