% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{determine_version}
\alias{determine_version}
\title{Check for a version of data to load}
\usage{
determine_version(
  mode = "somatic_hypermutation_locations",
  this_genome_build = "grch37"
)
}
\arguments{
\item{mode}{Determines which data to handle. Defaults to somatic_hypermutation_locations. Will grow with more options as more data is version tracked.}

\item{this_genome_build}{The genome build of the data if coordinate-based. Accepts grch37 (default) or hg38.}
}
\value{
data frame
}
\description{
This function determines if a user is requesting the latest version
of the bundled data or wants to access one of the earlier veresions.
}
