% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dhitsig_genes_with_weights}
\alias{dhitsig_genes_with_weights}
\title{Double Hit Signature Genes With Weights.}
\format{
\subsection{\code{dhitsig_genes_with_weights}}{

A data frame with 104 rows and 3 columns.
\describe{
\item{Ensembl_ID}{Ensembl IDs as factors, 104 levels.}
\item{ImportanceScore}{Numeric column with importance scores.}
\item{Hugo_Symbol}{Gene symbols in Hugo format as a factor with 104 levels.}
}
}
}
\usage{
dhitsig_genes_with_weights
}
\description{
A data frame with double hit signature genes (both as ensembl IDs and Hugo symbols) and importance scores.
}
\keyword{datasets}
