% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_fma_exp3.R
\name{Exp3Policy}
\alias{Exp3Policy}
\title{R6 Class representing the EXP3 policy.}
\description{
An \code{\link{Exp3Policy}} object implements the behaviour of
  an agent with a Exponential-weight policy for Exploration and Exploitation
  (EXP3). It is instantiated with the learning parameter eta. The method
  \code{\link{run}} processes a reward matrix corresponding to a binary
  stochastic with finitely many arms bandit.
}
\details{
The EXP3 algorithm is used in adversarial binary stochastic bandits
  with finitely many arms problems.

  When processing a reward matrix, a reward distribution is attributed to
  each arm and then sampled randomly at each round. The maximum of those samples
  determines which arm is played.

  This class uses a \code{\link{Recorder}} object to export results.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Short string describing the instance.}

\item{\code{record}}{R6 object of class Recorder used to export simulation
results.}

\item{\code{K}}{Integer containing the number of arms.}

\item{\code{eta}}{Learning rate.}

\item{\code{S_hat}}{Numeric vector holding distribution parameter of each arm.}

\item{\code{P}}{Numeric vector containing the sample of each arm.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Exp3Policy$new()}}
\item \href{#method-reset}{\code{Exp3Policy$reset()}}
\item \href{#method-compute_sampling_distribution}{\code{Exp3Policy$compute_sampling_distribution()}}
\item \href{#method-get_action}{\code{Exp3Policy$get_action()}}
\item \href{#method-update_with_reward}{\code{Exp3Policy$update_with_reward()}}
\item \href{#method-run}{\code{Exp3Policy$run()}}
\item \href{#method-clone}{\code{Exp3Policy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Exp3Policy object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exp3Policy$new(eta = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{eta}}{Learning rate.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Exp3Policy` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Set or reset object attributes before starting a simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exp3Policy$reset(K)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{Number of arms.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_sampling_distribution"></a>}}
\if{latex}{\out{\hypertarget{method-compute_sampling_distribution}{}}}
\subsection{Method \code{compute_sampling_distribution()}}{
Samples all arms reward distributions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exp3Policy$compute_sampling_distribution()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector containing each arm sample.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_action"></a>}}
\if{latex}{\out{\hypertarget{method-get_action}{}}}
\subsection{Method \code{get_action()}}{
Chooses which arm to pull by sampling the arms reward
  distribution then taking the max.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exp3Policy$get_action()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The index of the chosen arm.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_with_reward"></a>}}
\if{latex}{\out{\hypertarget{method-update_with_reward}{}}}
\subsection{Method \code{update_with_reward()}}{
Updates S_hat given reward r of arm a.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exp3Policy$update_with_reward(r, a)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{Reward.}

\item{\code{a}}{Arm.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Process a bandit problem represented by the reward matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exp3Policy$run(rewards, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rewards}}{Reward matrix. Time on rows, arms on columns.}

\item{\code{verbose}}{Logical. if FALSE, only action and reward history will be
returned. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble describing at each iteration, the distribution parameter
  of each arm, the sample of each arm, the action taken and the reward
  received. If verbose is FALSE, only the action and reward.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Exp3Policy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
