% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_multisamp_ideogram}
\alias{fancy_multisamp_ideogram}
\title{Generate ideograms for selected sample, visualizing copy number variation segments.}
\usage{
fancy_multisamp_ideogram(
  these_sample_ids,
  plot_title = "CN Segments Ideogram",
  plot_sub = "grch37",
  chr_anno_dist = 3,
  chr_select = paste0("chr", c(1:22)),
  coding_only = FALSE,
  from_flatfile = FALSE,
  use_augmented_maf = FALSE
)
}
\arguments{
\item{these_sample_ids}{Sample to be plotted (accepts 2, 3 or 4 samples).}

\item{plot_title}{Main title of plot.}

\item{plot_sub}{Subtitle of plot.}

\item{chr_anno_dist}{Optional parameter to adjust chromosome annotations, default value is 3, increase to adjust annotations to left.}

\item{chr_select}{Optional parameter to subset plot to specific chromosomes. Default value is chr1-22.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate ideograms for selected sample, visualizing copy number variation segments.
}
\examples{

two_samples = c("00-15201_tumorA", "00-15201_tumorB")
ideo_2_samp = fancy_multisamp_ideogram(these_sample_ids = two_samples,
                                       plot_title = "CN Segments Ideogram",
                                       plot_sub = "grch37",
                                       chr_anno_dist = 4,
                                       chr_select = paste0("chr", c(1:22)),
                                       coding_only = FALSE,
                                       from_flatfile = FALSE,
                                       use_augmented_maf = FALSE)


}
