% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_propcov_plot}
\alias{fancy_propcov_plot}
\title{Visualize proportional coverage (10X and 30X) for selected samples and add comparison group (optional).}
\usage{
fancy_propcov_plot(
  these_samples,
  metadata,
  filter_cohort,
  filter_pathology,
  comparison_samples,
  seq_type = "genome",
  plot_subtitle = ""
)
}
\arguments{
\item{these_samples}{Data frame with sample IDs (to be plotted) in the first column.}

\item{metadata}{Optional, user can provide a metadata df to subset sample IDs from.}

\item{filter_cohort}{Optional parameter to be used when these_sample is NULL. Calls get_gambl_metadata() and filters on the cohort supplied in this parameter.}

\item{filter_pathology}{Optional parameter to be used when these_sample is NULL. Calls get_gambl_metadata() and filters on the pathology supplied in this parameter.}

\item{comparison_samples}{Optional parameter, give the function a list of sample IDs to be compared against the main plotting group.}

\item{seq_type}{Selected seq type for incoming QC metrics.}

\item{plot_subtitle}{Plotting parameter, subtitle of generated plot.}
}
\value{
plot as ggplot object.
}
\description{
Visualize proportional coverage (10X and 30X) for selected samples and add comparison group (optional).
}
\examples{
#Example 1: Compare proportional coverage for all samples compared to all FL cases:
#prepare comparison data:
this_meta = get_gambl_metadata()
all_fl = dplyr::filter(this_meta, pathology == "FL") \%>\%
dplyr::select(sample_id) \%>\%
  arrange(sample_id) \%>\%
  as.data.frame()

sub_fl = all_fl[c(128:183),]
sub_fl = as.data.frame(sub_fl)
colnames(sub_fl)[1] = "sample_id"

plot:
my_plot = fancy_propcov_plot(seq_type = "genome", comparison_samples = sub_fl, plot_subtitle = "Proportional coverage for all samples compared to all FL cases")

Example 2: Visualize proportional coverage for all FL cases (no comparison group)
all_fl = fancy_propcov_plot(filter_pathology = "FL", seq_type = "genome", plot_subtitle = "Proportional coverage for all FL cases")

}
