% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_proportions_plot}
\alias{fancy_proportions_plot}
\title{Proportional Metrics Plot.}
\usage{
fancy_proportions_plot(
  these_sample_ids,
  metadata,
  these_samples_metadata,
  keep_cohort,
  keep_pathology,
  seq_type = "genome",
  plot_subtitle = ""
)
}
\arguments{
\item{these_sample_ids}{Data frame with sample IDs (to be plotted) in the first column.}

\item{metadata}{Optional, user can provide a metadata df to subset sample IDs from.}

\item{these_samples_metadata}{GAMBL metadata subset to the cases you want to process.}

\item{keep_cohort}{Optional parameter to be used when these_sample is NULL. Calls \link{get_gambl_metadata} and filters on the cohort supplied in this parameter.}

\item{keep_pathology}{Optional parameter to be used when these_sample is NULL. Calls \link{get_gambl_metadata} and filters on the pathology supplied in this parameter.}

\item{seq_type}{Selected seq type for incoming QC metrics.}

\item{plot_subtitle}{Plotting parameter, subtitle of generated plot.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
This function takes all the available proportional quality control metrics and a vector of sample IDs (plotted on the x-ais) and plots the vlaues along the y-axis.
This function provides straightforward subsetting parameters allowing for a straightforward execution.
Either provide a data frame with sample IDs in the first column to the \code{these_samples_ids} parameter.
Or, call one of the optional parameters for using an already subset metadata table (subset to the sample IDs of interest).
If \code{these_samples_ids} and \code{these_samples_metadata} is not provided, the user can subset al GAMBL samples on the fly with \code{keep_cohort} and/or \code{keep_pathology}.
This function also provides parameters for easy customization of the plot aesthetics. For example, the subtitle of the plot can easily be controlled with the \code{plot_subtitle} parameter.
For usage examples and more information, refer to the parameter descriptions and examples in the fancy vignette.
}
\details{
Visualize proportional metrics for selected samples.
}
\examples{
#Example 1 - using these_sample_ids parameter
#subset on FL cases with QC metrics available and plot
metadata = get_gambl_metadata() \%>\%
kridel_fl = dplyr::filter(metadata, pathology == "FL",
               cohort == "FL_Kridel")
kridel_fl_samples = dplyr::select(kridel_fl, sample_id)

fancy_proportions_plot(these_sample_ids = kridel_fl_samples)

#Example 2 - using already filtered metadata (these_samples_metadata)
fancy_proportions_plot(these_samples_metadata = kridel_fl)

#Example 3 - using in-house metadata filtering options
fancy_proportions_plot(keep_cohort = "FL_Kridel",
                   keep_pathology = "FL")

}
