% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_qc_alignment_plot}
\alias{fancy_qc_alignment_plot}
\title{Visualize (stacked barplot) genomic read-subsets across a selection of samples.}
\usage{
fancy_qc_alignment_plot(
  these_samples,
  metadata,
  comparison_data = NULL,
  seq_type = "genome",
  add_mean = TRUE,
  limit_cohort = NULL,
  limit_pathology = NULL,
  this_color_palette = c(TotalReads = "#3D405B", TotalUniquelyMapped = "#81B29A",
    TotalDuplicatedreads = "#E07A5F"),
  plot_sub = ""
)
}
\arguments{
\item{these_samples}{Data frame with sample IDs (to be plotted) in the first column.}

\item{metadata}{Optional argument, used to derive sample IDs if sample_table is Null.}

\item{comparison_data}{Optional argument for plotting mean alignment metrics. Default is plotting the mean for samples provided. This parameter takes a list of smaple IDs.}

\item{seq_type}{Subset qc metrics to a specific seq_type, default is genome.}

\item{add_mean}{Set to TRUE to superimpose mean values of plotted variables. Default is TRUE.}

\item{limit_cohort}{If no df with sample IDs is supplied (these_samples = NULL) the function calls get_gambl_metadata and subsets on selected cohort.}

\item{limit_pathology}{If no df with sample IDs is supplied (these_samples = NULL) the function calls get_gambl_metadata and subsets on selected pathology.}

\item{this_color_palette}{Optional parameter that holds the selected colours for the plotted bars.}

\item{plot_sub}{Optional parameter, add a subtitle to alignment metric plot.}
}
\value{
plot as ggplot object.
}
\description{
Visualize (stacked barplot) genomic read-subsets across a selection of samples.
}
\examples{
sample_table = data.frame(sample_id  = c("09-33003T", "05-18426T", "SP59282"))
fancy_qc_plot(these_samples = sample_table, seq_type = "genome")

}
