% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_snp_chrdistplot}
\alias{fancy_snp_chrdistplot}
\title{Generate a plot with SNP distribution per chromosome.}
\usage{
fancy_snp_chrdistplot(
  this_sample,
  plot_subtitle = "SNP Distribution Per Chromosome",
  chr_select = paste0("chr", c(1:22)),
  include_dnp = FALSE,
  coding_only = FALSE,
  from_flatfile = FALSE,
  use_augmented_maf = FALSE
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{include_dnp}{Optional argument for including DNPs. Default is FALSE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate a plot with SNP distribution per chromosome.
}
\examples{
snp_plot = fancy_snp_chrdistplot(this_sample = "HTMCP-01-06-00422-01A-01D")
snp_dnp_plot = fancy_snp_chrdistplot(this_sample = "HTMCP-01-06-00422-01A-01D", include_dnp = TRUE, plot_subtitle = "SNP + DNP Distribution Per Chromosome")

}
