% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_snv_chrdistplot}
\alias{fancy_snv_chrdistplot}
\title{Generate a plot with SNV distribution per chromosome.}
\usage{
fancy_snv_chrdistplot(
  this_sample,
  maf_data,
  maf_path = NULL,
  variant_type_col = 10,
  chromosome_col = 5,
  plot_title = paste0(this_sample),
  plot_subtitle = "SNV Distribution Per Chromosome",
  chr_select = paste0("chr", c(1:22)),
  include_dnp = FALSE,
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{include_dnp}{Optional argument for including DNPs. Default is FALSE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate a plot with SNV distribution per chromosome.
}
\examples{
snv_plot = fancy_snv_chrdistplot(this_sample = "HTMCP-01-06-00422-01A-01D")
snv_dnp_plot = fancy_snv_chrdistplot(this_sample = "HTMCP-01-06-00422-01A-01D", include_dnp = TRUE, plot_subtitle = "SNV + DNP Distribution Per Chromosome")

}
