% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_sv_size_plot}
\alias{fancy_sv_size_plot}
\title{Generate plot visualizing SV sizes retreived with get_combined_sv. Subset on variant type, filter on VAF, size etc.}
\usage{
fancy_sv_size_plot(
  this_sample,
  vaf_cutoff = 0,
  size_cutoff = 50,
  chr_select = paste0("chr", c(1:22)),
  plot_title = paste0(this_sample),
 
    plot_subtitle = paste0("SV sizes for Manta calls. Dashed line annotates mean variant size.\\nVAF cut off: ",
    vaf_cutoff, ", SV size cut off: ", size_cutoff),
  projection = "grch37"
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{vaf_cutoff}{Threshold for filtering variants on VAF (events with a VAF > cutoff will be retained).}

\item{size_cutoff}{Threshold for filtering variants on size, default is 50bp.}

\item{chr_select}{Optional argument for subsetting on selected chromosomes, default is all autosomes.}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{projection}{Genomic projection for SVs and circos plot. Accepted values are grch37 and hg38.}
}
\value{
Nothing.
}
\description{
Generate plot visualizing SV sizes retreived with get_combined_sv. Subset on variant type, filter on VAF, size etc.
}
\examples{
myplot = fancy_sv_size_plot(this_sample = "HTMCP-01-06-00422-01A-01D")
myplot2 = fancy_sv_size_plot(this_sample = "HTMCP-01-06-00422-01A-01D", size_cutoff = 0, chr_select = c("chr1", "chr2"))

}
