% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_svbar}
\alias{fancy_svbar}
\title{Generate a bar plot visualizing structural variant subtype distributions}
\usage{
fancy_svbar(
  this_sample,
  plot_title = paste0(this_sample),
  plot_subtitle = "Structural Variant Subtype Distribution",
  chr_select = paste0("chr", c(1:22)),
  variant_select = c("DEL", "INS"),
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{variant_select}{Subtypes of SVs to be incldued in plot, default is DEL and INS.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate a bar plot visualizing structural variant subtype distributions
}
\examples{
chr1_sv = fancy_svbar(this_sample = "HTMCP-01-06-00422-01A-01D", chr_select = c(1))
svs = fancy_svbar(this_sample = "HTMCP-01-06-00422-01A-01D")

}
