% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_v_count}
\alias{fancy_v_count}
\title{Generate a bar plot visualizing total variant (SSM or SVs) count for selected contigs.}
\usage{
fancy_v_count(
  this_sample,
  maf_data,
  maf_path = NULL,
  ssm = TRUE,
  projection = "grch37",
  min_vaf = 0,
  variant_type_col = 10,
  chromosome_col = 5,
  plot_title = paste0(this_sample),
  plot_subtitle = "Variant Count For Selected Contigs",
  chr_select = paste0("chr", c(1:22)),
  variant_select = c("DEL", "INS", "DUP"),
  snp_colours = c(SNP = "#2B9971", DNP = "#993F2B", TNP = "#A62656"),
  hide_legend = FALSE,
  coding_only = FALSE,
  log10_y = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{ssm}{Set to FALSE to get plotting data from get_combined_sv (SVs). Default value is TRUE (plots SSM retrieved from annotate_cn_by_ssm$maf).}

\item{projection}{Genome build for returned variants (only applicable for ssm = FALSE).}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Recommended: 0 (only applicable for ssm = FALSE).}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{variant_select}{Subtypes of SVs to be incldued in plot, default is DEL, INS and DUP.}

\item{snp_colours}{Optional vector with colours for SNPs (DNP and TNP).}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{log10_y}{Set to TRUE to force y axis to be in log10.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented}{Boolean statement if to use augmented maf, default is TRUE.}
}
\value{
plot as ggplot object.
}
\description{
Generate a bar plot visualizing total variant (SSM or SVs) count for selected contigs.
}
\examples{
chr1_sv = fancy_v_count(this_sample = "HTMCP-01-06-00422-01A-01D", chr_select = c(1))
svs = fancy_v_count(this_sample = "HTMCP-01-06-00422-01A-01D")

}
