% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{fetch_next_sv}
\alias{fetch_next_sv}
\title{Retrieve the nearest SV to a specified SV in a given patient}
\usage{
fetch_next_sv(
  db = "gambl_test",
  table_name = "bedpe_manta_hg19",
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  sv_data,
  with_chr_prefix = FALSE,
  in_from
)
}
\arguments{
\item{db}{The GAMBL database name}

\item{table_name}{The table we are querying}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned}

\item{min_score}{The lowest Manta somatic score for a SV to be returned}

\item{sv_data}{A single row of a data frame returned by a function such as get_manta_sv}

\item{with_chr_prefix}{Prepend all chromosome names with chr (required by some downstream analyses)}

\item{pair_status}{Use to restrict results (if desired) to matched or unmatched results (default is to return all)}
}
\value{
A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs
}
\description{
Retrieve the nearest SV to a specified SV in a given patient
}
