% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fetch_output_files}
\alias{fetch_output_files}
\title{Title}
\usage{
fetch_output_files(
  tool_name,
  base_path,
  results_dir = "99-outputs",
  seq_type = "genome",
  genome_build = "hg38",
  search_pattern = "cellularity_ploidy.txt"
)
}
\arguments{
\item{base_path}{Either the full or relative path to where all the results directories are for the tool e.g. "gambl/sequenza_current"}

\item{search_pattern}{}
}
\value{
A data frame with one row per file and sample IDs parsed from the file name along with other GAMBL wildcards
}
\description{
Title
}
