% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{finalize_study}
\alias{finalize_study}
\title{Finish setting up a new cBioPortal instance or updating an existing portal data set}
\usage{
finalize_study(
  short_name = "GAMBL",
  include_icgc_data = FALSE,
  human_friendly_name = "GAMBL data",
  project_name = "gambl_minus_icgc",
  description = "GAMBL data without ICGC",
  cancer_type = "mixed",
  out_dir,
  overwrite = FALSE,
  sample_ids
)
}
\arguments{
\item{short_name}{a concise name for your portal project}

\item{include_icgc_data}{whether or not you want ICGC and other external data included}

\item{human_friendly_name}{a slightly more verbose name for your project}

\item{project_name}{unique ID for your project}

\item{description}{A verbose description of your data set}

\item{out_dir}{The full path to the base directory where the files are being created}

\item{sample_ids}{A vector of all the sample_id that were included in any of the data files for cbioportal}
}
\value{
nothing
}
\description{
Finish setting up a new cBioPortal instance or updating an existing portal data set
}
