% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{findMotif}
\alias{findMotif}
\title{Annotate mutations target AID motif}
\usage{
findMotif(maf, motif = "WRCY", projection = "grch37", fastaPath)
}
\arguments{
\item{maf}{MAF data frame (required columns: Reference_Allele, Chromosome, Start_Position, End_Position)}

\item{motif}{The motif sequence (default is WRCY)}

\item{projection}{The genome build projection for the variants you are working with (default is grch37)}

\item{fastaPath}{Can be a path to a FASTA file}
}
\value{
A data frame with two extra columns (seq and motif).
}
\description{
Checks for the presence of mutations at a given AID motif
}
\details{
In positions that reference allele  "C" has been mutated, two nucleotides before and one nucleotide after it
will be captured. In positions that reference allele  "G" has been mutated, one nucleotide before and two nucleotides after it
will be captured. The other mutations will have "NO" for the sequence.
Then, the forward orientation and reverse compliment of the motif will be generated and their presence in aquired reference
allele "C" and reference allele "G" sequence will be checked, respectively. ("NO" for the rows that have different reference
alleles)
}
\examples{
findMotif(maf = maf,
          motif = "WRCY",
          projection = "hg38"
         )
}
