% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{focal_cn_plot}
\alias{focal_cn_plot}
\title{Generate a plot of all CN segments}
\usage{
focal_cn_plot(
  region,
  gene,
  these_samples_metadata,
  type = "gain",
  segment_size = 1,
  crop_segments = TRUE,
  sort_by_annotation = c("pathology"),
  crop_distance = 1e+08
)
}
\arguments{
\item{crop_distance}{}
}
\value{

}
\description{
Generate a plot of all CN segments
}
