% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FtestCNV}
\alias{FtestCNV}
\title{Using GISTIC2.0 outputs, perform Fisher's exact test to compare CNV frequencies between 2 groups.}
\usage{
FtestCNV(
  gistic_lesions,
  metadata,
  comparison,
  fdr.method = "fdr",
  fdr.cutoff = 0.1,
  text_size = 7,
  blacklisted_regions = NULL
)
}
\arguments{
\item{gistic_lesions}{Path to the GISTIC2.0 all_lesions output file.}

\item{metadata}{Data frame containing sample ids and column with annotated data for the 2 groups of interest. All other columns are ignored. Currently, function exits if asked to compare more than 2 groups.}

\item{comparison}{Specify column annotating groups of interest.}

\item{fdr.method}{FDR method to adjust p values. Uses p.adjust function, and therefore accepts its method for FDR ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). By default, this function uses "fdr".}

\item{fdr.cutoff}{Specify FDR significance cut-off. By default, this function uses 0.1.}

\item{text_size}{Size of the text on the forest plot of differentially enriched CNV. Default text-size is 7.}

\item{blacklisted_regions}{Optionally, specify any descriptors (value from column \code{Descriptor} of GISTIC2.0 all_lesions output file) to filter out before amy comparisons are done. It is possible to specify list of multiple descriptors, for example, c("3p12.3", "12p13.2"). Default is NULL.}
}
\value{
list of 3 objects:
DISTINCT - set of CNV identified as significantly different between 2 groups;
CNV.EVENTS - conveniently reformatted set of events in both groups that can be used for downstream analyses; and
GRAPH - forest plot visualization of distinct CNV between 2 groups.
Output can be accessed by index (e.g.\link{3}), or by name (e.g.\link{"GRAPH"})
}
\description{
Using GISTIC2.0 outputs, perform Fisher's exact test to compare CNV frequencies between 2 groups.
}
\examples{
# basic usage
FtestCNV(gistic_lesions = "path_to_GISTIC2.0_output/all_lesions.conf_{confidence_level}.txt", metadata = derived_data, comparison = "pathology")
# advanced usage
FtestCNV(gistic_lesions = "path_to_GISTIC2.0_output/all_lesions.conf_{confidence_level}.txt", metadata = derived_data,
comparison = "pathology", fdr.method = "bonferroni", fdr.cutoff = 0.05, blacklisted_regions = c("3p12.3", "12p13.2"))

}
