% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_blacklist}
\alias{gene_blacklist}
\title{Genes Blacklist.}
\format{
\subsection{\code{gene_blacklist}}{

A tibble with 291 rows.
\describe{
\item{Gene}{Genes symbols in Hugo format.}
}
}
}
\usage{
gene_blacklist
}
\description{
A tibble with gene symbols (Hugo) that falls within blacklisted regions of the genome.
}
\keyword{datasets}
