% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gene_to_region}
\alias{gene_to_region}
\title{Get regions (bed format) from genes.}
\usage{
gene_to_region(gene_symbol, ensembl_id, genome_build)
}
\arguments{
\item{gene_symbol}{Gene symbol (e.g BCL2).}

\item{ensembl_id}{Esemble ID (e.g ENSG00000171791).}

\item{genome_build}{Reference genome build (currently, only grch37 supported).}
}
\value{
Gene coordinates in the format "chr:start-end" format.
}
\description{
Get regions (bed format) from genes.
}
\examples{
bcl2_region = gene_to_region(gene_symbol = "BCL2", genome_build = "grch37")
bcl2_region = gene_to_region(ensembl_id = "ENSG00000171791", genome_build = "grch37")

}
