% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_cn_states}
\alias{get_cn_states}
\title{Get CN States.}
\usage{
get_cn_states(
  regions_list,
  regions_bed,
  region_names,
  these_samples_metadata,
  this_seq_type = "genome",
  all_cytobands = FALSE,
  use_cytoband_name = FALSE
)
}
\arguments{
\item{regions_list}{A vector of regions in the format chrom:start-end.}

\item{regions_bed}{A bed file with one row for each region you want to determine the CN state from.}

\item{region_names}{Subset CN states on specific regions (gene symbols e.g FCGR2B).}

\item{these_samples_metadata}{A metadata table to auto-subset the data to samples in that table before returning.}

\item{this_seq_type}{Seq type for returned CN segments. One of "genome" (default) or "capture".}

\item{all_cytobands}{Include all cytobands, default is set to FALSE. Currently only supports hg19.}

\item{use_cytoband_name}{Use cytoband names instead of region names, e.g p36.33.}
}
\value{
Copy number matrix.
}
\description{
Get a copy number matrix for all samples based on segmented data in the database.
}
\details{
This function returns CN states for the specified regions.
For how to specify regions, refer to the parameter descriptions and function examples.
}
\examples{
#basic usage, generic lymphoma gene list
cn_matrix = get_cn_states(regions_bed=grch37_lymphoma_genes_bed)
single_gene_cn = get_cn_states(regions_list=c(this_region), region_names = c("FCGR2B"))
capture_cn = get_cn_states(
 regions_list=c("8:128723128-128774067"),
 region_names = c("MYC"),
 this_seq_type = "capture"
)

}
