% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_combined_sv}
\alias{get_combined_sv}
\title{Retrieve Combined Manta and GRIDSS-derived SVs from a flatfile and filter}
\usage{
get_combined_sv(
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  this_sample_id,
  with_chr_prefix = FALSE,
  projection = "grch37"
)
}
\arguments{
\item{min_vaf}{The minimum tumour VAF for a SV to be returned}

\item{min_score}{The lowest Manta somatic score for a SV to be returned}

\item{this_sample_id}{}

\item{with_chr_prefix}{Prepend all chromosome names with chr (required by some downstream analyses)}
}
\value{
A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs
}
\description{
Retrieve Combined Manta and GRIDSS-derived SVs from a flatfile and filter
}
