% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_gambl_colours}
\alias{get_gambl_colours}
\title{Get GAMBL Colours.}
\usage{
get_gambl_colours(
  classification = "all",
  alpha = 1,
  as_list = FALSE,
  as_dataframe = FALSE,
  return_available = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{classification}{Optionally request only colours for pathology, lymphgen, mutation or copy_number.}

\item{alpha}{Alpha of plotted colours.}

\item{as_list}{Boolean parameter controling the format of the return. Default is FALSE.}

\item{as_dataframe}{Boolean parameter controling the format of the return. Default is FALSE.}

\item{return_available}{Set to TRUE for returning all avaialble colours. Default is FALSE.}

\item{verbose}{Default is FALSE}
}
\value{
A named vector of colour codes for lymphgen classes and pathology.
}
\description{
Get GAMBL colour schemes for annotating figures.
}
\details{
This function was designed to retreive specified GAMBL colour palettes.
By default, this function returns all the colours currently avaialble.
The user can easily specify for what classification to return colurs for with the \code{classificaion} parameter.
It is also possible to return any given colours in different formats.
To do so, refer to the Boolean arguments; \code{as_list} and \code{as_dataframe}.
For more information regarding the avaialble colours, refer to the utillities vignette.
}
\examples{
lymphgen_cols = get_gambl_colours("lymphgen")
#be sure to install ggsci from https://github.com/morinlab/ggsci
#install_github("morinlab/ggsci")

}
