% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_gambl_metadata}
\alias{get_gambl_metadata}
\title{Get GAMBL metadata}
\usage{
get_gambl_metadata(
  db = "gambl_test",
  seq_type_filter = "genome",
  tissue_status_filter = c("tumour")
)
}
\arguments{
\item{db}{The GAMBL database name}

\item{seq_type_filter}{Filtering criteria (default: all genomes)}

\item{tissue_status_filter}{Filtering criteria (default: only tumour genomes)}
}
\value{
A data frame with metadata for each biopsy in GAMBL
}
\description{
Get GAMBL metadata
}
\examples{
#basic usage
my_metadata = get_gambl_metadata()
}
