% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_gene_cn_and_expression}
\alias{get_gene_cn_and_expression}
\title{Get the copy number and expression for a single gene}
\usage{
get_gene_cn_and_expression(gene_symbol, ensembl_id)
}
\arguments{
\item{ensembl_gene_id}{}
}
\value{

}
\description{
Get the copy number and expression for a single gene
}
