% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_lymphgen}
\alias{get_lymphgen}
\title{Get Lymphgen.}
\usage{
get_lymphgen(
  these_samples_metadata,
  flavour,
  return_feature_matrix = FALSE,
  return_feature_annotation = FALSE,
  lymphgen_file,
  keep_all_rows = FALSE,
  keep_original_columns = FALSE
)
}
\arguments{
\item{these_samples_metadata}{A metadata table to auto-subset the data to samples in that table before returning.}

\item{flavour}{Lymphgen flavour.}

\item{return_feature_matrix}{Boolean parameter, default is FALSE.}

\item{return_feature_annotation}{Boolean parameter, default is FALSE.}

\item{lymphgen_file}{Path to lymphgen file.}

\item{keep_all_rows}{Boolean parameter, default is FALSE.}

\item{keep_original_columns}{Boolean parameter, default is FALSE.}
}
\value{
A data frame.
}
\description{
Get a specific flavour of LymphGen from the main GAMBL outputs.
}
\details{
Get a specific flavour of LymphGen from the main GAMBL outputs and tidy the composites.
Optionally return a matrix of features instead
}
\examples{
lymphgens = get_lymphgen(flavour = "no_cnvs.no_sv.with_A53")

}
