% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_manta_sv}
\alias{get_manta_sv}
\title{Get Manta SVs}
\usage{
get_manta_sv(
  these_sample_ids,
  these_samples_metadata,
  projection = "grch37",
  chromosome,
  qstart,
  qend,
  region,
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  pairing_status,
  from_flatfile = TRUE,
  verbose = TRUE,
  from_cache = TRUE,
  write_to_file = FALSE
)
}
\arguments{
\item{these_sample_ids}{A vector of multiple sample_id (or a single sample ID as a string) that you want results for.}

\item{these_samples_metadata}{A metadata table to auto-subset the data to samples in that table before returning.}

\item{projection}{The projection genome build. Default is grch37.}

\item{chromosome}{Optional, the chromosome you are restricting to (can be prefixed or not prefixed).}

\item{qstart}{Optional, query start coordinate of the range you are restricting to.}

\item{qend}{Optional, query end coordinate of the range you are restricting to.}

\item{region}{Optional, region formatted like chrX:1234-5678 (chromosome can be prefixed or not prefixed) instead of specifying chromosome, start and end separately.}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Default is 0.1.}

\item{min_score}{The lowest Manta somatic score for a SV to be returned. Default is 40.}

\item{pass}{If TRUE (default) only return SVs that are annotated with PASS in the FILTER column. Set to FALSE to keep all variants, regardless if they PASS the filters.}

\item{pairing_status}{Use to restrict results (if desired) to matched or unmatched results (default is to return all). This parameter takes the filtering condition as a string ("matched" or "unmatched").}

\item{from_flatfile}{Set to TRUE by default, FALSE is no longer supported (database).}

\item{verbose}{Set to FALSE to minimize the output to console. Default is TRUE. This parameter also dictates the verbose-ness of any helper function internally called inside the main function.}

\item{from_cache}{Boolean variable for using cached results, default is TRUE. If \code{write_to_file = TRUE}, this parameter auto-defaults to FALSE.}

\item{write_to_file}{Boolean statement that outputs bedpe file if TRUE, default is FALSE. Setting this to TRUE forces \code{from_cache = FALSE}.}
}
\value{
A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs.
}
\description{
Retrieve Manta SVs and filter.
}
\details{
Return Manta SVs with additional VCF information to allow for filtering of high-confidence variants.
To return SV calls for multiple samples, give \code{these_sample_ids} a vector of sample IDs, if only one sample is desired,
give this parameter one sample ID, as a string (or a vector of characters). The user can also call the \code{these_samples_metadata}
parameter to make use of an already subset metadata table. In this case, the returned calls will be restricted to the sample_ids
within that data frame. This function relies on a set of specific internal functions \link{id_ease} and \link{get_manta_sv_by_samples} (if \code{from_cache = FALSE}).
This function can also restrict the returned calls to any genomic regions specified within \code{chromosome}, \code{qstart}, \code{qend},
or the complete region specified under \code{region} (in chr:start-end format), note that chromosome can be either prefixed or not prefixed.
Useful filtering parameters are also available, use \code{min_vaf} to set the minimum tumour VAF for a SV to be returned and \code{min_score}
to set the lowest Manta somatic score for a SV to be returned. \code{pair_status} can be used to return variants from either matched or unmatched samples.
In addition, the user can chose to return all variants, even the ones not passing the filter criteria. To do so, set \code{pass = FALSE} (default is TRUE).
Is it adviseed to run this function with \code{from_cache = TRUE} (default) to read manta calls from a previous generated merge (cached result).
If set to FALSE in combination with \code{write_to_file = TRUE}, the function will generate new merged manta calls, if the data access restriction allows it.
Note, that if \code{write_to_file} is set to TRUE, the function autodefaults \code{from_cache = FALSE} to avoid nonsense parameter combinations.
Is this function not what you are looking for? Try one of the following, similar, functions;
\link{get_combined_sv}, \link{get_manta_sv_by_sample}, \link{get_manta_sv_by_samples}
}
\examples{
#lazily get every SV in the table with default quality filters
all_sv = get_manta_sv()

#get all SVs for a single sample
some_sv = get_manta_sv(these_sample_ids = "94-15772_tumorA")

#get the SVs in a region around MYC
myc_locus_sv = get_manta_sv(region = "8:128723128-128774067")

}
