% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_manta_sv_by_sample}
\alias{get_manta_sv_by_sample}
\title{Load the manta output for 1 sample from a bedpe file. This function requires 1 sample specified in the this_sample_id parameter, or a metadata subset to 1 sample (these_samples_metadata). For multiple samples, please refer to get_manta_sv_by_samples.}
\usage{
get_manta_sv_by_sample(
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  this_sample_id,
  these_samples_metadata,
  with_chr_prefix = FALSE,
  from_flatfile = TRUE,
  projection = "grch37"
)
}
\arguments{
\item{min_vaf}{The minimum tumour VAF for a SV to be returned.}

\item{min_score}{The lowest Manta somatic score for a SV to be returned.}

\item{pass}{If set to TRUE, include SVs that are annotated with PASS in FILTER column. Default is TRUE.}

\item{this_sample_id}{The single sample_id you want to obtain the result from.}

\item{these_samples_metadata}{A metadata table containing metadata for this_sample_id, or sample of interest. If no sample ID is provided with this_sample_id, the function will extract sample ID from these_samples_metadata.}

\item{with_chr_prefix}{Prepend all chromosome names with chr (required by some downstream analyses).}

\item{from_flatfile}{Set to TRUE by default.}

\item{projection}{The projection genome build. Currently only grch37 is supported.}
}
\value{
a data frame containing the Manta outputs from this_sample_id in a bedpe-like format with additional columns extracted from the VCF column.
}
\description{
Load the manta output for 1 sample from a bedpe file. This function requires 1 sample specified in the this_sample_id parameter, or a metadata subset to 1 sample (these_samples_metadata). For multiple samples, please refer to get_manta_sv_by_samples.
}
\examples{
#minimum parameters
my_bedpe_df = get_manta_sv_by_sample(this_sample_id = "00-14595_tumorA")

#extended parameter usage
my_metadata = get_gambl_metadata() \%>\% dplyr::filter(sample_id = "00-14595_tumorA")

my_bedpe_df = get_manta_sv_by_sample(these_samples_metadata = my_metadata,
                                     min_vaf = 0.05,
                                     min_score = 30,
                                     pass = FALSE,
                                     with_chr_prefix = TRUE)

}
