% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_mutation_frequency_bin_matrix}
\alias{get_mutation_frequency_bin_matrix}
\title{Count hypermutated bins and generate heatmaps/cluster the data}
\usage{
get_mutation_frequency_bin_matrix(
  regions,
  regions_df,
  these_samples_metadata,
  region_padding = 1000,
  metadataColumns = c("pathology"),
  sortByColumns = c("pathology"),
  customColourColumns = NULL,
  slide_by = 100,
  window_size = 500,
  min_count_per_bin = 3,
  min_bin_recurrence = 5,
  min_bin_patient = 0,
  cluster_rows_heatmap = FALSE,
  cluster_cols_heatmap = FALSE
)
}
\arguments{
\item{metadata}{}
}
\value{

}
\description{
Count hypermutated bins and generate heatmaps/cluster the data
}
