% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_samples_md5_hash}
\alias{get_samples_md5_hash}
\title{Generate an md5 hash for a set of samples to help ensure reproducibility}
\usage{
get_samples_md5_hash(
  these_samples_metadata,
  these_samples,
  sample_set_name,
  sample_sets_df
)
}
\arguments{
\item{these_samples_metadata}{Optionally provide a metadata table or any data frame with a column named sample_id that has been subset to the samples you're working with}

\item{these_samples}{Optionally provide a vector of sample_id you are working with}

\item{sample_set_name}{Optionally provide the name of a sample set in GAMBL and the function will load the samples from that set and provide the hash}

\item{sample_sets_df}{Optionally provide a data frame of the sample sets instead of relying on/loading the local file from the GAMBL repo}
}
\value{
The md5 hash of the ordered set of sample_id
}
\description{
Generate an md5 hash for a set of samples to help ensure reproducibility
}
