% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_regions}
\alias{get_ssm_by_regions}
\title{Title}
\usage{
get_ssm_by_regions(regions_list, regions_bed, streamlined = FALSE)
}
\arguments{
\item{streamlined}{}
}
\value{

}
\description{
Title
}
