% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{get_study_info}
\alias{get_study_info}
\title{Get Study Info.}
\usage{
get_study_info(dir, list_to_global = FALSE)
}
\arguments{
\item{dir}{The relative path to the study directory (expects to find meta_study.txt in this folder).}

\item{list_to_global}{Boolean parameter, if set to TRUE all study identifiers will be returned to the global environment. Default is FALSE.}
}
\value{
A list with study specific identifiers, or nothing (i.e list_to_global = TRUE).
}
\description{
Function for retrieving study specific identifiers.
}
\details{
This function takes one required parameter (\code{dir}).
This is the relative path to the main directory for the study of interest.
The function reads in the meta_study.txt file and extracts unique study identifiers.
This returns a list that holds all the identifiers.
The user can also return the study identifiers to the global environment.
To do so, set the \code{list_to_global = TRUE}.
}
\examples{
#return study identifiers as  list:
my_study_info= get_study_info(dir = "path/to/study/")

#return all identifiers to the global environment:
get_study_info(dir = "path/to/study/", list_to_global = TRUE)

}
