% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37_gene_coordinates}
\alias{grch37_gene_coordinates}
\title{grch37 Gene Coordinates.}
\format{
\subsection{\code{grch37_gene_coordinates}}{

A data frame with 63,763 rows and 6 columns.
\describe{
\item{ensembl_gene_id}{Ensembl gene ID}
\item{chromosome}{The chromosome that the gene is residing on}
\item{start}{The start coordinates for the gene}
\item{end}{The end coordinates for the gene}
\item{gene_name}{The gene name}
\item{hugo_symbol}{Gene symbol in Hugo format}
}
}
}
\usage{
grch37_gene_coordinates
}
\description{
All gene coordinates in respect to grch37.
}
\keyword{datasets}
