% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37_lymphoma_genes_bed}
\alias{grch37_lymphoma_genes_bed}
\title{Lymphoma Genes (grch37).}
\format{
\subsection{\code{grch37_lymphoma_genes_bed}}{

A data frame with 195 rows and 4 columns.
\describe{
\item{chromosome_name}{The chromosome for which the gene is residing on}
\item{start_position}{The start coordinate for the gene}
\item{end_position}{The end coordinate for the gene}
\item{hgnc_symbol}{Gene symbol in Hugo format}
}
}
}
\usage{
grch37_lymphoma_genes_bed
}
\description{
Lymphoma associated genes in respect to grch37.
}
\keyword{datasets}
