% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37_oncogene}
\alias{grch37_oncogene}
\title{Oncogenes in grch37 genome build.}
\format{
\subsection{\code{grch37_oncogene}}{

A data frame with 19 rows and 5 columns.
\describe{
\item{chrom}{Chromosomes without chr-prefix, 1:22.}
\item{start}{Start coordinate for the specified oncogene.}
\item{end}{End coordinate for the specified oncogene.}
\item{gene}{Lymphoma oncogene.}
\item{entrez}{ENTREZ ID for the specified oncogene.}
}
}
}
\usage{
grch37_oncogene
}
\description{
A data frame with the coordinates of lymphoma oncogenes relative to the grch37 genome build.
}
\keyword{datasets}
