% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch37_partners}
\alias{grch37_partners}
\title{grch37 Partner Genes.}
\format{
\subsection{\code{grch37_partners}}{

A data frame with 31 rows and 5 columns.
\describe{
\item{chrom}{The chromosome for which the gene is residing on}
\item{start}{The start coordinate for the gene}
\item{end}{The end coordinate for the gene}
\item{gene}{Gene symbol in Hugo format}
\item{entrez}{Entrez ID}
}
}
}
\usage{
grch37_partners
}
\description{
Translocation partners for oncogenes in with coordinates in respect to grch37.
}
\keyword{datasets}
