% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_mutation_frequency_bin.R
\name{heatmap_mutation_frequency_bin}
\alias{heatmap_mutation_frequency_bin}
\title{Heatmap of mutation counts across sliding windows for multiple regions.}
\usage{
heatmap_mutation_frequency_bin(
  regions = NULL,
  regions_bed = NULL,
  these_samples_metadata = NULL,
  these_sample_id = NULL,
  this_seq_type = c("genome", "capture"),
  maf_data,
  mut_freq_matrix,
  projection = "grch37",
  region_padding = 1000,
  drop_unmutated = FALSE,
  metadataColumns = c("pathology"),
  sortByColumns = c("pathology"),
  expressionColumns = NULL,
  orientation = "sample_rows",
  skip_regions,
  only_regions,
  customColours = NULL,
  backgroundColour = "grey90",
  slide_by = 100,
  window_size = 500,
  min_count_per_bin = 0,
  min_bin_recurrence = 5,
  min_mut_tumour = 0,
  region_fontsize = 8,
  cluster_rows_heatmap = FALSE,
  cluster_cols_heatmap = FALSE,
  show_gene_colours = FALSE,
  label_regions_by = "name",
  legend_row = 3,
  legend_col = 3,
  legend_direction = "horizontal",
  legendFontSize = 10,
  legend_side = "bottom",
  return_heatmap_obj = FALSE,
  from_indexed_flatfile = TRUE,
  mode = "slms-3"
)
}
\arguments{
\item{regions}{Named vector of regions in the format c(name1 = "chr:start-end", name2 = "chr:start-end"). If neither regions nor regions_bed is specified, the function will use GAMBLR aSHM region information.}

\item{regions_bed}{Data frame of regions with four columns (chrom, start, end, name).}

\item{these_samples_metadata}{Metadata with at least sample_id column. If not providing a maf data frame, seq_type is also required.}

\item{maf_data}{Optional maf data frame. Will be subset to rows where Tumor_Sample_Barcode matches provided sample IDs or metadata table. If not provided, maf data will be obtained with get_ssm_by_regions().}

\item{mut_freq_matrix}{Optional matrix of binned mutation frequencies generated outside of this function, usually by \link{calc_mutation_frequency_by_regions}.}

\item{projection}{Genome build the function will operate in. Ensure this matches your provided regions and maf data for correct chr prefix handling. Default grch37.}

\item{region_padding}{Amount to pad the start and end coordinates by. Default 1000}

\item{drop_unmutated}{Whether to drop bins with 0 mutations. If returning a matrix format, this will only drop bins with no mutations in any samples.}

\item{metadataColumns}{Mandatory character vector of metadata columns to use in heatmap annotation. Default c("pathology").}

\item{sortByColumns}{Mandatory character vector of metadata columns to order annotations by. Will be ordered by factor levels and sorted in the order specified. Default c("pathology").}

\item{expressionColumns}{Optional character vector of numeric metadata columns, usually gene expression, for heatmap annotation.}

\item{orientation}{Specify whether heatmap should have samples in rows ("sample_rows") or in columns ("sample_cols"). Default sample_rows.}

\item{skip_regions}{Optional character vector of genes to exclude from the default aSHM regions.}

\item{only_regions}{Optional character vector of genes to include from the default aSHM regions.}

\item{customColours}{Optional list of character vectors specifying colours for heatmap annotation with metadataColumns, e.g. list(pathology = c(DLBCL = "green", BL = "purple")). If left blank, the function will attempt to match heatmap annotations with existing colours from \link{get_gambl_colours}, or will default to the Blood colour palette.}

\item{backgroundColour}{Optionally specify the colour for heatmap bins with 0 mutations. Default grey90.}

\item{slide_by}{Slide size for sliding window. Default 100.}

\item{window_size}{Size of sliding window. Default 500.}

\item{min_count_per_bin}{Specify the minimum number of mutations per bin to be included in the heatmap. Only bins with all samples falling below this threshold will be dropped. Default 0.}

\item{min_bin_recurrence}{Specify how many samples a bin must be mutated in to be displayed. Default 5.}

\item{min_mut_tumour}{Specify how many bins a tumour must be mutated in to be displayed. Default 0.}

\item{region_fontsize}{Fontsize of region labels on the heatmap. Default 8.}

\item{cluster_rows_heatmap}{Boolean. Default FALSE.}

\item{cluster_cols_heatmap}{Boolean.  Default FALSE.}

\item{show_gene_colours}{Boolean. Whether to add heatmap annotation colours for each region. Default FALSE.}

\item{label_regions_by}{Specify which feature of the regions to label the heatmap with. Heatmap will be split according to this value, and ordered by factor levels if the specified column is a factor. Default name.}

\item{legend_row}{Control aesthetics of the heatmap legend. Default 3.}

\item{legend_col}{Control aesthetics of the heatmap legend. Default 3.}

\item{legend_direction}{Control aesthetics of the heatmap legend. Default "horizontal".}

\item{legendFontSize}{Control aesthetics of the heatmap legend. Default 10.}

\item{legend_side}{Control aesthetics of the heatmap legend. Default "bottom".}

\item{return_heatmap_obj}{Boolean. FALSE will plot the heatmap automatically. TRUE will return a heatmap object to allow further tweaking with the draw() function. Default FALSE.}

\item{from_indexed_flatfile}{Set to TRUE to avoid using the database and instead rely on flat files (only works for streamlined data, not full MAF details).}

\item{mode}{Only works with indexed flat files. Accepts 2 options of "slms-3" and "strelka2" to indicate which variant caller to use. Default is "slms-3".}

\item{these_sample_ids}{Vector of sample IDs. Metadata will be subset to sample IDs present in this vector.}
}
\value{
A table of mutation counts for sliding windows across one or more regions. May be long or wide.
}
\description{
Obtain a heatmap of mutation counts across sliding windows for multiple regions.
}
\details{
This function takes a metadata table with \code{these_samples_metadata} parameter and internally calls \link{calc_mutation_frequency_bin_region} (that internally calls \link{get_ssm_by_regions}).
to retrieve mutation counts for sliding windows across one or more regions and generate a heatmap. May optionally provide any combination of a maf data frame, existing metadata, or a regions data frame or named vector.
}
\examples{
#load metadata.
metadata = get_gambl_metadata()
dlbcl_bl_meta = dplyr::filter(metadata, pathology \%in\% c("DLBCL", "BL"))

#bring together all derived sample-level results from many GAMBL pipelines.
dlbcl_bl_meta = collate_results(join_with_full_metadata = TRUE,
                                these_samples_metadata = dlbcl_bl_meta)

#get ashm regions
some_regions = grch37_ashm_regions

mut_count_matrix <- calc_mutation_frequency_bin_by_regions(
   these_samples_metadata = dlbcl_bl_meta,
   regions_bed = some_regions
)

}
