% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedge.R
\name{hedge}
\alias{hedge}
\title{Hedge Calculator}
\usage{
hedge(initialbet, initialodds, hedgeodds, oddstype = default.GamblR.odds())
}
\arguments{
\item{initialbet}{Initial bet amount}

\item{initialodds}{Odds that initial bet was made at}

\item{hedgeodds}{Odds of the prospective hedge (this may be able to take any length of input?)}

\item{oddstype}{Format of input odds (e.g. decimal, see Convert odds)}
}
\value{
A list containing the hedge amount, the amount won if the initial bet wins, and the amount won if the hedge bet wins
}
\description{
Determine the best hedge bet to make
}
