% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_ashm_regions}
\alias{hg38_ashm_regions}
\title{hg38 ASHM Regions.}
\format{
\subsection{\code{hg38_ashm_regions}}{

A data frame with 88 rows and 7 columns.
\describe{
\item{chr_name}{The chromosome for which the region is residing on}
\item{hg19_start}{start coordinate for the region}
\item{hg19_end}{end coordinate for the region}
\item{gene}{Gene symbol (Hugo)}
\item{region}{Region name}
\item{regulatory_comment}{Regulatory element}
\item{name}{gene-region format}
}
}
}
\usage{
hg38_ashm_regions
}
\description{
ASHM regions in respect to hg38.
}
\keyword{datasets}
