% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_partners}
\alias{hg38_partners}
\title{hg38 Partner Genes.}
\format{
\subsection{\code{hg38_partners}}{

A data frame with 31 rows and 5 columns.
\describe{
\item{chrom}{The chromsome for which the gene is residing on}
\item{start}{The start coordinate for the gene}
\item{end}{The end coordinate for the gene}
\item{gene}{Gene symbol in Hugo format}
\item{entrez}{Entrez ID}
}
}
}
\usage{
hg38_partners
}
\description{
Translocation partners for oncogenes in with coordinates in respect to hg38.
}
\keyword{datasets}
