% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotspot_regions_grch37}
\alias{hotspot_regions_grch37}
\title{grch37 Hotspot Regions.}
\format{
\subsection{\code{hotspot_regions_grch37}}{

A data frame with 6 rows and 3 columns.
\describe{
\item{chrom}{Chromosome for the described region}
\item{start}{The start coordinate for the region}
\item{end}{The end coordinate for the region}
}
}
}
\usage{
hotspot_regions_grch37
}
\description{
Mutation hotspot regions in respect to grch37.
}
\keyword{datasets}
