% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{id_ease}
\alias{id_ease}
\alias{id_ease,}
\alias{id}
\alias{ease}
\title{ID Ease}
\usage{
id_ease(
  these_samples_metadata,
  these_sample_ids,
  this_seq_type = "genome",
  return_this = "sample_id",
  verbose = TRUE
)
}
\arguments{
\item{these_samples_metadata}{A data frame with metadata, subset to sample IDs of interest.
If not provided will retrieve GAMBL metadata for all available samples.}

\item{these_sample_ids}{Sample IDs as a character of vectors.}

\item{this_seq_type}{The seq type of interest. Default is genome.}

\item{return_this}{Let's the user control what they want back, i.e sample IDs as a vector of characters or a metadata object as a data frame.
Default is "sample_id". The other acceptable value for this a parameter is "metadata".
The desired return might depend on the required input for any given GAMBLR function, further downstream (i.e some functions expect a vector of sample IDs, other a metadata table)}

\item{verbose}{Set to FALSE to limit the information that gets printed to the console. Default is TRUE.}
}
\value{
A character of vectors with Sample IDs (default) or a data frame with metadata (see \code{id_ease}).
}
\description{
Convenience function that standardize the way GAMBLR functions deals with sample IDs (these_sample_ids)
and metadata (these_samples_metadata).
}
\details{
This function can take sample IDs as a vector of characters, or a metadata table in data frame format.
Based on what is given to \code{return_this}, the function can return either a vector of characters (sample ids) or a data frame (metadata).
If no sample IDs are provided to the function, the function will operate on all gambl sample IDs available for the given seq type.
It is highly recommended to run this function with \code{verbose = TRUE} (default).
Since this will not only improve the overall logic on how the function operates.
But also might help with debugging functions that are internally calling this function.
The function also performs sanity checks and notifies the user if any of the requested sample IDs are not found in the metadata.
In addition, the function also notifies the dimensions of the returned object, providing further insight to what is returned.
}
\examples{
#give the function nothing (i.e return all sample IDs in the metadata for the default seq type)
this_is_wrong = id_ease()

#return metadata for all samples in the default seq type
all_meta = id_ease(return_this = "metadata")

#return metadata based on a sample ID
sample_meta = id_ease(these_sample_ids = "94-15772_tumorA", 
                      return_this = "metadata")

#return sample IDs based on an already filtered metadata
this_metadata = get_gambl_metadata(seq_type_filter = "genome") \%>\% 
  head(5)

thes_ids = id_ease(these_samples_metadata = this_metadata)

}
