% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{intersect_maf}
\alias{intersect_maf}
\title{perform set operations on two MAFs}
\usage{
intersect_maf(maf1, maf2, set_returned = "maf1_only")
}
\arguments{
\item{set_returned}{}
}
\value{

}
\description{
perform set operations on two MAFs
}
