% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kelly.R
\name{kelly}
\alias{kelly}
\title{Kelly Criterion}
\usage{
kelly(p, odds, oddstype = default.GamblR.odds(), fraction = 1)
}
\arguments{
\item{p}{A vector of event probabilites}

\item{odds}{A vector of event odds}

\item{oddstype}{Format of input odds (e.g. decimal, see Convert odds)}

\item{fraction}{Bet fraction (for fractional Kelly)}
}
\value{
A vector of bankroll fractions (for each individual bet)
}
\description{
Kelly Criterion calculation
}
