% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kellymulti.R
\name{kellymulti}
\alias{kellymulti}
\title{Kelly-multi}
\usage{
kellymulti(pmatrix, oddsmatrix, oddstype = default.GamblR.odds(),
  fraction = 1)
}
\arguments{
\item{pmatrix}{A matrix of event probabilities}

\item{oddsmatrix}{A matrix of event odds}

\item{oddstype}{Format of input odds (e.g. decimal, see Convert odds)}

\item{fraction}{Bet fraction (for fractional Kelly)}
}
\value{
A matrix containing bet fractions for each event
}
\description{
The workhorse for kellymulti2
}
