% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kellymulti2.R
\name{kellymulti2}
\alias{kellymulti2}
\title{Kelly Criterion for multiple simultaneous events/games}
\usage{
kellymulti2(probhome, oddhome, oddaway, probtie = NULL, oddtie = NULL,
  homenames = NULL, awaynames = NULL, oddstype = default.GamblR.odds(),
  fraction = 1, ou = FALSE)
}
\arguments{
\item{probhome}{Probability for the home teams to win}

\item{oddhome}{Odds for the home teams}

\item{oddaway}{Odds for the away teams}

\item{probtie}{Probability of ties (optional)}

\item{oddtie}{Odds for ties (optional)}

\item{homenames}{Names of the home teams}

\item{awaynames}{Names of the away teams}

\item{oddstype}{Format of input odds (e.g. decimal, see Convert odds)}

\item{fraction}{Bet fraction (for fractional Kelly)}

\item{ou}{Over/under bet? (TRUE/FALSE)}
}
\value{
A dataframe containing the matchup names along with the fractonal bet amount for the simultaneous Kelly Criterion
}
\description{
Determine the bet fractions for multiple (uncorrelated) games/event occuring at the same time
}
