% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes}
\alias{lymphoma_genes}
\title{Lymphoma Genes}
\format{
\subsection{\code{lymphoma_genes}}{

A data frame with 196 rows and 12 columns.
\describe{
\item{Gene}{Gene symbol}
\item{DLBCL}{Boolean flag annotating id the described genes are significant for the pathology (DLBCL)}
\item{FL}{Boolean flag annotating id the described genes are significant for the pathology (FL)}
\item{BL}{Boolean flag annotating id the described genes are significant for the pathology (BL)}
\item{MCL}{Boolean flag annotating id the described genes are significant for the pathology (MCL)}
\item{CLL}{Boolean flag annotating id the described genes are significant for the pathology (CLL)}
\item{ensembl_gene_id}{Ensembl gene ID}
\item{hgnc_symbol}{Gene symbol (Hugo)}
\item{LymphGen}{Boolean flag, TRUE if lymphGen}
\item{Reddy}{Boolean flag, TRUE if gene verified by the stated study (Reddy)}
\item{Chapuy}{Boolean flag, TRUE if gene verified by the stated study (Chapuy)}
\item{entrezgene_id}{Entrez ID for the described gene}
}
}
}
\usage{
lymphoma_genes
}
\description{
A data frame with known lymphoma genes, genes are annotated by pathology, as well as literature support.
}
\keyword{datasets}
