% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoma_genes_comprehensive}
\alias{lymphoma_genes_comprehensive}
\title{Lymphoma Genes Comprehensive.}
\format{
\subsection{\code{lymphoma_genes_comprehensive}}{

A data frame with 127 rows and 9 columns.
\describe{
\item{ensemble_gene_id}{Gene in Ensemble format}
\item{Gene}{Gene symbol in Hugo format.}
\item{Chapuy}{Boolean flag, TRUE if gene verified by the stated study (Chapuy)}
\item{Reddy}{Boolean flag, TRUE if gene verified by the stated study (Reddy)}
\item{LymphGen}{Boolean flag, TRUE if lymphGen}
\item{curated}{Boolean flag, describing if the gene has been curated or not}
\item{other}{Other support for the described gene}
\item{Lacy}{Boolean flag, TRUE if gene verified by the stated study (Lacy)}
\item{aSHM}{Boolean flag for annotating aSHM}
}
}
}
\usage{
lymphoma_genes_comprehensive
}
\description{
A detailed data frame with lymphoma genes, annotated with evidence from literature and aSHM.
}
\keyword{datasets}
