% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_igv_snapshot}
\alias{make_igv_snapshot}
\title{Title}
\usage{
make_igv_snapshot(
  bams,
  genome_build,
  region,
  padding = 200,
  chrom,
  start,
  end,
  sample_id,
  out_path = "/tmp/",
  igv_port = 60506
)
}
\arguments{
\item{igv_port}{}
}
\value{

}
\description{
Title
}
\examples{
#IMPORTANT: you must be running IGV on the host that is running R and you need to have it listening on a port
# The simplest scenario is to run this command on a terminal (if using a Mac), assuming you are using R on gphost10 and you have a ssh config that routes gp10 to that host
# ssh -X gp10
# then launch IGV (e.e. from a conda installation):
# conda activate igv; igv &
this_sv = annotated_sv \%>\% filter(gene=="ETV6")
tumour_bam = get_bam_path(sample=this_sv$tumour_sample_id)
make_igv_snapshot(chrom=this_sv$chrom2, start=this_sv$start2, end=this_sv$end2, sample_id=this_sv$tumour_sample_id,out_path="~/IGV_snapshots/")
}
