% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{map_metadata_to_colours}
\alias{map_metadata_to_colours}
\title{Assign a colour palette to metadata columns automatically and consistently}
\usage{
map_metadata_to_colours(
  metadataColumns,
  these_samples_metadata,
  as_vector = TRUE,
  verbose = F,
  annoAlpha = 1
)
}
\arguments{
\item{annoAlpha}{Optional alpha to apply to annotation colours}
}
\value{

}
\description{
Assign a colour palette to metadata columns automatically and consistently
}
