% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mutation.table.df}
\alias{mutation.table.df}
\title{Default mapping table between mutation type (aka, variant classification) to mutation class}
\format{
A data frame with three columns:
\describe{
\item{Mutation_Type}{Mutation type, aka, variant classification}
\item{Mutation_Class}{mutation class}
\item{Short_Name}{short name of mutation type}
}
}
\usage{
mutation.table.df
}
\description{
A dataset containing the mapping table between genomic mutation type (aka, variant classification) to mutation class.
This dataset comes from the g3viz package and was obtained via this URL:
https://github.com/morinlab/g3viz/tree/master/data
}
\examples{
mutation.table.df
}
\keyword{datasets}
