% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_mutation_dynamics_heatmap}
\alias{plot_mutation_dynamics_heatmap}
\title{Plot a heatmap comparing the VAF of mutations in T1/T2 pairs}
\usage{
plot_mutation_dynamics_heatmap(
  maf1,
  maf2,
  vafcolname,
  patient_colname = "patient_id",
  these_samples_metadata,
  sortByColumns,
  metadata_columns = c("sample_id"),
  gene_orientation = "bottom",
  annotate_zero = FALSE,
  genes,
  top_n_genes,
  drop_unless_lowvaf = FALSE,
  vaf_cutoff_to_drop = 0.05,
  cluster_columns = F,
  cluster_rows = F
)
}
\arguments{
\item{cluster_rows}{}
}
\value{

}
\description{
Plot a heatmap comparing the VAF of mutations in T1/T2 pairs
}
