% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyChromoplot}
\alias{prettyChromoplot}
\title{Chromosome Plot}
\usage{
prettyChromoplot(
  scores,
  genes_to_label,
  cutoff = 0.5,
  adjust_amps = 0.5,
  adjust_dels = 2.75,
  label_size = 3,
  force_pull = 0,
  segment.curvature = 0.25,
  segment.ncp = 4,
  segment.angle = 25
)
}
\arguments{
\item{scores}{Output file scores.gistic from the run of GISTIC2.0}

\item{genes_to_label}{Optional. Provide a data frame of genes to label (if mutated). The first 3 columns must contain chromosome, start, and end coordinates. Another required column must contain gene names and be named \code{gene}. All other columns are ignored. If no data frame provided, oncogenes from GAMBLR packages are used by default to annotate on the plot.}

\item{cutoff}{Optional. Used to determine which regions to color as aberrant. Must be float in the range \link{0-1}. The higher the number, the less regions will be considered as aberrant. The default is 0.5.}

\item{adjust_amps}{Optional. The value of G-score for highest amplification peak will be multiplied by this value to determine how far up the gene label will be displayed. Default 0.5.}

\item{adjust_dels}{Optional. The value of G-score for highest deletion peak will be multiplied by this value to determine how far down the gene label will be displayed. Default 2.75.}

\item{label_size}{Optional. The font size for the gene label to be displayed. Default 3.}

\item{force_pull}{Optional. How strong the gene name label will be pulled towards a data point. Default 0 (no pulling).}

\item{segment.curvature}{Optional. Indicates whether arrow to the data point should be curved. Accepts numeric value, where negative is for left-hand and positive for right-hand curves, and 0 for straight lines. Default 0.25.}

\item{segment.ncp}{Optional. Indicates number of control points to make a smoother curve. Higher value allows for more flexibility for the curve. Default 4.}

\item{segment.angle}{Optional. Numeric value in the range 0-180, where less than 90 skews control points of the arrow from label to data point toward the start point. Default 25.}
}
\value{
Nothing
}
\description{
Use GISTIC2.0 scores output to reproduce maftools::chromoplot with more flexibility.
}
\details{
This function uses GISTIC2.0 scores to create a chromosome plot, based on a similar plotting function from \code{maftools}.
The only required parameter for this function is \code{scores}, which is the path to a file with GISTIC2.0 scores.
Other parameters are all optional. For a detailed explanation of how to use these, refer to the parameter descriptions.
}
\examples{
\dontrun{
#basic usage
prettyChromoplot("path_to_gistic_results/scores.gistic")

#advanced usages
prettyChromoplot(scorees = "path_to_gistic_results/scores.gistic",
                 genes_to_label = "path_to_gene_coordinates_table.tsv",
                 cutoff = 0.75) +
                  ... #any ggplot options to customize plot appearance
}

}
