% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyForestPlot}
\alias{prettyForestPlot}
\title{Create a forest plot comparing mutation frequencies for a set of genes between two groups.}
\usage{
prettyForestPlot(
  maf,
  mutmat,
  metadata,
  genes,
  comparison_column,
  comparison_values = FALSE,
  separate_hotspots = FALSE,
  comparison_name = FALSE,
  custom_colours = FALSE,
  custom_labels = FALSE,
  max_q = 1
)
}
\arguments{
\item{maf}{A maf data frame. Minimum required columns are Hugo_Symbol and Tumor_Sample_Barcode.}

\item{mutmat}{Optional argument for binary mutation matrix. If not supplied, function will generate this amtrix from the file used in argument "maf".}

\item{metadata}{Metadata for the comparisons. Minimum required columns are Tumor_Sample_Barcode and the column assigning each case to one of two groups.}

\item{genes}{An optional list of genes to restrict your plot to.}

\item{comparison_column}{Mandatory: the name of the metadata column containing the comparison values.}

\item{comparison_values}{Optional: If the comparison column contains more than two values or is not a factor, specify a character vector of length two in the order you would like the factor levels to be set, reference group first.}

\item{separate_hotspots}{Optional: If you would like to treat hotspots separately from other mutations in any gene. Requires that the maf file is annotated with GAMBLR::annotate_hotspots.}

\item{comparison_name}{Optional: Specify the legend title if different from the comparison column name.}

\item{custom_colours}{Optional: Specify a named vector of colours that match the values in the comparison column.}

\item{custom_labels}{Optional: Specify custom labels for the legend categories. Must be in the same order as comparison_values.}

\item{max_q}{cut off for q values to be filtered in fish test}
}
\value{
A ggplot object with a side-by-side forest plot and bar plot showing mutation incidences across two groups.
}
\description{
Create a forest plot comparing mutation frequencies for a set of genes between two groups.
}
\examples{
metadata = get_gambl_metadata(case_set = "tFL-study") #\%>\%
  dplyr::filter(pairing_status == "matched") \%>\%
  dplyr::filter(consensus_pathology \%in\% c("FL", "DLBCL"))

maf = get_coding_ssm(limit_samples = metadata$sample_id, basic_columns = TRUE)

prettyForestPlot(maf, 
                 metadata, 
                 genes = c("ATP6V1B2", "EZH2", "TNFRSF14", "RRAGC"),
                 comparison_column = "consensus_pathology", 
                 comparison_values = c("DLBCL", "FL"), 
                 separate_hotspots = FALSE, 
                 comparison_name = "FL vs DLBCL")

}
