% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyGeneCloud}
\alias{prettyGeneCloud}
\title{Make a word cloud of gene names from a MAF file based on mutation frequency}
\usage{
prettyGeneCloud(
  maf_df,
  these_genes,
  other_genes,
  these_genes_colour = "#B2DF8A",
  other_genes_colour = "#bc42f5",
  colour_index
)
}
\arguments{
\item{maf_df}{}

\item{these_genes}{}

\item{other_genes}{}

\item{these_genes_colour}{}

\item{other_genes_colour}{}

\item{colour_index}{}
}
\value{
data frame with counts for each gene
}
\description{
Make a word cloud of gene names from a MAF file based on mutation frequency
}
