% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyOncoplot}
\alias{prettyOncoplot}
\title{Make a non-maftools oncoplot from a MAFtools object}
\usage{
prettyOncoplot(
  maftools_obj,
  genes,
  keepGeneOrder = TRUE,
  keepSampleOrder = TRUE,
  these_samples_metadata,
  metadataColumns,
  sortByColumns,
  removeNonMutated = FALSE,
  minMutationPercent,
  fontSizeGene = 6,
  annoAlpha = 1,
  mutAlpha = 1,
  recycleOncomatrix = FALSE
)
}
\arguments{
\item{genes}{Optional list of genes to restrict to}

\item{metadataColumns}{A vector naming the columns to show as annotations on the bottom}

\item{sortByColumns}{A vector naming the columns to order on, in the order of priority}

\item{removeNonMutated}{}

\item{maf_object}{A maftools object}
}
\value{

}
\description{
Make a non-maftools oncoplot from a MAFtools object
}
