% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyRainfallPlot}
\alias{prettyRainfallPlot}
\title{Plot a rainfall plot for one sample}
\usage{
prettyRainfallPlot(
  this_sample_id,
  label_ashm_genes = TRUE,
  chrom,
  projection = "grch37",
  chromosome,
  this_maf,
  maf_path
)
}
\arguments{
\item{maf_path}{}
}
\value{
a ggplot2 plot. Print it using print() or save it using ggsave()
}
\description{
Plot a rainfall plot for one sample
}
