% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_merge_manta_with_liftover}
\alias{read_merge_manta_with_liftover}
\title{Title}
\usage{
read_merge_manta_with_liftover(
  bedpe_paths = c(),
  pattern = "--matched",
  out_dir
)
}
\arguments{
\item{out_dir}{}
}
\value{

}
\description{
Title
}
