% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reddy_genes}
\alias{reddy_genes}
\title{Reddy Genes.}
\format{
\subsection{\code{reddy_genes}}{

A data frame with 150 rows and 4 columns.
\describe{
\item{Input}{Input}
\item{hgnc_symbol}{The HGNC symbol}
\item{Approved name}{Approved name}
\item{HGNC ID}{HGNC ID}
}
}
}
\usage{
reddy_genes
}
\description{
Genes identified as significantly mutated in DLBCL by the study of Reddy et al.
}
\keyword{datasets}
