% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{region_to_bins}
\alias{region_to_bins}
\title{Split a contiguous genomic region on a chromosome into non-overlapping bins}
\usage{
region_to_bins(
  chromosome = "chr1",
  start = 10000,
  end = 121500000,
  bin_size = 2000
)
}
\arguments{
\item{chromosome}{}

\item{start}{}

\item{end}{}

\item{bin_size}{}
}
\value{
Data frame describing the bins various ways
}
\description{
Split a contiguous genomic region on a chromosome into non-overlapping bins
}
\examples{
chr8q_bins = region_to_bins(chromosome="8",start=48100000,end=146364022,bin_size = 20000)
}
