% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{region_to_gene}
\alias{region_to_gene}
\title{Return gennes residing in defined region(s)}
\usage{
region_to_gene(
  region,
  gene_format = "hugo",
  genome_build = "grch37",
  chr_select = paste0("chr", c(1:22))
)
}
\arguments{
\item{region}{Regions to intersect genes with, this should be a bed-like df (chromosome, start, end).}

\item{gene_format}{Parameter for specifying the format of returned genes, default is "hugo", other acceptable inputs are "ensembl".}

\item{genome_build}{Reference genome build.}

\item{chr_select}{Optional parameter to subset plot to specific chromosomes. Default value is chr1-22.}
}
\description{
Return gennes residing in defined region(s)
}
\examples{
myc_region = gene_to_region(gene_symbol = "MYC", genome_build = "grch37", return_as = "df")
region = region_to_gene(region = myc_region, gene_format = "hugo", genome_build = "grch37")

}
