% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversekelly.R
\name{reversekelly}
\alias{reversekelly}
\title{Reverse Kelly Criterion}
\usage{
reversekelly(oldodds, newodds, betpercent, oddstype1 = default.GamblR.odds(),
  oddstype2 = default.GamblR.odds())
}
\arguments{
\item{oldodds}{Original odds}

\item{newodds}{New odds}

\item{betpercent}{Bet percentage "tipped" for the original odds}

\item{oddstype1}{Format of original odds (e.g. decimal, see Convert odds)}

\item{oddstype2}{Format of new odds (e.g. decimal, see Convert odds)}
}
\value{
A list containing the Kelly optimal bet for the new odds, the implied probability of the 'tip', the old implied probability, the old edge, the new implied probability, and the new edge.
}
\description{
Reverse Kelly Criterion calculation
}
\details{
Calculates the implied propability from a "tip" assuming it is Kelly optimal. Useful for tailing a tip when the odds at your book are different. (e.g. If 5\% at $1.75 is Kelly optimal then 2\% at $1.70 may also be so.
}
