% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rose.R
\name{rose}
\alias{rose}
\title{S Criterion}
\usage{
rose(b, p, sd, odds = "decimal")
}
\arguments{
\item{b}{Vector of event odds}

\item{p}{Vector of event probabilities}

\item{sd}{Vector of event probability standard deviations}

\item{odds}{Format of input odds (e.g. decimal, see Convert odds)}

\item{fraction}{Bet fraction (for fractional Kelly)}
}
\value{
A list contaning the S optimal bet fraction, the Kelly optimal bet fraction, and the ratio of S to Kelly (for use with Simultaneous Kelly).
}
\description{
Similar to the Kelly Criterion but for known standard deviation
}
