% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sanitize_maf_data}
\alias{sanitize_maf_data}
\title{Convert mutation data to a shareable format.}
\usage{
sanitize_maf_data(
  mutation_maf_path,
  mutation_maf_data,
  output_oncomatrix,
  genes_keep,
  genes_drop = c()
)
}
\arguments{
\item{mutation_maf_path}{Provide either the full path to a MAF file.}

\item{mutation_maf_data}{Otherwise provide a data frame of the MAF data.}

\item{output_oncomatrix}{Optionally provide the path for your sanitized output file (otherwise it writes to the working directory).}

\item{genes_keep}{Specify which genes you want to remain in the output.}

\item{genes_drop}{Optionally specify which genes to drop (this doesn't mean all other genes will remain. Maftools decides that part).}
}
\value{
The full path to the oncomatrix file (a matrix with Variant_Classification or Multi_Hit indicating coding mutation status per patient).
}
\description{
\code{sanitize_maf_data} returns an oncomatrix of patient/gene data indicating only data needed to produce an oncoplot.
}
\details{
Write an oncomatrix from a MAF File for further plotting. This is meant to be run by individuals who have access to data sets to
"sanitize" a subset of data for subsequent use by them or others who don't have permission to access the raw data.
Example: User J has full permissions for ICGC data and has read permissions on a MAF file. User B needs to make some oncoplots
and/or perform some statistical analysis on the frequency and assortment of mutations in that data set but don't need all the details.
User J can run this function on a maf file and provide the path of the output to user B.
}
\examples{
#note, this will be used by default if the user wants to be lazy
lymph_genes = lymphoma_genes$Gene

secure_maf = "/projects/rmorin/projects/gambl-repos/gambl-rmorin/results/icgc_dart/
slms-3_vcf2maf_current/level_3/final_merged_grch37.CDS.maf"
safe_oncomatrix_path = sanitize_maf_data(mutation_maf_path = secure_maf,
                                         genes_keep = lymph_genes)

}
