% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sanity_check_metadata}
\alias{sanity_check_metadata}
\title{Sanity Check Metadata.}
\usage{
sanity_check_metadata()
}
\value{
A table.
}
\description{
Function that performs sanity checks on metadata.
}
\details{
Helper function for sanity checking GAMBL metadata.
}
\examples{
sane_meta_data = sanity_check_metadata()

}
