% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rv.R
\name{set_rv}
\alias{set_rv}
\title{Declare Random Variables}
\usage{
set_rv(
  x,
  ...,
  params = list(),
  overwrite = FALSE,
  verbose = FALSE,
  list_dist = pdist
)
}
\arguments{
\item{x}{An object from "sim"}

\item{...}{Formulas declaring random variables (e.g., `u ~ uniform(0, 1)`).}

\item{params}{A named list of local constants or parameters.}

\item{overwrite}{Logical; whether to overwrite previously defined variables.}

\item{verbose}{Logical; whether to display the defined random variables.}

\item{list_dist}{Built-in distribution in `gamblr` package to be used.}
}
\value{
Updated simulation context with declared random variables.
}
\description{
Defines one or more random variables using formula syntax (e.g., `x ~ normal(0, 1)`).
}
