% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{setup_fusions}
\alias{setup_fusions}
\title{Annotate SVs and create the input for fusions to be displayed in cBioPortal instance}
\usage{
setup_fusions(
  short_name = "GAMBL",
  include_icgc_data = FALSE,
  human_friendly_name = "GAMBL data",
  project_name = "gambl_minus_icgc",
  description = "GAMBL data without ICGC",
  out_dir
)
}
\arguments{
\item{short_name}{a concise name for your portal project}

\item{include_icgc_data}{whether or not you want ICGC and other external data included}

\item{human_friendly_name}{a slightly more verbose name for your project}

\item{project_name}{unique ID for your project}

\item{description}{A verbose description of your data set}

\item{out_dir}{The full path to the base directory where the files are being created}
}
\value{
a vector of sample_id for the patients that have been included
}
\description{
Annotate SVs and create the input for fusions to be displayed in cBioPortal instance
}
