% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{setup_study}
\alias{setup_study}
\title{Initialize a new cBioPortal instance or update existing portal data set, can also be used to retrieve sample ids included in study.}
\usage{
setup_study(
  seq_type_filter = c("capture"),
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_all",
  description = "GAMBL data from exomes",
  overwrite = TRUE,
  out_dir
)
}
\arguments{
\item{short_name}{A concise name for your portal project.}

\item{human_friendly_name}{A slightly more verbose name for your project.}

\item{project_name}{Unique ID for your project.}

\item{description}{A verbose description of your data set.}

\item{overwrite}{Flag to specify that files should be overwritten if they exist. Default is TRUE.}

\item{out_dir}{The full path to the base directory where the files are being created.}

\item{include_icgc_data}{Whether or not you want ICGC and other external data included.}

\item{exclude_cohorts}{Cohorts to be excluded.}

\item{gambl_maf}{maf origin.}

\item{gambl_icgc_maf}{Icgc maf origin.}
}
\value{
A vector of sample_id for the patients that have been included.
}
\description{
Initialize a new cBioPortal instance or update existing portal data set, can also be used to retrieve sample ids included in study.
}
\examples{
Setup study and save included ids as a vector of characters:
ids = setup_study(out_dir = "GAMBLR/cBioPortal/instance01/")

}
