% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standardize_chr_prefix}
\alias{standardize_chr_prefix}
\title{Standardize Chromosome Prefix.}
\usage{
standardize_chr_prefix(incoming_vector, projection)
}
\arguments{
\item{incoming_vector}{Input vector of any length with chromosome names.}

\item{projection}{Projection to which chr prefix should be standardized.}
}
\value{
A vector of chromosome names with prefix standardized to projection
}
\description{
Standardize the chr prefix in a vector of chromosome names based on projection.
}
\details{
INTERNAL FUNCTION, not meant for out-of-package use.
}
\examples{
these_chrs = c(8, "13", "chr4", "chrY")
standardize_chr_prefix(these_chrs, projection = "hg38")

}
