% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subset_cnstates}
\alias{subset_cnstates}
\title{Helper function called by fancy_multisample_ideo, for sub-setting copy number information based on segments avaialble in cn data}
\usage{
subset_cnstates(cn_segments, samplen)
}
\arguments{
\item{cn_segments}{DF with copy number segments, usually retrieved from get_sample_cn_segments.}

\item{samplen}{Numeric value that annotates the sample order.}
}
\value{
Nothing.
}
\description{
Helper function called by fancy_multisample_ideo, for sub-setting copy number information based on segments avaialble in cn data
}
\examples{
cn_states = get_sample_cn_segments(multiple_samples = TRUE, sample_list = c("00-15201_tumorA", "HTMCP-01-06-00422-01A-01D"), streamlined = FALSE)
subset_cnstates(cn_segments = cn_states, samplen = 1)

}
