% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{supplement_maf}
\alias{supplement_maf}
\title{Complement maf with missing samples.}
\usage{
supplement_maf(incoming_maf, these_samples_metadata)
}
\arguments{
\item{incoming_maf}{The initial MAF data frame to be supplemented with missing samples.}

\item{these_samples_metadata}{The metadata data frame that contains Tumor_Sample_Barcode column with ids to be present in the complemented MAF.}
}
\value{
maf_df with complemented Tumor_Sample_Barcode and other columns ready to be used downstream
}
\description{
Complement maf with missing samples.
}
\examples{
small_maf = get_coding_ssm(limit_cohort = "dlbcl_reddy", seq_type = "capture") \%>\% dplyr::filter(Hugo_Symbol=="MYC")
reddy_meta = get_gambl_metadata(seq_type_filter = "capture") \%>\% dplyr::filter(cohort=="dlbcl_reddy")
complete_maf = supplement_maf(incoming_maf = small_maf, these_samples_metadata = reddy_meta)

}
