% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_regions_grch37}
\alias{target_regions_grch37}
\title{Target Regions grch37.}
\format{
\subsection{\code{target_regions_grch37}}{

A data frame with 295994 rows and 3 columns.
\describe{
\item{chrom}{Chromosome for the descriebd region}
\item{start}{Start coordinate of the region}
\item{end}{End coordiante of the region}
}
}
}
\usage{
target_regions_grch37
}
\description{
Target regions in respect to grch37.
}
\keyword{datasets}
