% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_regions_hg38}
\alias{target_regions_hg38}
\title{Target Regions hg38.}
\format{
\subsection{\code{target_regions_hg38}}{

A data frame with 296453 rows and 3 columns.
\describe{
\item{chrom}{Chromosome for the descriebd region}
\item{start}{Start coordinate of the region}
\item{end}{End coordiante of the region}
}
}
}
\usage{
target_regions_hg38
}
\description{
Target regions in respect to hg38.
}
\keyword{datasets}
