% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tidy_lymphgen}
\alias{tidy_lymphgen}
\title{Tidy Lymphgen.}
\usage{
tidy_lymphgen(
  df,
  lymphgen_column_in = "Subtype.Prediction",
  lymphgen_column_out = "Subtype.Prediction",
  relevel = FALSE
)
}
\arguments{
\item{df}{Input data frame.}

\item{lymphgen_column_in}{The name of the column with lymphgen data to be processed.}

\item{lymphgen_column_out}{The name of the column to write the tidied results (optional).}

\item{relevel}{If TRUE, will return the output column as a factor with plot-friendly levels.}
}
\value{
A data frame with a tidied lymphGen column
}
\description{
Consolidate a column of LymphGen data in the original Subtype.Prediction output format to the GAMBLR tidy format.
}
\details{
This function takes an incoming data frame (\code{df}) and consolidates a column of LymphGen data.
Specify the column with the lymphgen data to be processed with \code{lymphgen_column_in} and
what column to write the tidied data to with \code{lymphgen_column_out}.
In addition, the user can also run this function with \code{relevel = TRUE} (default is FALSE),
to return the output column as a factor with plot friendly levels.
}
\examples{
metadata = get_gambl_metadata()
tidy_lymphgen(df = metadata,
              lymphgen_column_in = "lymphgen_with_cnv",
              lymphgen_column_out = "lymphgen_with_cnv_tidy",
              relevel = TRUE)

}
