% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trim_scale_expression}
\alias{trim_scale_expression}
\title{Trim Scale Expressions.}
\usage{
trim_scale_expression(x)
}
\arguments{
\item{x}{Numeric value (of expression) to be trimmed.}
}
\value{
Numeric value.
}
\description{
INTERNAL FUNCTION called by prettyOncoplot, not meant for out-of-package usage.
}
\details{
INTERNAL FUNCTION called by prettyOncoplot, not meant for out-of-package usage.
}
\examples{
\dontrun{
trimmed = trim_scale_expression(2)
}

}
