% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_fma_ucb.R
\name{UCBPolicy}
\alias{UCBPolicy}
\title{R6 Class representing the Upper Confidence Bound policy}
\description{
An UCBPolicy object is instantiated using exploration
  parameter alpha. The method run is used to process a matrix containing the
  arms rewards.
}
\details{
The Upper Confidence Bound Algorithm is used in stochastic bandits
  with finitely many arms problems. Like other bandit algorithms, arms are
  chosen according to their indices computed at each round. Here this index
  is the upper confidence bound parameterized by the constant alpha.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Short string describing the instance.}

\item{\code{record}}{R6 object of class Recorder used to export simulation
results.}

\item{\code{alpha}}{Numeric scalar containing UCB parameter.}

\item{\code{mu_hat}}{Numeric vector containing reward expectation of each arm.}

\item{\code{N}}{Numeric vector containing number of times each arm was pulled.}

\item{\code{K}}{Integer containing the number of arms}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{UCBPolicy$new()}}
\item \href{#method-reset}{\code{UCBPolicy$reset()}}
\item \href{#method-compute_UCB_for_an_arm}{\code{UCBPolicy$compute_UCB_for_an_arm()}}
\item \href{#method-compute_UCB_for_all_arms}{\code{UCBPolicy$compute_UCB_for_all_arms()}}
\item \href{#method-update_with_reward}{\code{UCBPolicy$update_with_reward()}}
\item \href{#method-run}{\code{UCBPolicy$run()}}
\item \href{#method-clone}{\code{UCBPolicy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new UCBPolicy object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCBPolicy$new(alpha = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Exploration parameter.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `UCBPolicy` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset"></a>}}
\if{latex}{\out{\hypertarget{method-reset}{}}}
\subsection{Method \code{reset()}}{
Set or reset object attributes before starting a simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCBPolicy$reset(K)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{K}}{Number of arms.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_UCB_for_an_arm"></a>}}
\if{latex}{\out{\hypertarget{method-compute_UCB_for_an_arm}{}}}
\subsection{Method \code{compute_UCB_for_an_arm()}}{
Computes the upper confidence bound of arm a at time t.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCBPolicy$compute_UCB_for_an_arm(t, a)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Time. Integer.}

\item{\code{a}}{Arm. Integer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The upper confidence bound of arm a at time t.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_UCB_for_all_arms"></a>}}
\if{latex}{\out{\hypertarget{method-compute_UCB_for_all_arms}{}}}
\subsection{Method \code{compute_UCB_for_all_arms()}}{
Computes the upper confidence bound of all arms at time t.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCBPolicy$compute_UCB_for_all_arms(t)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{Time. Integer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector containing the upper confidence bound of all arms.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_with_reward"></a>}}
\if{latex}{\out{\hypertarget{method-update_with_reward}{}}}
\subsection{Method \code{update_with_reward()}}{
Updates mu_hat and N given reward r of arm a.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCBPolicy$update_with_reward(r, a)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{Reward.}

\item{\code{a}}{Arm.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Process a bandit problem represented by the reward matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCBPolicy$run(rewards, verbose = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rewards}}{Reward matrix. Time on rows, arms on columns.}

\item{\code{verbose}}{Logical. if FALSE, only action and reward history will be
returned. Default is TRUE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tribble describing at each iteration, the UCB value and reward
  expectation of each arm, the action taken and the reward received. If
  verbose is FALSE, only the action and reward.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UCBPolicy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
