% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getwinper.R
\name{winper}
\alias{winper}
\title{Determine win percentages}
\usage{
winper(odds, oddstype = default.GamblR.odds())
}
\arguments{
\item{odds}{A vector of odds for a single event (e.g. win/loss, win/loss/tie)}

\item{oddstype}{Format of input odds (e.g. decimal, see Convert odds)}
}
\value{
A list containing the implied probability for each event along with the book's margin
}
\description{
Calculates the probablities each time will win along with the book's margin
}
