% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wright_genes_with_weights}
\alias{wright_genes_with_weights}
\title{Wright Genes With Weights.}
\format{
\subsection{\code{wright_genes_with_weights}}{

A data frame with 210 rows and 3 columns.
\describe{
\item{Ensembl_ID}{Gene in Ensembl ID format}
\item{Hugo_Symbol}{Gene symbol in Hugo format}
\item{Weight_tValue}{Weight Value for the specified gene}
}
}
}
\usage{
wright_genes_with_weights
}
\description{
Description.
}
\keyword{datasets}
