% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_driver_ssm}
\alias{annotate_driver_ssm}
\title{Annotate Driver SSM.}
\usage{
annotate_driver_ssm(
  maf_df,
  lymphoma_type,
  driver_genes,
  noncoding_regions = c(NFKBIZ = "chr3:101578206-101578365", HNRNPH1 =
    "chr5:179,045,946-179,046,683")
)
}
\arguments{
\item{maf_df}{Data frame of MAF-format mutations.}

\item{lymphoma_type}{Optional keyword to find genes for annotation (e.g. "BL","DLBCL","FL") is not specifying driver genes.}

\item{driver_genes}{Optional vector of Hugo_Symbol of genes for coding mutations.}

\item{noncoding_regions}{Optional named vector of regions to use to further restrict noncoding mutations per gene.}
}
\value{
Incoming MAF, subset to putative driver mutations.
}
\description{
Retrieve data or use the supplied data frame of genes to get/annotate MAF data indicating which rows are putative driver mutations.
}
\details{
Provide a maf-like data frame with \code{maf_df} as the only required parameter.
For information on how to use the additional parameters, refer to the parameter descriptions and function examples.
}
\examples{
driver_ssm = annotate_driver_ssm(maf_df = grande_maf,
                                 lymphoma_type = "DLBCL",
                                 noncoding_regions=c("NFKBIZ"="chr3:101578206-101578365"))

}
