% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{annotate_hotspots}
\alias{annotate_hotspots}
\title{Annotate Hotspots.}
\usage{
annotate_hotspots(
  mutation_maf,
  recurrence_min = 5,
  analysis_base = c("FL--DLBCL", "BL--DLBCL"),
  p_thresh = 0.05
)
}
\arguments{
\item{mutation_maf}{A data frame in MAF format.}

\item{recurrence_min}{minimum number of recurrences for mutation to be included, default is 5.}

\item{analysis_base}{Base name for hot spot output directory.}

\item{p_thresh}{P value threshold, default is 0.05.}
}
\value{
The same data frame with one additional column "hot_spot".
}
\description{
Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations.
}
\details{
This function takes an already loaded MAF data frame with the \code{mutation_maf} parameter.
The user can then control the minimum number of recurrences for mutations to be included with \code{recurrance_min},
The default is 5. \code{analysis_base} controls the base name go hotspot output directory.
Lastly, \code{p_thresh} sets the p value threshold, default is 0.05.
}
\examples{
hot_ssms = annotate_hotspots(all_ssm)
hot_maf = read.maf(hot_ssms)
oncoplot(hot_maf,genes=c("MEF2B","TP53","MYD88"),additionalFeature = c("hot_spot",TRUE))

}
