% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_recurrent_cnv}
\alias{annotate_recurrent_cnv}
\title{Annotate Recurrent CNVs.}
\usage{
annotate_recurrent_cnv(seg_df, seg_file)
}
\arguments{
\item{seq_df}{Data frame of sequences with start and end coordinates.}

\item{seq_file}{Optional argument to read sequences from file (currently not used in function).}
}
\value{
Nothing.
}
\description{
Annotates recurrent CNVs from a data frame with CNV data.
}
\details{
This function takes a data frame with CNVs (\code{seq_df}) and annotates recurrent CNVs.
}
\examples{
annotated = annotate_recurrent_cnv(my_seq_df)
annotated = annotate_recurrent_cnv(seq_file = "my_seq_data.txt")

}
