% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_recurrent_cnv}
\alias{annotate_recurrent_cnv}
\title{Annotate Recurrent CNVs.}
\usage{
annotate_recurrent_cnv(seg_df, seg_file)
}
\arguments{
\item{seg_df}{Data frame of sequences with start and end coordinates.}

\item{seg_file}{Optional argument to read sequences from file (currently not used in function).}
}
\value{
Nothing.
}
\description{
Annotates recurrent CNVs from a data frame with CNV data.
}
\details{
This function takes a data frame with CNVs (\code{seq_df}) and annotates recurrent CNVs.
}
\examples{
\dontrun{
my_segs = get_sample_cn_segments(this_sample_id = "HTMCP-01-06-00422-01A-01D")
annotated = annotate_recurrent_cnv(seg_df = my_segs)
}

}
