% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_sv}
\alias{annotate_sv}
\title{Annotate a data frame of SV breakpoints after retrieval from the database}
\usage{
annotate_sv(sv_data, genome_build = "grch37")
}
\arguments{
\item{sv_data}{A data frame of SVs. This should be the output of get_manta_sv.}

\item{genome_build}{Which reference genome these SVs are from}
}
\value{
A data frame with annotated SVs (gene symbol and entrez ID)
}
\description{
Annotate a data frame of SV breakpoints after retrieval from the database
}
\examples{
# Basic usage
annotated_sv = annotate_sv(sv_df)
}
