% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_sv}
\alias{annotate_sv}
\title{Annotate a data frame of SV breakpoints after retrieval from the database}
\usage{
annotate_sv(
  sv_data,
  partner_bed,
  with_chr_prefix = FALSE,
  collapse_redundant = FALSE,
  return_as = "bedpe"
)
}
\arguments{
\item{sv_data}{A data frame of SVs. This should be the output of get_manta_sv.}

\item{partner_bed}{Optional bed-format data frame to use for annotating oncogene partners (e.g. enhancers). required columns are: chrom,start,end,gene}

\item{with_chr_prefix}{Optionally request that chromosome names are returned with a chr prefix}

\item{collapse_redundant}{Remove reciprocal events and only return one per event}
}
\value{
A data frame with annotated SVs (gene symbol and entrez ID)
}
\description{
Annotate a data frame of SV breakpoints after retrieval from the database
}
\examples{
# Basic usage
sv_df = get_manta_sv()
annotated_sv = annotate_sv(sv_df)
}
