% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{append_to_table}
\alias{append_to_table}
\title{Housekeeping function to add results to a table}
\usage{
append_to_table(table_name, data_df)
}
\arguments{
\item{table_name}{The name of the database table to update/populate}

\item{data_df}{A dataframe of values to load into the table}
}
\value{

}
\description{
Housekeeping function to add results to a table
}
