% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ashm_multi_rainbow_plot}
\alias{ashm_multi_rainbow_plot}
\title{ASHM Multi-panel Rainbow Plot}
\usage{
ashm_multi_rainbow_plot(
  regions_bed,
  regions_to_display,
  exclude_classifications,
  metadata,
  seq_type,
  custom_colours,
  classification_column = "lymphgen",
  maf_data
)
}
\arguments{
\item{regions_bed}{Bed file with chromosome coordinates, should contain columns chr, start, end, name (with these exact names).}

\item{regions_to_display}{Optional vector of names from default regions_bed to use.}

\item{exclude_classifications}{Optional argument for excluding specific classifications from a metadeta file.}

\item{metadata}{A metadata file already subsetted and arranged on the order you want the samples vertically displayed.}

\item{seq_type}{the seqtype you want results back for if \code{maf_data} is not provided.}

\item{custom_colours}{Provide named vector (or named list of vectors) containing custom annotation colours if you do not want to use standartized pallette.}

\item{classification_column}{Optional. Override default column for assigning the labels used for colouring in the figure.}

\item{maf_data}{An already loaded maf, if no provided, this function will call \code{get_ssm_by_region}, using the regions supplied into \code{regions_bed}.}
}
\value{
Nothing
}
\description{
Generates a colourful multi-panel overview of hypermutation in regions of interest across many samples.
}
\details{
The input for this function is a bed-file with the following columns; chr, start, end, name.
Note that for this function to work, the column names must be exactly this.
The user also needs to specify a vector of names (\code{regions_to_display}) to further control what regions are to be displayed on the returned plot.
It is also possible to exclude specific classifications from the metadata file. This is achieved with \code{exclude_classifications}.
In addition the user can also use the \code{metadata} parameter to use an already subset and arranged metadata table.
This function will call \link{get_ssm_by_region} if \code{maf_data} is not called. For more info, refer to the parameter descriptions of this function.
}
\examples{
#get lymphgen colours
lymphgen_colours = get_gambl_colours(classification = "lymphgen")

#build plot
ashm_multi_rainbow_plot(regions_to_display = c("BCL2-TSS",
                                               "MYC-TSS",
                                               "SGK1-TSS",
                                               "IGL"),
                        custom_colours = lymphgen_colours,
                        seq_type = "genome")

}
