% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{ashm_multi_rainbow_plot}
\alias{ashm_multi_rainbow_plot}
\title{Generate a colourful multi-panel overview of hypermutation in regions of interest across many samples}
\usage{
ashm_multi_rainbow_plot(
  regions_bed,
  regions_to_display,
  exclude_classifications,
  metadata,
  custom_colours,
  classification_column = "lymphgen",
  db = "gambl_test",
  table_name = "maf_slms3_hg19_icgc",
  maf_data
)
}
\arguments{
\item{regions_bed}{Bed file with chromosome coordinates, should contain columns chr, start, end, name (with these exact names)}

\item{regions_to_display}{Optional vector of names from default regions_bed to use}

\item{metadata}{A metadata file already subsetted and arranged on the order you want the samples vertically displayed}

\item{classification_column}{optional. Override default column for assigning the labels used for colouring in the figure.}

\item{db}{optional. Specify a different database (changing this is not recommended)}

\item{table_name}{optional. Specify a different table to query (changing this is not recommended)}
}
\value{
nothing
}
\description{
Generate a colourful multi-panel overview of hypermutation in regions of interest across many samples
}
