% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{ashm_rainbow_plot}
\alias{ashm_rainbow_plot}
\title{ASHM Rainbow Plot}
\usage{
ashm_rainbow_plot(
  mutations_maf,
  metadata,
  exclude_classifications,
  drop_unmutated = FALSE,
  classification_column,
  bed,
  region,
  custom_colours,
  hide_ids = TRUE
)
}
\arguments{
\item{mutations_maf}{A data frame containing mutations (MAF format) within a region of interest (i.e. use the get_ssm_by_region).}

\item{metadata}{should be a data frame with sample_id as a column.}

\item{exclude_classifications}{Optional argument for excluding specific classifications from a metadeta file.}

\item{drop_unmutated}{Boolean argument for removing unmutated sample ids in mutated cases.}

\item{classification_column}{The name of the metadata column to use for ordering and colouring samples.}

\item{bed}{Optional data frame specifying the regions to annotate (required columns: start, end, name).}

\item{region}{Genomic region for plotting in bed format.}

\item{custom_colours}{Provide named vector (or named list of vectors) containing custom annotation colours if you do not want to use standartized pallette.}

\item{hide_ids}{Boolean argument, if TRUE, ids will be removed.}
}
\value{
ggplot2 object
}
\description{
Make a rainbow plot of all mutations in a region, ordered and coloured by metadata.
}
\details{
This function creates a rainbow plot for all mutations in a region. Region can either be specified with the \code{region} parameter,
or the user can provide a maf that has already been subset to the region(s) of interest with \code{mutation_maf}.
As a third alternative, the regions can also be specified as a bed file with \code{bed}.
Lastly, this function has a variety of parameters that can be used to further customize the returned plot in many different ways.
Refer to the parameter descriptions, examples as well as the vignettes for more demonstrations how this function can be called.
}
\examples{
#basic usage
region = "chr6:90975034-91066134"
metadata = get_gambl_metadata()
plot = ashm_rainbow_plot(metadata = metadata, region = region)

#advanced usages
mybed = data.frame(start=c(128806578,128805652,128748315), end=c(128806992,128809822,128748880), name=c("TSS","enhancer","MYC-e1"))
ashm_rainbow_plot(mutations_maf=my_mutations,metadata=my_metadata,bed=mybed)

}
