% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{assemble_file_details}
\alias{assemble_file_details}
\title{Assemble File Details.}
\usage{
assemble_file_details(
  file_details_df,
  file_paths,
  tool_name,
  unix_group,
  these_sample_ids,
  output_type = "ploidy",
  is_production = "yes"
)
}
\arguments{
\item{file_details_df}{Optionally supply the data frame directly instead (e.g. from find_files_extract_wildcards).}

\item{file_paths}{A vector of full file paths, e.g. the output of dir.}

\item{tool_name}{The tool or pipeline that generated the files (should be the same for all).}

\item{unix_group}{The unix group (should be the same for all).}

\item{these_sample_ids}{A vector of sample_id the same length and in the same order as the file paths.}

\item{output_type}{The file type to distinguish different output file types from the same pipeline (e.g. seg, maf, ploidy).}

\item{is_production}{Boolean parameter. Default is yes.}
}
\value{
Updates the database by appending to the gambl_files table. Use with caution!
}
\description{
Update the database by appending to the gambl_files table.
}
\details{
Specify the file paths with \code{file_paths} followed by the name of the tool (\code{tool_name}).
Next, set the output type (e.g seq, maf, etc.) and unix group (should be the same for all).
Lastly, specify the sample IDs with \code{these_sample_ids}.
For more information on how to use the optional parameters, refer to the parameter descriptions.
}
\examples{
assemble_file_details(file_paths = c(one.maf, another.maf), tool_name = "manta", unix_group = "genome", output_type = "maf", these_sample_ids = c(one_sample, another_sample))

}
