% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{assemble_file_details}
\alias{assemble_file_details}
\title{Title}
\usage{
assemble_file_details(
  file_details_df,
  file_paths,
  tool_name,
  unix_group,
  these_sample_ids,
  output_type = "ploidy",
  is_production = "yes"
)
}
\arguments{
\item{file_details_df}{Optionally supply the data frame directly instead (e.g. from find_files_extract_wildcards)}

\item{file_paths}{A vector of full file paths, e.g. the output of dir}

\item{tool_name}{The tool or pipeline that generated the files (should be the same for all)}

\item{unix_group}{The unix group (should be the same for all)}

\item{these_sample_ids}{A vector of sample_id the same length and in the same order as the file paths}

\item{output_type}{The file type to distinguish different output file types from the same pipeline (e.g. seg, maf, ploidy)}

\item{tool_version}{Optional: provide the version of the pipeline or tool}
}
\value{
Updates the database by appending to the gambl_files table. Use with caution!
}
\description{
Title
}
