% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_mutation_frequency_sliding_windows}
\alias{calc_mutation_frequency_sliding_windows}
\title{Count the number of mutations in a sliding window across a region for all samples. Unlikely to be used directly in most cases. See get_mutation_frequency_bin_matrix instead}
\usage{
calc_mutation_frequency_sliding_windows(
  this_region,
  chromosome,
  start_pos,
  end_pos,
  metadata,
  slide_by = 100,
  window_size = 1000,
  plot_type = "none",
  return_format = "long-simple",
  min_count_per_bin = 3,
  return_count = FALSE,
  drop_unmutated = FALSE,
  classification_column = "lymphgen"
)
}
\arguments{
\item{plot_type}{Set to true for a plot of your bins. By default no plots are made.}

\item{drop_unmutated}{This may not currently work properly.}

\item{classification_column}{Only used for plotting}
}
\value{

}
\description{
Count the number of mutations in a sliding window across a region for all samples. Unlikely to be used directly in most cases. See get_mutation_frequency_bin_matrix instead
}
