% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc_mutation_frequency_sliding_windows}
\alias{calc_mutation_frequency_sliding_windows}
\title{Count the number of mutations in a sliding window across a region for all samples. Unlikely to be used directly in most cases. See get_mutation_frequency_bin_matrix instead}
\usage{
calc_mutation_frequency_sliding_windows(
  this_region,
  chromosome,
  start_pos,
  end_pos,
  metadata,
  slide_by = 100,
  window_size = 1000,
  plot_type = "none",
  sortByColumns = "pathology",
  return_format = "long-simple",
  min_count_per_bin = 3,
  return_count = FALSE,
  drop_unmutated = FALSE,
  classification_column = "lymphgen",
  from_indexed_flatfile = FALSE,
  mode = "slms-3"
)
}
\arguments{
\item{plot_type}{Set to true for a plot of your bins. By default no plots are made.}

\item{sortByColumns}{Which of the metadata to sort on for the heatmap}

\item{drop_unmutated}{This may not currently work properly.}

\item{classification_column}{Only used for plotting}

\item{from_indexed_flatfile}{Set to TRUE to avoid using the database and instead rely on flatfiles (only works for streamlined data, not full MAF details)}

\item{mode}{Only works with indexed flatfiles. Accepts 2 options of "slms-3" and "strelka2" to indicate which variant caller to use. Default is "slms-3".}
}
\value{

}
\description{
Count the number of mutations in a sliding window across a region for all samples. Unlikely to be used directly in most cases. See get_mutation_frequency_bin_matrix instead
}
