% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_host}
\alias{check_host}
\title{Check if code is running remotely and (optionally) attempt a connection and set global ssh_session variable}
\usage{
check_host(auto_connect = FALSE)
}
\arguments{
\item{auto_connect}{Set to TRUE if you want the function to create an ssh session (if necessary)}
}
\value{
NULL. This function makes a new ssh session, if necessary, and stores it in a global variable (ssh_session)
}
\description{
Check if code is running remotely and (optionally) attempt a connection and set global ssh_session variable
}
\examples{
ssh_session = get_ssh_session()
}
