% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{classify_dlbcl}
\alias{classify_dlbcl}
\title{Classify DLBCLs according to genetic subgroups.}
\usage{
classify_dlbcl(
  these_samples_metadata,
  maf_data,
  seg_data,
  sv_data,
  this_seq_type = "genome",
  projection = "grch37",
  output = "predictions",
  method = "chapuy",
  adjust_ploidy = TRUE,
  include_N1 = FALSE
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains sample_id column with ids for the samples to be classified.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. At least must contain the first 45 columns of standard MAF format.}

\item{seg_data}{The SEG data frame to be used for matrix assembling. Must be of standard SEG formatting, for example, as returned by get_sample_cn_segments.}

\item{sv_data}{The SV data frame to be used for matrix assembling. Must be of standard BEDPE formatting, for example, as returned by get_combined_sv.}

\item{this_seq_type}{The seq_type of the samples. Only used to retrerive data through GAMBLR when it is not provided. Defaults to genome.}

\item{projection}{The projection of the samples. Only used to retrerive data through GAMBLR when it is not provided. Defaults to grch37.}

\item{output}{The output to be returned after prediction is done. Can be one of predictoins, matrix, or both. Defaults to predictions.}

\item{adjust_ploidy}{Whether to perform ploidy adjustment for the CNV data. Defaults to TRUE (recommended).}

\item{include_N1}{When running Lacy classifieer, indicate whether to set samples with NOTCH1 mutations to N1 group as described in Runge et al (2021). Defaults to FALSE.}
}
\value{
data frame with classification, binary matrix used in classification, or both
}
\description{
Using the user-provided or GAMBLR-retrieved data, this function will assemble the matrix according to the approach of
Chapuy et al (2018) or Lacy et al (2020) classifiers. Since neither of this classifiers is publicly released, we have implemented a solution
that closely (> 92\% accuracy) recapitulates each of these systems. For the classifier of Chapuy et al, the constructed matrix will be
used to calculate class probability using the feature weights from bundled NMF model. For the Lacy et al classifier, the matrix will be
used for prediction of random forest model,  which is supplied with the GAMBLR package.
}
\examples{
test_meta <- get_gambl_metadata(case_set = "DLBCL-unembargoed")
predictions_chapuy <- classify_dlbcl(these_samples_metadata = test_meta)
predictions_lacy <- classify_dlbcl(these_samples_metadata = test_meta, method = "lacy")
matrix_and_predictions <- classify_dlbcl(these_samples_metadata = test_meta, output = "both")

}
