% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{classify_fl}
\alias{classify_fl}
\title{Classify FL.}
\usage{
classify_fl(
  these_samples_metadata,
  maf_data,
  model = RFmodel_FL,
  this_seq_type = "genome",
  output = "predictions"
)
}
\arguments{
\item{these_samples_metadata}{The metadata data frame that contains a sample_id column with ids for the samples to be classified.}

\item{maf_data}{The MAF data frame to be used for matrix assembling. At least must contain the first 45 columns of standard MAF format.}

\item{model}{The RF model. Classifier from the paper describing cFL is used. It is not recommended to change the value of this parameter.}

\item{this_seq_type}{The seq_type of the samples. Only really used to retrieve mutations when maf data is not provided and to be retrieved through GAMBLR. Defaults to genome.}

\item{output}{The output to be returned after prediction is done. Can be one of predictions, matrix, or both. Defaults to predictions.}
}
\value{
data frame with classification, binary matrix used in classification, or both.
}
\description{
Classify FL samples into cFL/dFL subgroups.
}
\details{
Use the random forest prediction model to assemble the binary matrix and use it to classify FL tumors into cFL/dFL.
}
\examples{
test_meta <- get_gambl_metadata(case_set="tFL-study")
predictions = classify_fl(these_samples_metadata = test_meta)
predictions = classify_fl(these_samples_metadata = test_meta, output = "both")

}
