% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanup_maf}
\alias{cleanup_maf}
\title{Cleanup MAF.}
\usage{
cleanup_maf(maf_df)
}
\arguments{
\item{maf_df}{input MAF data frame.}
}
\value{
maf_df with transformed columns
}
\description{
Transform input maf columns to allow for usage of dplyr verbs.
}
\details{
Transform input maf columns to allow for usage of dplyr verbs.
Allowing for a stright-forward plotting workflow as well as downstream data aggregation and manipulation.
}
\examples{
ssm_sample = get_ssm_by_sample(this_sample_id = "HTMCP-01-06-00485-01A-01D",
                               tool_name = "slims-3",
                               projection = "grch37")

clean_maf = cleanup_maf(maf_df = ssm_sample)

}
