% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cnvKompare}
\alias{cnvKompare}
\title{This function adjusts ploidy of the sample using the percent of genome altered (PGA). The PGA is calculated internally, but can also be optionally provided as data frame
if calculated from other sources. Only the samples above the threshold-provided PGA will have ploidy adjusted. The function can work with either individual or
multi-sample seg file. The telomeres are always excluded from calculation, and sex chromosomes can be optionally included or excluded. The supported projections are grch37 and hg38.
The chromosome prefix is handled internally per projection and does not need to be consistent.}
\usage{
cnvKompare(
  patient_id,
  sample_ids,
  this_seg,
  seg_path,
  genes_of_interest,
  projection = "grch37",
  ignore_cytoband_labels = c("acen", "gvar", "stalk"),
  max_overlap = 20,
  exclude_sex = FALSE,
  return_heatmap = TRUE,
  compare_pairwise = TRUE
)
}
\arguments{
\item{patient_id}{Specify patient_id to retrieve sample ids from GAMBL metadata.}

\item{sample_ids}{Optionally, specify sample ids for comparison.}

\item{this_seg}{Optional input data frame of seg file. Must adhere to seg format.}

\item{seg_path}{Optionally, specify the path to a local seg file. Must adhere to seg format.}

\item{genes_of_interest}{Provide specific genes to be displayed on the time-series plot.}

\item{projection}{Argument specifying the projection of seg file, which will determine coordinates of the cytobands. Default is grch37, but hg38 is also accepted.}

\item{ignore_cytoband_labels}{Cytobands to be ignored. By default, "acen", "gvar", "stalk" are excluded.}

\item{max_overlap}{For time-series plot, how many maximum overlapping points are allowed?}

\item{exclude_sex}{Boolean argument specifying whether to exclude sex chromosomes from calculation. Default is FALSE.}

\item{return_heatmap}{Boolean argument specifying whether to return a heatmap of cnvKompare scores. Default is TRUE.}

\item{compare_pairwise}{Boolean argument specifying whether to perform pairwise comparisons is there are more than 2 time points in the group. Default is TRUE.}

\item{pga_cutoff}{Minimum PGA for the sample to adjust ploidy. Default is 0.05 (5\%).}
}
\value{
A list of overall and pairwise percent concordance, concordant and discordant cytobands, comparison heatmap of cnvKompare socres, and time series ggplot object.
}
\description{
This function adjusts ploidy of the sample using the percent of genome altered (PGA). The PGA is calculated internally, but can also be optionally provided as data frame
if calculated from other sources. Only the samples above the threshold-provided PGA will have ploidy adjusted. The function can work with either individual or
multi-sample seg file. The telomeres are always excluded from calculation, and sex chromosomes can be optionally included or excluded. The supported projections are grch37 and hg38.
The chromosome prefix is handled internally per projection and does not need to be consistent.
}
