% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_ashm_results}
\alias{collate_ashm_results}
\title{INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Determine the hypermutation status of a few genes.}
\usage{
collate_ashm_results(
  sample_table,
  seq_type_filter = "genome",
  ssh_session = NULL
)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}
}
\value{
Samples table.
}
\description{
INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Determine the hypermutation status of a few genes.
}
\examples{
sample_table = collate_ashm_results(sample_table=sample_table)

}
