% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_csr_results}
\alias{collate_csr_results}
\title{Collate a few CSR annotations, including MiXCR}
\usage{
collate_csr_results(sample_table)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column}
}
\value{
The sample table with additional columns
}
\description{
Collate a few CSR annotations, including MiXCR
}
