% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_csr_results}
\alias{collate_csr_results}
\title{Collate CSR Results}
\usage{
collate_csr_results(sample_table, seq_type_filter = "genome")
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{seq_type_filter}{Filtering criteria, default is genomes.}
}
\value{
The sample table with additional columns.
}
\description{
Collate a few CSR annotations, including MiXCR.
}
\details{
INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
}
\examples{
\dontrun{
gambl_results_derived = collate_csr_results(gambl_results_derived)
}

}
