% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_curated_sv_results}
\alias{collate_curated_sv_results}
\title{Title}
\usage{
collate_curated_sv_results(
  sample_table,
  path_to_files = "icgc_dart/derived_and_curated_metadata/"
)
}
\arguments{
\item{path_to_files}{}
}
\value{

}
\description{
Title
}
