% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_derived_results}
\alias{collate_derived_results}
\title{Title}
\usage{
collate_derived_results(sample_table, database_name = "gambl_test")
}
\arguments{
\item{sample_table}{}
}
\value{

}
\description{
Title
}
