% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_extra_metadata}
\alias{collate_extra_metadata}
\title{Collate Extra Metadata.}
\usage{
collate_extra_metadata(sample_table, file_path)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{file_path}{Path to extra metadata.}
}
\value{
A table.
}
\description{
Gather additional metadata information and expand the incoming sample table (or metadata).
}
\details{
INTERNAL FUNCTION called by \code{collate_results}, not meant for out-of-package usage.
}
\examples{
table = collate_extra_metadata(sample_table = "my_sample_table.txt")

}
