% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_igblast_results}
\alias{collate_igblast_results}
\title{INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Identify samples with mutated IGHV from MiXCR + IgBLAST results}
\usage{
collate_igblast_results(
  sample_table,
  seq_type_filter = "genome",
  results_directory = "",
  missing_threshold = 2
)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{results_directory}{Directory containging MiXCR/IgBLASTN results. Depends on files having the columns "mutatedStatus" and "numMissingRegions"}

\item{missing_threshold}{Limit for missing IGH regions during MiXCR assembly. Default is 2. Higher thresholds may return less accurate mutation status.}
}
\value{
Samples table.
}
\description{
INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Identify samples with mutated IGHV from MiXCR + IgBLAST results
}
\examples{
sample_table = collate_igblast_results(sample_table = sample_table, results_directory="*/01-mixcr/{seq_type}/", missing_threshold=3)

}
