% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_lymphgen}
\alias{collate_lymphgen}
\title{Collate Lymphgen.}
\usage{
collate_lymphgen(
  these_samples_metadata,
  lymphgen_version = "default",
  tidy = FALSE
)
}
\arguments{
\item{these_samples_metadata}{Optional parameter with meta data filtered for sample_ids of interest. If provided, this function will join lymphgen with this meta data, regardless of tidy TRUE/FALSE.}

\item{lymphgen_version}{Version of selected lymphgen, default is "default".}

\item{tidy}{Boolean parameter, set to TRUE for tidy format (i.e long format with no columns dropped). Default is FALSE, which returns the data in a wide format, keeps both the original Subtype. Prediction and tidied LymphGen values and puts the values from each "flavour" in its own column.}
}
\value{
A df with lymphgen information.
}
\description{
Expand a sample_table (meta data) horizontally with different flavours of lymphgen data.
}
\details{
This function takes a sample table (metadata) and adds different flavours of lymphgen data.
It is possible to call this function with an alreeady subuset metdata table (with samplee IDs of interest) with \code{these_samples_metadata}.
If this done, the fucntion will join the lymphgen data with this tablee. Currently, the only supported \code{lymphgen_version} is "default".
For more information refer to the function examples.
}
\examples{
this_meta = get_gambl_metadata() \%>\% dplyr::filter(pathology == "DLBCL")
wide_lymphgen = collate_lymphgen(these_samples_metadata = this_meta, lymphgen_version = "default", tidy = FALSE)

}
