% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_qc_results}
\alias{collate_qc_results}
\title{Collate Quality Control Results.}
\usage{
collate_qc_results(sample_table, seq_type_filter = "genome")
}
\arguments{
\item{sample_table}{df with sample ids in the first column.}

\item{seq_type_filter}{default is genome, capture is also available for unix_group icgc_dart.}
}
\value{
The sample table with additional columns.
}
\description{
Expand a metadata table horizontally with quality control metrics.
}
\details{
INTERNAL FUNCTION called by \code{collate_results}, not meant for out-of-package usage.
}
\examples{
qc_metrics = collate_qc_results(sample_table = sample_table)

}
