% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_results}
\alias{collate_results}
\title{Bring together all derived sample-level results from many GAMBL pipelines}
\usage{
collate_results(
  sample_table,
  write_to_file = FALSE,
  join_with_full_metadata = FALSE,
  case_set,
  sbs_manipulation = "",
  seq_type_filter = "genome"
)
}
\value{
A table keyed on biopsy_id that contains a bunch of per-sample results from GAMBL
}
\description{
Bring together all derived sample-level results from many GAMBL pipelines
}
\examples{
everything_collated = collate_results(join_with_full_metadata=TRUE)
}
