% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_results}
\alias{collate_results}
\title{Bring together all derived sample-level results from many GAMBL pipelines.}
\usage{
collate_results(
  sample_table,
  write_to_file = FALSE,
  join_with_full_metadata = FALSE,
  these_samples_metadata,
  case_set,
  sbs_manipulation = "",
  seq_type_filter = "genome",
  from_cache = TRUE
)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{write_to_file}{Boolean statement that outputs tsv file if TRUE, default is FALSE.}

\item{join_with_full_metadata}{Join with all columns of meta data, default is FALSE.}

\item{these_samples_metadata}{Optional argument to use a user specified metadata df, overwrites get_gambl_metadata in join_with_full_metadata.}

\item{case_set}{Optional short name for a pre-defined set of cases.}

\item{sbs_manipulation}{Optional variable for transforming sbs values (e.g log, scale).}

\item{seq_type_filter}{Filtering criteria, default is genomes.}

\item{from_cache}{Boolean variable for using cached results, default is TRUE.}
}
\value{
A table keyed on biopsy_id that contains a bunch of per-sample results from GAMBL
}
\description{
Bring together all derived sample-level results from many GAMBL pipelines.
}
\examples{
everything_collated = collate_results(join_with_full_metadata = TRUE)

}
