% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_sbs_results}
\alias{collate_sbs_results}
\title{INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Bring in the results from mutational signature analysis.}
\usage{
collate_sbs_results(
  sample_table,
  seq_type_filter = "genome",
  file_path,
  scale_vals = FALSE,
  sbs_manipulation = ""
)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{file_path}{Optional path to SBS file.}

\item{scale_vals}{Parameter not used?}

\item{sbs_manipulation}{Optional variable for transforming sbs values (e.g log, scale).}
}
\value{
A data frame with new columns added.
}
\description{
INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Bring in the results from mutational signature analysis.
}
\examples{
collated = collate_sbs_results(sample_table=sample_table,sbs_manipulation=sbs_manipulation)

}
