% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_ssm_results}
\alias{collate_ssm_results}
\title{Compute some summary statistics based on SSM calls}
\usage{
collate_ssm_results(sample_table, from_flatfile = TRUE, include_silent = FALSE)
}
\arguments{
\item{from_flatfile}{Optional argument whether to use database or flat file to retrieve mutations}

\item{include_silent}{Logical parameter indicating whether to include siment mutations into coding mutations. Default is FALSE}
}
\value{

}
\description{
Compute some summary statistics based on SSM calls
}
