% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_sv_results}
\alias{collate_sv_results}
\title{Title}
\usage{
collate_sv_results(
  sample_table,
  tool = "manta",
  oncogenes = c("MYC", "BCL2", "BCL6", "CCND1", "IRF4")
)
}
\arguments{
\item{oncogenes}{}
}
\value{

}
\description{
Title
}
