% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collate_sv_results}
\alias{collate_sv_results}
\title{INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Determine and summarize which cases have specific oncogene SVs.}
\usage{
collate_sv_results(
  sample_table,
  tool = "manta",
  seq_type_filter = "genome",
  oncogenes = c("MYC", "BCL2", "BCL6", "CCND1", "IRF4")
)
}
\arguments{
\item{sample_table}{A data frame with sample_id as the first column.}

\item{tool}{Name of tool (optional, default is manta).}

\item{oncogenes}{Which oncogenes to collate SVs from.}
}
\value{
Data frame with additional columns ({tool}\emph{{oncogene} and {tool}}{oncogene}_{partner}).
}
\description{
INTERNAL FUNCTION called by collate_results, not meant for out-of-package usage.
Determine and summarize which cases have specific oncogene SVs.
}
\examples{
results = collate_samples_sv_results(samples, "manta", c("MYC", "BCL2"))

}
