% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{comp_report}
\alias{comp_report}
\title{Construct pdf with sample-level plots, using minimum of arguments}
\usage{
comp_report(
  this_sample,
  export_individual_plots = FALSE,
  out,
  seq_data,
  seq_path = NULL,
  maf_data,
  maf_path = NULL
)
}
\arguments{
\item{this_sample}{Sample ID to be plotted in report.}

\item{export_individual_plots}{Boolean parameter, set to TRUE to export individual plots.}

\item{out}{Path to output folder.}

\item{seq_data}{Optional parameter with copy number df already loaded into R.}

\item{seq_path}{Optional parameter with path to external cn file.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}
}
\value{
Nothing.
}
\description{
Construct pdf with sample-level plots, using minimum of arguments
}
\examples{
comp_report(this_sample = "HTMCP-01-06-00422-01A-01D", out = "reports/", export_individual_plots = TRUE)

}
