% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{comp_report}
\alias{comp_report}
\title{sample-level SV/SSM/CN reports in PDF}
\usage{
comp_report(
  this_sample_id,
  export_individual_plots = FALSE,
  out,
  seq_data,
  seq_path = NULL,
  maf_data,
  maf_path = NULL,
  this_seq_type = "genome"
)
}
\arguments{
\item{this_sample_id}{Sample ID to be plotted in report.}

\item{export_individual_plots}{Boolean parameter, set to TRUE to export individual plots.}

\item{out}{Path to output folder.}

\item{seq_data}{Optional parameter with copy number df already loaded into R.}

\item{seq_path}{Optional parameter with path to external cn file.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{this_seq_type}{Seq type for returned CN segments. One of "genome" (default) or "capture".}
}
\value{
Nothing.
}
\description{
Construct pdf with sample-level plots, using minimum of arguments
}
\details{
This function runs the complete collection of \code{fancy_x_plots} for a specific sample ID (\code{this_sample}), with default parameters.
The generated plots are put together into a two-page PDF. In addition, it is also possible to export all individual plots.
This can be done by setting \code{export_individual_plots} to TRUE. It is also possible to use an already loaded seq file instead of using the
\code{this_sample_id} parameter, this is done with the \code{seq_data} and \code{maf_data} parameters. Similarly, you can also point this function to a local
file on disk with the \code{seq_path} and \code{maf_path} parameters.
}
\examples{
\dontrun{
#create a PDF report for one sample, as well as exporting all individual plots.
comp_report(this_sample = "HTMCP-01-06-00422-01A-01D",
            out = "reports/",
            export_individual_plots = TRUE)
}

}
