% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{compare_mutation_flavour}
\alias{compare_mutation_flavour}
\title{Compare Mutation Flavour.}
\usage{
compare_mutation_flavour(
  these_sample_ids,
  flavour1 = "clustered",
  flavour2 = ""
)
}
\arguments{
\item{these_sample_ids}{A vector of sample IDs to be included.}

\item{flavour1}{First flavour of variant calls, to be returned as unique if not present in flavour2. Default is "clustered".}

\item{flavour2}{Second flavour of variant calls.}
}
\value{
a list with MAFs that are only present in flavour1.
}
\description{
Get a MAF that is just the variants unique to one of two flavours of variant calls available.
}
\details{
Subset a MAF to only have variants that are unique to one flavour (specified with \code{flavour1}).
This function is currently not exported, since there is only one flavour available at the moment (see docs for \code{get_ssm_by_sample}).
}
