% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{copy_number_vaf_plot}
\alias{copy_number_vaf_plot}
\title{Create a genome-wide copy number plot for one sample and (optionally) display mutation VAF.}
\usage{
copy_number_vaf_plot(
  this_sample_id,
  just_segments = FALSE,
  coding_only = FALSE,
  one_chrom,
  genes_to_label,
  from_flatfile = FALSE,
  use_augmented_maf = FALSE,
  add_chr_prefix = FALSE
)
}
\arguments{
\item{this_sample_id}{The sample_id for the sample to plot.}

\item{just_segments}{Specify whether only the segments will be plotted (instead of mutation VAF).}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{one_chrom}{Subset plot to one chromosome.}

\item{genes_to_label}{Optional. Provide a list of genes to label (if mutated). Can only be used with coding_only (see above).}

\item{from_flatfile}{If set to true the function will use flatfiles instead of the database.}

\item{use_augmented_maf}{Boolean statement if to use augmented maf, default is FALSE.}

\item{add_chr_prefix}{If TRUE, "chr" prefix will be added to chr column. Default is FALSE.}
}
\value{
Nothing
}
\description{
Create a genome-wide copy number plot for one sample and (optionally) display mutation VAF.
}
\examples{
cnv_vaf_plot = copy_number_vaf_plot(this_sample_id = "some-sample-name",
                                    just_segments = TRUE,
                                    coding_only = FALSE,
                                    one_chrom = "chr2",
                                    genes_to_label = "MYC",
                                    from_flatfile = FALSE,
                                    use_augmented_maf = FALSE,
                                    add_chr_prefix = TRUE)

}
