% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{copy_number_vaf_plot}
\alias{copy_number_vaf_plot}
\title{Create a genome-wide copy number plot for one sample and (optionally) display mutation VAF}
\usage{
copy_number_vaf_plot(
  this_sample,
  just_segments = FALSE,
  coding_only = FALSE,
  one_chrom,
  genes_to_label,
  from_flatfile = FALSE,
  use_augmented_maf = FALSE
)
}
\arguments{
\item{this_sample}{The sample_id for the sample to plot}

\item{just_segments}{Specify whether only the segments will be plotted (instead of mutation VAF)}

\item{coding_only}{optional. Set to TRUE to restrict to plotting only coding mutations}

\item{genes_to_label}{optional. Provide a list of genes to label (if mutated). Can only be used with coding_only (see below)}

\item{from_flatfile}{If set to true the function will use flatfiles instead of the database}
}
\value{
nothing
}
\description{
Create a genome-wide copy number plot for one sample and (optionally) display mutation VAF
}
