% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{count_ssm_by_region}
\alias{count_ssm_by_region}
\title{Count the variants in a region with a variety of filtering options}
\usage{
count_ssm_by_region(
  region,
  chromosome,
  start,
  end,
  all_mutations_in_these_regions,
  these_samples_metadata,
  count_by,
  seq_type = "genome"
)
}
\arguments{
\item{region}{}

\item{chromosome}{}

\item{start}{}

\item{end}{}

\item{all_mutations_in_these_regions}{If you are calling this function many times (e.g. bins spanning a larger region), to save a ton of time you are strongly encouraged to provide the output of get_ssm_by_region on the entire region of interest and passing it to this function}

\item{these_samples_metadata}{}

\item{count_by}{Defaults to counting all variants. Specify 'sample_id' if you want to collapse and count only one per sample}

\item{seq_type}{}
}
\description{
Count the variants in a region with a variety of filtering options
}
