% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{determine_version}
\alias{determine_version}
\title{Check for a version of data to load}
\usage{
determine_version(
  mode = "somatic_hypermutation_locations",
  this_genome_build = "grch37"
)
}
\arguments{
\item{auto_connect}{Set to TRUE to ensure an ssh_session is created if absent}
}
\value{
data frame
}
\description{
This function determines if a user is requesting the latest version
of the bundled data or wants to access one of the earlier veresions.
}
