% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_cnbar}
\alias{fancy_cnbar}
\title{Generate a bar plot visualizing sample-specific copy number states and affected bases for each CN segment.}
\usage{
fancy_cnbar(
  this_sample,
  plot_title = paste0(this_sample),
 
    plot_subtitle = "n CNV Segments (barplots, left y-axis), n Affected bases for each CN state",
  chr_select = paste0("chr", c(1:22)),
  cutoff = 160,
  include_cn2 = TRUE
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{Vector of chromosomes to be included in plot, defaults to autosomes.}

\item{cutoff}{Set threshold for maximum CN state to be retrieved.}

\item{include_cn2}{Optional boolean statement for including CN = 2 states in plot.}
}
\value{
Nothing.
}
\description{
Generate a bar plot visualizing sample-specific copy number states and affected bases for each CN segment.
}
\examples{
chr1_cns = fancy_cnbar(this_sample = "HTMCP-01-06-00422-01A-01D", chr_select = c(1))
cns = fancy_cnbar(this_sample = "HTMCP-01-06-00422-01A-01D")

}
