% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_ideogram}
\alias{fancy_ideogram}
\title{genome-wide ideogram annotated with SSM and CN information}
\usage{
fancy_ideogram(
  this_sample_id,
  gene_annotation,
  seq_data,
  seq_path = NULL,
  maf_data,
  maf_path = NULL,
  variant_type_col_maf = 10,
  chromosome_col_maf = 5,
  start_col_maf = 6,
  end_col_maf = 7,
  chrom_col_seq = 2,
  start_col_seq = 3,
  end_col_seq = 4,
  cn_col_seq = 7,
  plot_title = paste0(this_sample_id),
  plot_subtitle = "Genome-wide Ideogram (grch37).",
  intersect_regions,
  include_ssm = TRUE,
  ssm_count = TRUE,
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample_id}{Sample to be plotted (for multiple samples, see fancy_multisample_ideogram.}

\item{gene_annotation}{Annotate ideogram with a set of genes. These genes can either be specified as a vector of characters or a data frame.}

\item{seq_data}{Optional parameter with copy number df already loaded into R.}

\item{seq_path}{Optional parameter with path to external cn file.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{variant_type_col_maf}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col_maf}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{start_col_maf}{Index of column with variant start coordinates (to be used with either maf_data or maf_path).}

\item{end_col_maf}{Index of column with variant end coordinates (to be used with either maf_data or maf_path).}

\item{chrom_col_seq}{Index of column annotating Chromosome (to be used with either maf_data or maf_path).}

\item{start_col_seq}{Index of column with copy number start coordinates (to be used with either maf_data or maf_path).}

\item{end_col_seq}{Index of column with copy number end coordinates (to be used with either maf_data or maf_path).}

\item{cn_col_seq}{Index of column holding copy number information (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Optional argument for plot subtitle.}

\item{intersect_regions}{Optional parameter for subset variant calls to specific regions. Should be either a vector of characters (chr:start-end) or data frame with regions.}

\item{include_ssm}{Set to TRUE to plot ssms (dels and ins).}

\item{ssm_count}{Optional parameter to summarize n variants per chromosome, inlcude_ssm must be set to TRUE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented_maf}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate sample-level ideogram with copy number information, ssm and gene annotations, etc.
}
\details{
This function generates genome-wide ideograms, visualizing SSM data as well as CN segments.
It is also possible to superimpose the plot with gene annotations. Offering a comprehenssive overview of all SSM and CN segments of different anneuploidity.
The plotting of SSM can be toggled with setting \code{include_ssm} to TRUE. If so, it is also possible to count the number of SSMs per chromosome with \code{ssm_count = TRUE}.
To get data for plotting, there are a few different options avaialble; like all \code{fanncy_x_plots} a sample ID can be provided to the \code{this_sample}
parameter. If done so, the function will retreive data (SSm and CN segments) by wrapping the appropriate functions.
This data can also be provided with \code{seq_data}, \code{seg_path}, \code{maf_data} and \code{maf_path}.
For more info on how to run with these paraemters, refer to the parameter descriptions.
In order to annotate the ideogram with genes, simply give the \code{gene_annotations} parameter a set of genes as a vector of charachters or a data frame with gene names in the first column.
Another useful parameter for restricting the plotted regions is to call the function with \code{intersect_regions}.
This parameter takes a vector of charachters or data frame with regions that the plotted calls are restricted to.
}
\examples{
#get FL genes
fl_genes = dplyr::filter(lymphoma_genes, FL == TRUE) \%>\%
  dplyr::select(Gene) \%>\%
  pull(Gene)

#get FL genes on chromosome 1
fl_genes_chr1 = gene_to_region(gene_symbol = fl_genes, return_as = "df") \%>\%
  dplyr::filter(chromosome == "1") \%>\%
  pull(hugo_symbol)

#build an ideogram showing ssm and copy number states for chromosome 1, and superimpose with FL genes.
ideogram_fl_chr1 = fancy_ideogram(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                                  gene_annotation = fl_genes_chr1,
                                  intersect_regions = "chr1:10000-249250621")

#
fl_regions = gene_to_region(gene_symbol = fl_genes, return_as = "df")
ideogram_fl = fancy_ideogram(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                              gene_annotation = fl_genes,
                              intersect_regions = fl_regions)

}
