% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_ideogram}
\alias{fancy_ideogram}
\title{Generate sample-level ideogram with copy number information, ssm and gene annotations, etc.}
\usage{
fancy_ideogram(
  this_sample,
  gene_annotation,
  seq_data,
  seq_path = NULL,
  maf_data,
  maf_path = NULL,
  variant_type_col_maf = 10,
  chromosome_col_maf = 5,
  start_col_maf = 6,
  end_col_maf = 7,
  chrom_col_seq = 2,
  start_col_seq = 3,
  end_col_seq = 4,
  cn_col_seq = 7,
  plot_title = paste0(this_sample),
  plot_subtitle = "Genome-wide Ideogram (grch37).",
  intersect_regions,
  include_ssm = TRUE,
  ssm_count = TRUE,
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample}{Sample to be plotted (for multiple samples, see fancy_multisample_ideogram.}

\item{gene_annotation}{Annotate ideogram with a set of genes. These genes can either be specified as a vector of characters or a data frame.}

\item{seq_data}{Optional parameter with copy number df already loaded into R.}

\item{seq_path}{Optional parameter with path to external cn file.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{variant_type_col_maf}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col_maf}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{start_col_maf}{Index of column with variant start coordinates (to be used with either maf_data or maf_path).}

\item{end_col_maf}{Index of column with variant end coordinates (to be used with either maf_data or maf_path).}

\item{chrom_col_seq}{Index of column annotating Chromosome (to be used with either maf_data or maf_path).}

\item{start_col_seq}{Index of column with copy number start coordinates (to be used with either maf_data or maf_path).}

\item{end_col_seq}{Index of column with copy number end coordinates (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Optional argument for plot subtitle.}

\item{intersect_regions}{Optional parameter for subset variant calls to specific regions. Should be either a vector of characters (chr:start-end) or data frame with regions.}

\item{include_ssm}{Set to TRUE to plot ssms (dels and ins).}

\item{ssm_count}{Optional parameter to summarize n variants per chromosome, inlcude_ssm must be set to TRUE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{cn_col}{Index of column holding copy number information (to be used with either maf_data or maf_path).}

\item{use_augmented}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate sample-level ideogram with copy number information, ssm and gene annotations, etc.
}
\examples{
#
fl_genes = dplyr::filter(lymphoma_genes, FL == TRUE) \%>\%
  dplyr::select(Gene) \%>\%
  pull(Gene)

fl_genes_chr1 = gene_to_region(gene_symbol = fl_genes, return_as = "df") \%>\%
  dplyr::filter(chromosome == "1") \%>\%
  pull(hugo_symbol)

ideogram_fl_chr1 = fancy_ideogram(this_sample = "HTMCP-01-06-00422-01A-01D",
                                  gene_annotation = fl_genes_chr1, 
                                  intersect_regions = "chr1:10000-249250621",
                                  include_ssm = TRUE, 
                                  ssm_count = TRUE, 
                                  coding_only = FALSE, 
                                  from_flatfile = FALSE, 
                                  use_augmented = FALSE)
 
 fl_regions = gene_to_region(gene_symbol = fl_genes, return_as = "df")
 ideogram_fl = fancy_ideogram(this_sample = "HTMCP-01-06-00422-01A-01D",
                              gene_annotation = fl_genes, 
                              intersect_regions = fl_regions,
                              include_ssm = TRUE, 
                              ssm_count = TRUE, 
                              coding_only = FALSE, 
                              from_flatfile = FALSE, 
                              use_augmented = FALSE)

}
