% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_ideogram}
\alias{fancy_ideogram}
\title{Generate sample-level ideogram with SVs and gene annotations}
\usage{
fancy_ideogram(
  this_sample,
  gene_annotation,
  plot_title = paste0(this_sample),
  plot_subtitle = "Genome-wide Ideogram (grch37).",
  include_sv = TRUE,
  sv_count = TRUE,
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample}{Sample to be plotted (currently only accepts one sample).}

\item{gene_annotation}{Annotate ideogram with a single gene.}

\item{plot_title}{Title of plot (default to sample ID).}

\item{include_sv}{Set to TRUE to plot SVs (dels and ins).}

\item{sv_count}{Optional parameter to summarize n variants per chromosome, inlcude_sv must be set to TRUE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate sample-level ideogram with SVs and gene annotations
}
\examples{
ideogram_MYC = fancy_ideogram(this_sample = "HTMCP-01-06-00422-01A-01D", gene_annotation = "MYC", include_sv = TRUE, sv_count = TRUE, coding_only = FALSE, from_flatfile = FALSE, use_augmented = FALSE)
cnv_ideogram = fancy_ideogram(this_sample = "HTMCP-01-06-00422-01A-01D")

}
