% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_multisamp_ideogram}
\alias{fancy_multisamp_ideogram}
\title{Genome-wide ideogram (CN segments) for multiple samples}
\usage{
fancy_multisamp_ideogram(
  these_sample_ids,
  plot_title = "CN Segments Ideogram",
  plot_sub = "grch37",
  chr_anno_dist = 3,
  chr_select = paste0("chr", c(1:22)),
  include_cn2 = FALSE,
  kompare = FALSE,
  concordance = TRUE,
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{these_sample_ids}{Sample to be plotted (accepts 2, 3 or 4 samples).}

\item{plot_title}{Main title of plot.}

\item{plot_sub}{Subtitle of plot.}

\item{chr_anno_dist}{Optional parameter to adjust chromosome annotations, default value is 3, increase to adjust annotations to left.}

\item{chr_select}{Optional parameter to subset plot to specific chromosomes. Default value is chr1-22.}

\item{include_cn2}{Set to TRUE for plotting CN states == 2.}

\item{kompare}{Boolean statement, set to TRUE to call cnvKompare on the selected samples for plotting concordant (or discordant) cn segments across selected chromosomes.}

\item{concordance}{Boolean parameter to be used when kompare = TRUE. Default is TRUE, to plot discordant segments, set parameter to FALSE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented_maf}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
Nothing.
}
\description{
Generate ideograms for selected sample, visualizing copy number variation segments.
}
\details{
To create multi-sample ideograms, i.e showing CN segments across multiple samples, this function was created.
This can be used to infer inheritance patterns, hotspots, etc. across multiple samples or the sample ID but for different timepoints.
In addition, this plot can also allow to only plot concordant (or discordant) cn segments between two samples. i.e how two samples differ, or are a like.
The function automatically detects the number of samples provided and sets the plotting parameters accordingly.
The maximun number of samples this plot can deal with is 4 and sample IDs of interest are given to the \code{these_sample_ids}.
In order to only plot segments that are concordant between the selected samples, set \code{komapre} to TRUE and \code{concordance} to TRUE.
To instead plot discordant CN segments, set this parameter to FALSE.
}
\examples{
#two samples ideogram
ideo_2_samp = fancy_multisamp_ideogram(these_sample_ids = c("00-15201_tumorA", "00-15201_tumorB"),
                                       chr_anno_dist = 4)

}
