% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_snv_chrdistplot}
\alias{fancy_snv_chrdistplot}
\title{n SNVs per chromosome plot}
\usage{
fancy_snv_chrdistplot(
  this_sample_id,
  maf_data,
  maf_path = NULL,
  variant_type_col = 10,
  chromosome_col = 5,
  plot_title = paste0(this_sample_id),
  plot_subtitle = "SNV Distribution Per Chromosome",
  chr_select = paste0("chr", c(1:22)),
  include_dnp = FALSE,
  hide_legend = FALSE,
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample_id}{Sample to be plotted.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{include_dnp}{Optional argument for including DNPs. Default is FALSE.}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented_maf}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Visualizing the number of SNVs per chromosome.
}
\details{
This function takes on an already loaded maf-like data frame, or a path to the maf file of interest.
In addition, the user can also give this function a sample ID and the function will run assign_cn_to_ssm
to get data for plotting. If a maf file or data frame is used, the user has the chance to specify what column
that holds the Variant Type information (\code{variant_type_col}), in addition the user can also specify what column
in the incoming maf that is corresponding to the chromosome annotations. This function also includes useful subsetting
options. For example, \code{chr_select} allows the user to restrict the plot to specific chromosomes. \code{include_dnp} is an optional
argument (Boolean) for if variants of this subtype should be included or not. The plot can also be restricted to only
counting coding mutations (\code{coding_only}). Flat-file and augmented maf options can be toggled with \code{from_flatfile}
and \code{use_augmented_maf}. Both are TRUE by default and should rarely be set to FALSE. Lastly, this plotting function
also have convenient parameters for customizing the returned plot, e.g \code{plot_title}, \code{y_interval}, \code{hide_legend}, and\code{plot_subtitle}.
}
\examples{
#plot SNVs
fancy_snv_chrdistplot(this_sample_id = "HTMCP-01-06-00422-01A-01D")

#plot SNVs and DNPs
fancy_snv_chrdistplot(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                      include_dnp = TRUE,
                      plot_subtitle = "SNV + DNP Distribution Per Chromosome")

}
