% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_sv_size_plot}
\alias{fancy_sv_size_plot}
\title{Generate plot visualizing SV sizes retreived with get_combined_sv. Subset on variant type, filter on VAF, size etc.}
\usage{
fancy_sv_size_plot(
  this_sample,
  maf_data,
  maf_path = NULL,
  chrom_a_col = 3,
  start_a_col = 4,
  end_a__col = 5,
  variant_type_col = 9,
  vaf_cutoff = 0,
  size_cutoff = 50,
  chr_select = paste0("chr", c(1:22)),
  plot_title = paste0(this_sample),
 
    plot_subtitle = paste0("SV sizes for Manta calls. Dashed line annotates mean variant size.\\nVAF cut off: ",
    vaf_cutoff, ", SV size cut off: ", size_cutoff),
  projection = "grch37"
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{maf_data}{Optional parameter with copy number df already loaded into R.}

\item{maf_path}{Optional parameter with path to external cn file.}

\item{chrom_a_col}{Index of column holding chromosome (to be used with either maf_data or maf_path).}

\item{start_a_col}{Index of column holding start coordiantes (to be used with either maf_data or maf_path).}

\item{variant_type_col}{Index of column holding variant type information (to be used with either maf_data or maf_path).}

\item{vaf_cutoff}{Threshold for filtering variants on VAF (events with a VAF > cutoff will be retained).}

\item{size_cutoff}{Threshold for filtering variants on size, default is 50bp.}

\item{chr_select}{Optional argument for subsetting on selected chromosomes, default is all autosomes.}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{projection}{Genomic projection for SVs and circos plot. Accepted values are grch37 and hg38.}

\item{end_a_col}{Index of column holding end coordinates (to be used with either maf_data or maf_path).}
}
\value{
Nothing.
}
\description{
Generate plot visualizing SV sizes retreived with get_combined_sv. Subset on variant type, filter on VAF, size etc.
}
\examples{
myplot = fancy_sv_size_plot(this_sample = "HTMCP-01-06-00422-01A-01D")
myplot2 = fancy_sv_size_plot(this_sample = "HTMCP-01-06-00422-01A-01D", size_cutoff = 0, chr_select = c("chr1", "chr2"))

}
