% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_v_chrcount}
\alias{fancy_v_chrcount}
\title{n variants per chromosome plot.}
\usage{
fancy_v_chrcount(
  this_sample_id,
  maf_data,
  maf_path = NULL,
  ssm = TRUE,
  projection = "grch37",
  min_vaf = 0,
  variant_type_col = 10,
  chromosome_col = 5,
  plot_title = paste0(this_sample_id),
  y_interval = 1,
  hide_legend = FALSE,
  plot_subtitle = "Variant Count Distribution Per Chromosome",
  chr_select = paste0("chr", c(1:22)),
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE,
  add_qc_metric = FALSE,
  seq_type = "genome"
)
}
\arguments{
\item{this_sample_id}{Sample to be plotted.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{ssm}{Set to FALSE to get plotting data from \link{get_combined_sv} (SVs). Default value is TRUE (plots SSM retrieved from annotate_cn_by_ssm$maf)}

\item{projection}{Genome build for returned variants (only applicable for ssm = FALSE)}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Recommended: 0 (only applicable for ssm = FALSE).}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{y_interval}{Optional parameter for specifying intervals on y-axis.}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented_maf}{Boolean statement if to use augmented maf, default is FALSE.}

\item{add_qc_metric}{Boolean statement, if set to TRUE specified QC metric will be added (second y-axis).}

\item{seq_type}{Default is "genome".}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Visualizing variant (SSM or SVs) counts per chromosome.
}
\details{
Takes a maf data frame (or path to a maf), counts the number of variants per chromosome.
Selected chromosomes (\code{chr_select}) are plotted along the x-axis and the variant counts are represented on the y-axis.
This function can plot both Structural Variants (SV) and Simple Shared Motifs (SSM).
It plots SVs per default and SSM can be added with setting \code{ssm = TRUE}.
This plot can also be restricted to only show coding mutations. To do so, set \code{coding_only} to TRUE.
In addition, the returned plot can also be superimposed with a sample-specific mean coverage (from \link{collate_results}).
To do so, set \code{add_qc_metric} to TRUE. A collection of parameters for customizing the returned plot are also available.
e.g \code{plot_title}, \code{y_interval}, \code{hide_legend}, and \code{plot_subtitle}.
}
\examples{
#plot ssm
fancy_v_chrcount(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                 ssm = TRUE)

#plot SVs for chr 1-5
fancy_v_chrcount(this_sample_id = "HTMCP-01-06-00422-01A-01D",
                 ssm = FALSE,
                 min_vaf = 0,
                 projection = "grch37",
                 chr_select = paste0("chr", c(1:5)),
                 plot_subtitle = "SV Count Distribution (chr1-5)")

}
