% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_v_count}
\alias{fancy_v_count}
\title{Total n variants count plot.}
\usage{
fancy_v_count(
  this_sample_id,
  maf_data,
  maf_path = NULL,
  ssm = TRUE,
  projection = "grch37",
  this_seq_type = "genome",
  min_vaf = 0,
  variant_type_col = 10,
  chromosome_col = 5,
  plot_title = paste0(this_sample_id),
  plot_subtitle = "Variant Count For Selected Contigs",
  chr_select = paste0("chr", c(1:22)),
  variant_select = c("DEL", "INS", "DUP"),
  snp_colours = c(SNP = "#2B9971", DNP = "#993F2B", TNP = "#A62656"),
  hide_legend = FALSE,
  coding_only = FALSE,
  log10_y = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample_id}{Sample to be plotted.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{ssm}{Set to FALSE to get plotting data from get_combined_sv (SVs). Default value is TRUE (plots SSM retrieved from annotate_cn_by_ssm$maf).}

\item{projection}{Genome build for returned variants (only applicable for ssm = FALSE).}

\item{this_seq_type}{Seq type for returned CN segments. One of "genome" (default) or "capture".}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Recommended: 0 (only applicable for ssm = FALSE).}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{variant_select}{Subtypes of SVs to be included in plot, default is DEL, INS and DUP.}

\item{snp_colours}{Optional vector with colours for SNPs (DNP and TNP).}

\item{hide_legend}{Set to True to remove legend from plot, default is FALSE.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{log10_y}{Set to TRUE to force y axis to be in log10.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{use_augmented_maf}{Boolean statement if to use augmented maf, default is TRUE.}
}
\value{
A plot as a ggplot object (grob).
}
\description{
Generate a bar plot visualizing total variant (SSM or SVs) count for selected contigs.
}
\details{
This function creates a barplot showing the total number of variants for a selected sample.
Convenience parameters for restricting the returned plot are available. For example, with \code{ssm} (Boolean)
you can toggle if the plot will be in respect to SSM (\code{ssm = TRUE}) or if you wish to count SVs (\code{ssm = FALSE}).
In addition, this plot can also accept a variety of incoming data types. Either, you supply the function with a sample ID
(\code{this_sample_id}) and the function will retrieve data using \link{assign_cn_to_ssm} or \link{get_combined_sv} (depending on how the \code{ssm} parameter is used).
This function also supports a maf or maf-like data frame directly, this is done with \code{maf_data} or \code{maf_path}. If data is supplied with either of these parameters,
the user can specify what column holds the variant type information as well as chromosome information (\code{variant_type_col} and \code{chromosome_col}).
Restricting the plot to coding mutations is done with \code{coding_only = TRUE}. Flat-file and augmented maf options can be toggled with \code{from_flatfile}
and \code{use_augmented_maf}. Both are TRUE by default and should rarely be set to FALSE. Lastly, this plotting function also have convenient parameters for
customizing the returned plot, e.g \code{plot_title}, \code{y_interval}, \code{hide_legend}, and\code{plot_subtitle} and \code{snp_colours}. lastly, it is also possible
to control what variants are to be counted with \code{variant_select}. Default is deletions, insertions and duplications, c("DEL", "DUP", "INS"). Not that
the variant types specified in this parameter must match with whatever is present in the corresponding \code{variant_type_col}.
}
\examples{
#count all variants for one sample (default parameters)
fancy_v_count(this_sample_id = "HTMCP-01-06-00422-01A-01D")

#count and plot all variants on chromosome 1
fancy_v_count(this_sample_id = "HTMCP-01-06-00422-01A-01D",
              chr_select = c("chr1"))

}
