% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{fancy_v_sizedis}
\alias{fancy_v_sizedis}
\title{Generate a violine plot showing variant (SSM or SVs) size distributions for selected contigs.}
\usage{
fancy_v_sizedis(
  this_sample,
  maf_data,
  maf_path = NULL,
  ssm = TRUE,
  projection = "grch37",
  min_vaf = 0,
  variant_type_col = 10,
  chromosome_col = 5,
  start_col = 6,
  end_col = 7,
  plot_title = paste0(this_sample),
  plot_subtitle = "Variant Size Distribution",
  scale_value = "width",
  log_10 = TRUE,
  plot_trim = FALSE,
  chr_select = paste0("chr", c(1:22)),
  coding_only = FALSE,
  from_flatfile = TRUE,
  use_augmented_maf = TRUE
)
}
\arguments{
\item{this_sample}{Sample to be plotted.}

\item{maf_data}{Optional parameter with maf like df already loaded into R.}

\item{maf_path}{Optional parameter with path to external maf like file.}

\item{ssm}{Set to FALSE to get plotting data from get_combined_sv (SVs). Default value is TRUE (plots SSM retrieved from annotate_cn_by_ssm$maf).}

\item{projection}{Genome build for returned variants (only applicable for ssm = FALSE).}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Recommended: 0 (only applicable for ssm = FALSE).}

\item{variant_type_col}{Index of column holding Variant Type (to be used with either maf_data or maf_path).}

\item{chromosome_col}{Index of column holding Chromosome (to be used with either maf_data or maf_path).}

\item{start_col}{Index of column with variant start coordinates (to be used with either maf_data or maf_path).}

\item{end_col}{Index of column with variant end coordinates (to be used with either maf_data or maf_path).}

\item{plot_title}{Title of plot (default to sample ID).}

\item{plot_subtitle}{Subtitle for created plot.}

\item{scale_value}{Scale type for violin plot, accepted values are "area", "width", and "count", defualt is "count.}

\item{log_10}{Boolean statement for yaxis, default is TRUE.}

\item{chr_select}{vector of chromosomes to be included in plot, defaults to autosomes.}

\item{coding_only}{Optional. Set to TRUE to restrict to plotting only coding mutations.}

\item{from_flatfile}{If set to true the function will use flat files instead of the database.}

\item{trim}{Boolean statment for trimming violin plot. Default is TRUE.}

\item{use_augmented}{Boolean statement if to use augmented maf, default is FALSE.}
}
\value{
plot as ggplot object.
}
\description{
Generate a violine plot showing variant (SSM or SVs) size distributions for selected contigs.
}
\examples{
plot = fancy_v_sizedis(this_sample = "HTMCP-01-06-00422-01A-01D")

}
