% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{fetch_output_files}
\alias{fetch_output_files}
\title{Fetch Output Files.}
\usage{
fetch_output_files(
  tool,
  base_path,
  results_dir = "99-outputs",
  seq_type = "genome",
  build = "hg38",
  search_pattern = "cellularity_ploidy.txt"
)
}
\arguments{
\item{tool}{Name of tool.}

\item{base_path}{Either the full or relative path to where all the results directories are for the tool e.g. "gambl/sequenza_current".}

\item{results_dir}{Directory with results.}

\item{seq_type}{Either genome or capture.}

\item{build}{Default is hg38.}

\item{search_pattern}{File-extensions search pattern.}
}
\value{
A data frame with one row per file and sample IDs parsed from the file name along with other GAMBL wildcards.
}
\description{
Get output files from a set of conditions.
}
\details{
This function lets the user specify multiple conditions for returning result subsets.
First, specify the name of the tool with \code{tool}, then set the seq type (\code{seq_type}) to either genome or capture,
together with the genome build (\code{genome_build}). A data frame will be returned with one row per file and sample IDs together with GAMBL wildcards.
}
\examples{

ex_outs = fetch_output_files(tool = "manta", base_path = "gambl/sequenza_current", seq_type = "capture", build = "hg38")

}
