% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{fetch_output_files}
\alias{fetch_output_files}
\title{Title}
\usage{
fetch_output_files(
  tool,
  unix_group,
  base_path,
  results_dir = "99-outputs",
  seq_type = "genome",
  build = "hg38",
  search_pattern = "cellularity_ploidy.txt",
  ssh_session
)
}
\arguments{
\item{base_path}{Either the full or relative path to where all the results directories are for the tool e.g. "gambl/sequenza_current"}

\item{results_dir}{directory with results}

\item{seq_type}{either genome or capture}

\item{search_pattern}{file-extensions search pattern}

\item{tool_name}{name of tool}

\item{genome_build}{default is hg38}
}
\value{
A data frame with one row per file and sample IDs parsed from the file name along with other GAMBL wildcards
}
\description{
Title
}
