% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{finalize_study}
\alias{finalize_study}
\title{Finalize Study (cBioPortal).}
\usage{
finalize_study(
  seq_type_filter = "genome",
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  cancer_type = "mixed",
  these_sample_ids,
  overwrite = TRUE,
  metacols,
  metacol_names,
  metacol_types,
  metacol_prior,
  out_dir
)
}
\arguments{
\item{seq_type_filter}{the seq type you are setting up a study for, default is "genome".}

\item{short_name}{A concise name for your portal project.}

\item{human_friendly_name}{A slightly more verbose name for your project.}

\item{project_name}{Unique ID for your project.}

\item{description}{A verbose description of your data set.}

\item{cancer_type}{Cancer types included in study, default is "mixed".}

\item{these_sample_ids}{A vector of all the sample_id that were included in any of the data files for cBioPortal (i.e the output from \code{setup_study} and \code{setup_fusions}).}

\item{overwrite}{Flag to specify that files should be overwritten if they exist. Default is TRUE.}

\item{metacols}{Optional, specify any additional metadata/collate_result fields to be included in the cBioPortal metadata. Extra columns are specified as a vector of characters. If not provided, only "standard" metadata fields will be kept.}

\item{metacol_names}{Separately specify the names of the columns given to \code{metacols} as a vector of characters. The number of elements needs to match the total number of columns specified with \code{metacols}. Required parameter if \code{metacols} is being called.}

\item{metacol_types}{Specify the data type for selected metadata columns as a vector of characters. The number of elements specified needs to match the number of selected columns with \code{metacols}. Acceptable values are; STRING, NUMBER and BOOLEAN. Required parameter if \code{metacols} is being called.}

\item{metacol_prior}{Explicitly state the priority of selected metadata columns as a vector of characters. A higher number indicates a higher priority. Required parameter if \code{metacols} is being called.}

\item{out_dir}{The full path to the base directory where the files are being created.}
}
\value{
Nothing.
}
\description{
Finish setting up a new cBioPortal instance or updating an existing portal data set.
}
\details{
This function should be run as the last (or third step) in setting up a new cBioPortal instance.
The functions that should be run prior to these functions are; \link{setup_study} and \link{setup_fusions}.
\link{finalize_study} creates all the necessary tables and metadata files and case lists that are required to import a new study into cBioPortal.
Note, that all parameter arguments used in this function have to match the same parameter arguments for the previously run functions (\code{setup_study} and \code{setup_fusions}).
This function allows the user to specify additional fields from the collated metadata file (besides the "standard" fields).
For more information on how to use, see \code{metacols} and related parameters (\code{metacol_names}, \code{metacol_types}, and \code{meta_prior}).
}
\examples{
\dontrun{
#basic usage
finalize_study(these_sample_ids = c(ids, fusion_ids), out_dir = "GAMBLR/cBioPortal/instance01/")

#advanced usage
#get some samples
all_meta = get_gambl_metadata()
meta_sub = head(all_meta, 5)
my_samples = pull(meta_sub, sample_id)

#create a clinical file with additional collated metadata fields
finalize_study(these_sample_ids = my_samples,
               out_dir = "../",
               metacols = c("MeanCorrectedCoverage", "total_ssm"),
               metacol_names = c("Mean Corrected Coverage", "Total SSM"),
               metacol_types = c("NUMBER", "NUMBER"),
               metacol_prior = c("2", "1"))
}

}
