% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{finalize_study}
\alias{finalize_study}
\title{Finish setting up a new cBioPortal instance or updating an existing portal data set.}
\usage{
finalize_study(
  short_name = "GAMBL",
  include_icgc_data = FALSE,
  human_friendly_name = "GAMBL data",
  project_name = "gambl_minus_icgc",
  description = "GAMBL data without ICGC",
  cancer_type = "mixed",
  sample_ids,
  overwrite = TRUE,
  out_dir
)
}
\arguments{
\item{short_name}{A concise name for your portal project.}

\item{include_icgc_data}{Whether or not you want ICGC and other external data included.}

\item{human_friendly_name}{A slightly more verbose name for your project.}

\item{project_name}{Unique ID for your project.}

\item{description}{A verbose description of your data set.}

\item{cancer_type}{Cancer types included in study, default is "mixed".}

\item{sample_ids}{A vector of all the sample_id that were included in any of the data files for cbioportal.}

\item{overwrite}{Flag to specify that files should be overwritten if they exist. Default is TRUE.}

\item{out_dir}{The full path to the base directory where the files are being created.}
}
\value{
Nothing.
}
\description{
Finish setting up a new cBioPortal instance or updating an existing portal data set.
}
