% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{finalize_study}
\alias{finalize_study}
\title{Finalize Study (cBioPortal).}
\usage{
finalize_study(
  seq_type_filter = "genome",
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  cancer_type = "mixed",
  these_sample_ids,
  overwrite = TRUE,
  meta_columns,
  out_dir
)
}
\arguments{
\item{seq_type_filter}{the seq type you are setting up a study for, default is "genome".}

\item{short_name}{A concise name for your portal project.}

\item{human_friendly_name}{A slightly more verbose name for your project.}

\item{project_name}{Unique ID for your project.}

\item{description}{A verbose description of your data set.}

\item{cancer_type}{Cancer types included in study, default is "mixed".}

\item{these_sample_ids}{A vector of all the sample_id that were included in any of the data files for cBioPortal (i.e the output from \code{setup_study} and \code{setup_fusions}).}

\item{overwrite}{Flag to specify that files should be overwritten if they exist. Default is TRUE.}

\item{meta_columns}{Optional parameter for specifying metadata fields in the study. If not provided, the function will resort to using default metadata columns.}

\item{out_dir}{The full path to the base directory where the files are being created.}
}
\value{
Nothing.
}
\description{
Finish setting up a new cBioPortal instance or updating an existing portal data set.
}
\details{
This function should be run as the last (or third step) in setting up a new cBioPortal instance.
The functions that should be run prior to these functions are; \code{setup_study} and \code{setup_fusions}.
\code{finalize_study} creates all the necessary tables and metadata files (case lists) that are required to import a new study into cBioPortal.
Note, that all parameter arguments used in this function have to match the same parameter arguments for the previously run functions (\code{setup_study} and \code{setup_fusions}).
}
\examples{
\dontrun{
finalize_study(these_sample_ids = c(ids, fusion_ids), out_dir = "GAMBLR/cBioPortal/instance01/")
}
}
