% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{find_files_extract_wildcards}
\alias{find_files_extract_wildcards}
\title{Find Files Extract Wildcards}
\usage{
find_files_extract_wildcards(
  tool_results_path,
  search_pattern,
  genome_build,
  seq_type = "genome",
  unix_group = "gambl",
  tool_name
)
}
\arguments{
\item{tool_results_path}{Optional parameter, path to results.}

\item{search_pattern}{Search pattern.}

\item{genome_build}{Genome projection to be used.}

\item{seq_type}{Default is genome.}

\item{unix_group}{Default value is gambl.}

\item{tool_name}{Name of the tool.}
}
\value{
A tibble with wildcards.
}
\description{
Get wildcards for a set of samples.
}
\details{
Specify the file extension with \code{search_pattern} and the seq type, unix group, and genome build and the function will return a tibble with sample wildcards.
}
\examples{
file_details_manta = find_files_extract_wildcards(tool_name = "manta",
                                                  genome_build = c("hg38", "grch37"),
                                                  search_pattern = ".bed")

}
