% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{find_files_extract_wildcards}
\alias{find_files_extract_wildcards}
\title{Title}
\usage{
find_files_extract_wildcards(
  tool_results_path,
  search_pattern,
  genome_build,
  seq_type = "genome",
  unix_group = "gambl",
  tool_name
)
}
\arguments{
\item{tool_results_path}{path to results}

\item{search_pattern}{search pattern}

\item{genome_build}{genome projection to be sued}

\item{seq_type}{default si geneome}

\item{unix_group}{default value is gambl}
}
\description{
Title
}
\examples{
file_details_manta = find_files_extract_wildcards(tool_name = "manta", genome_build = c("hg38", "grch37"), search_pattern = ".bed")

}
