% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gene_to_region}
\alias{gene_to_region}
\title{Get regions from genes.}
\usage{
gene_to_region(
  gene_symbol,
  ensembl_id,
  genome_build = "grch37",
  return_as = "region"
)
}
\arguments{
\item{gene_symbol}{A vector of one or more gene symbols.}

\item{ensembl_id}{A vector of one or more Ensembl IDs.}

\item{genome_build}{Reference genome build.}

\item{return_as}{Specify the type of return. Default is region (chr:start-end), other acceptable arguments are "bed" and "df".}
}
\value{

}
\description{
Get regions from genes.
}
\examples{
bcl2_region = gene_to_region(gene_symbol = "BCL2", genome_build = "grch37")
bcl2_region = gene_to_region(ensembl_id = "ENSG00000171791", genome_build = "grch37")

}
