% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gene_to_region}
\alias{gene_to_region}
\title{Get regions from genes.}
\usage{
gene_to_region(
  gene_symbol,
  ensembl_id,
  genome_build = "grch37",
  return_as = "bed"
)
}
\arguments{
\item{gene_symbol}{Gene symbol (e.g BCL2).}

\item{ensembl_id}{Esemble ID (e.g ENSG00000171791).}

\item{genome_build}{Reference genome build.}

\item{return_as}{Specify the type of return. Default is bed (chr:start-end), other acceptable inputs are "df".}
}
\description{
Get regions from genes.
}
\examples{
bcl2_region = gene_to_region(gene_symbol = "BCL2", genome_build = "grch37")
bcl2_region = gene_to_region(ensembl_id = "ENSG00000171791", genome_build = "grch37")

}
