% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gene_to_region}
\alias{gene_to_region}
\title{Gene To Region.}
\usage{
gene_to_region(
  gene_symbol,
  ensembl_id,
  genome_build = "grch37",
  return_as = "region",
  sort_regions = TRUE
)
}
\arguments{
\item{gene_symbol}{A vector of one or more gene symbols.}

\item{ensembl_id}{A vector of one or more Ensembl IDs.}

\item{genome_build}{Reference genome build. Possible values are "grch37" (default) or "hg38".}

\item{return_as}{Specify the type of return. Default is "region" (chr:start-end), other acceptable arguments are "bed" and "df".}

\item{sort_regions}{A boolean parameter (TRUE is the default) indicating whether regions should be sorted by chomosome and start location.}
}
\value{
A data frame, or a string with region(s) for the provided gene(s).
}
\description{
Return coordinates for a given gene or a set of genes.
}
\details{
This function takes one or multiple gene names, either as hugo symbols or Ensembl IDs
and returns the coordinates for the selected genes in multiple formats (\code{return_as}).
The possible return formats are; bed, data frame and in "region" format (chr:start-end).
For returning genes residing in specific regions, see \link{region_to_gene}.
}
\examples{
bcl2_region = gene_to_region(gene_symbol = "BCL2",
                             genome_build = "grch37")

bcl2_region = gene_to_region(ensembl_id = "ENSG00000171791",
                             genome_build = "grch37")

}
