% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{genome_to_exome}
\alias{genome_to_exome}
\title{Subset maf file to only features that would be available in the WEX data.}
\usage{
genome_to_exome(
  maf,
  custom_bed,
  genome_build = "grch37",
  padding = 100,
  chr_prefixed = FALSE
)
}
\arguments{
\item{maf}{Incoming maf object. Can be maf-like data frame or maftools maf object. Required parameter. Minimum columns that should be present are Chromosome, Start_Position, and End_Position.}

\item{custom_bed}{Optional argument specifying a path to custom bed file for covered regions. Must be bed-like and contain chrom, start, and end position information in first 3 columns. Other columns are disregarded if provided.}

\item{genome_build}{String indicating genome build of the maf file. Default is grch37, but can accept modifications of both grch37- and hg38-based duilds.}

\item{padding}{Numeric value that will be used to pad probes in WEX data from both ends. Default is 100. After padding, overlapping features are squished together.}

\item{chr_prefixed}{Is the data chr-prefixed or not? Default is FALSE.}
}
\value{
Aa data frame of maf-like object with same columns as in input, but where rows are only kept for features that would be present as if the sample is WEX.
}
\description{
Subset maf file to only features that would be available in the WEX data.
}
\examples{
myc_ashm_maf = get_ssm_by_region(region="8:128748352-128749427") # get all ssm in the MYC aSHM region
GAMBLR::genome_to_exome(maf=myc_ashm_maf) # get mutations with 100 bp padding (default)
GAMBLR::genome_to_exome(maf=myc_ashm_maf, padding = 0) # get mutations covered in WEX with no padding

}
