% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_bams}
\alias{get_bams}
\title{Get BAMs.}
\usage{
get_bams(this_sample_id, this_patient_id)
}
\arguments{
\item{this_sample_id}{Sample ID of interest.}

\item{this_patient_id}{patient ID of interest.}
}
\value{
A list that contains the genome_build and an igv-friendly build (igv_build), a list of bam file paths for tumour, normal and mrna data.
}
\description{
Get full paths for bam files for a sample or patient.
}
\details{
Returns a list with BAM paths for tumour, normal and mrna data.
This function expects a sample ID (\code{this_sample_id}) or a patient ID (\code{this_patient_id}).
}
\examples{

#example 1, using a sample ID
bam_details = get_bams(this_sample_id = "HTMCP-01-06-00422-01A-01D")

#example 2, using a patient ID
bam_details = get_bams(this_patient_id = "HTMCP-01-06-00422")

}
