% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_cn_segments}
\alias{get_cn_segments}
\title{Retrieve all copy number segments from the GAMBL database that overlap with a single genomic coordinate range.}
\usage{
get_cn_segments(
  region,
  chromosome,
  qstart,
  qend,
  projection = "grch37",
  this_seq_type = "genome",
  with_chr_prefix = FALSE,
  streamlined = FALSE,
  from_flatfile = TRUE
)
}
\arguments{
\item{region}{Region formatted like chrX:1234-5678 or X:1234-56789.}

\item{chromosome}{The chromosome you are restricting to. Required parameter if region is not specified.}

\item{qstart}{Start coordinate of the range you are restricting to. Required parameter if region is not specified.}

\item{qend}{End coordinate of the range you are restricting to. Required parameter if region is not specified.}

\item{projection}{Selected genome projection for returned Cn segments.}

\item{this_seq_type}{Seq type for returned Cn segments. Currently, only genome is supported. Capture samples will be added once processed through CNV protocols.}

\item{with_chr_prefix}{Boolean parameter for toggling if chr prefixes should be present in the return, default is FLASE.}

\item{streamlined}{Return a basic rather than full MAF format.}

\item{from_flatfile}{Set to TRUE by default.}
}
\value{
A data frame with CN segments for the specified region.
}
\description{
Retrieve all copy number segments from the GAMBL database that overlap with a single genomic coordinate range.
}
\examples{
#Example using chromosome, qstart and qend parameters:
segments_region_grch37 = get_cn_segments(chromosome = "chr8",
                                         qstart = 128723128,
                                         qend = 128774067) 
                                    
#Example using the regions parameter:
segments_region_hg38 = get_cn_segments(region = "chr8:128,723,128-128,774,067",
                                       projection = "hg38", 
                                       with_chr_prefix = TRUE) 


}
