% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_cn_states}
\alias{get_cn_states}
\title{Get a copy number matrix for all samples based on segmented data in database}
\usage{
get_cn_states(regions_list, regions_bed, region_names)
}
\arguments{
\item{regions_list}{A list of regions in the format chrom:start-end}

\item{regions_bed}{A bed file with one row for each region you want to determine the CN state from}

\item{region_names}{}
}
\value{

}
\description{
Get a copy number matrix for all samples based on segmented data in database
}
\examples{
#basic usage, generic lymphoma gene list
cn_matrix = get_cn_states(regions_bed=grch37_lymphoma_genes_bed)
}
