% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_coding_ssm}
\alias{get_coding_ssm}
\title{Retrieve all coding SSMs from the GAMBL database in MAF-like format}
\usage{
get_coding_ssm(
  db = "gambl_test",
  table_name = "maf_slms3_hg19_icgc",
  limit_cohort,
  limit_pathology
)
}
\arguments{
\item{db}{The GAMBL database name}

\item{table_name}{The table we are querying}

\item{limit_cohort}{Supply this to restrict mutations to one cohort}

\item{limit_pathology}{Supply this to restrict mutations to one pathology}
}
\value{
A data frame containing all the MAF data columns (one row per mutation)
}
\description{
Retrieve all coding SSMs from the GAMBL database in MAF-like format
}
\examples{
#basic usage
maf_data = get_coding_ssm(limit_cohort=c("BL_Adult","BL_Pediatric","BL_ICGC"))
}
