% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_coding_ssm}
\alias{get_coding_ssm}
\title{Retrieve all coding SSMs from the GAMBL database in MAF-like format.}
\usage{
get_coding_ssm(
  limit_cohort,
  exclude_cohort,
  limit_pathology,
  limit_samples,
  these_samples_metadata,
  force_unmatched_samples,
  projection = "grch37",
  seq_type = "genome",
  basic_columns = TRUE,
  maf_cols = NULL,
  return_cols = FALSE,
  from_flatfile = TRUE,
  augmented = TRUE,
  min_read_support = 3,
  groups = c("gambl", "icgc_dart"),
  include_silent = TRUE,
  engine = "fread_maf"
)
}
\arguments{
\item{limit_cohort}{Supply this to restrict mutations to one or more cohorts in a list.}

\item{exclude_cohort}{Supply this to exclude mutations from one or more cohorts in a list.}

\item{limit_pathology}{Supply this to restrict mutations to one pathology.}

\item{limit_samples}{Supply this to restrict mutations to a vector of sample_id (instead of subsetting using the provided metadata)}

\item{these_samples_metadata}{Supply a metadata table to auto-subset the data to samples in that table before returning}

\item{force_unmatched_samples}{Optional argument for forcing unmatched samples, using get_ssm_by_samples.}

\item{projection}{Reference genome build for the coordinates in the MAF file. The default is hg19 genome build.}

\item{seq_type}{The seq_type you want back, default is genome.}

\item{basic_columns}{Set to FALSE to override the default behavior of returning only the first 45 columns of MAF data.}

\item{maf_cols}{if basic_columns is set to FALSE, the user can specify what columns to be returned within the MAF. This parameter can either be a list of indexes (integer) or a list of characters (matching columns in MAF).}

\item{return_cols}{If set to TRUE, a vector with all avaialble column names will be returned. Default is FALSE.}

\item{from_flatfile}{Set to TRUE to obtain mutations from a local flatfile instead of the database. This can be more efficient and is currently the only option for users who do not have ICGC data access.}

\item{augmented}{default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs)}

\item{groups}{Unix groups for the samples to be included. Default is both gambl and icgc_dart samples.}

\item{include_silent}{Logical parameter indicating whether to include silent mutations into coding mutations. Default is TRUE.}
}
\value{
A data frame containing all the MAF data columns (one row per mutation).
}
\description{
Retrieve all coding SSMs from the GAMBL database in MAF-like format.
}
\examples{
#basic usage
maf_data = get_coding_ssm(limit_cohort = c("BL_ICGC"))
maf_data = get_coding_ssm(limit_samples = "HTMCP-01-06-00485-01A-01D")

}
