% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_coding_ssm_status}
\alias{get_coding_ssm_status}
\title{Tabulate mutation status for non-silent SSMs for a set of genes}
\usage{
get_coding_ssm_status(
  gene_symbols,
  these_samples_metadata,
  from_flatfile = TRUE
)
}
\arguments{
\item{from_flatfile}{}
}
\value{

}
\description{
Tabulate mutation status for non-silent SSMs for a set of genes
}
\examples{
coding_tabulated_df = get_coding_ssm_status(gene_symbols=c("MYC","KMT2D"))
}
