% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_coding_ssm_status}
\alias{get_coding_ssm_status}
\title{Tabulate mutation status for non-silent SSMs for a set of genes}
\usage{
get_coding_ssm_status(
  gene_symbols,
  these_samples_metadata,
  from_flatfile = TRUE,
  maf_path = NULL,
  include_hotspots = TRUE,
  recurrence_min = 5,
  review_hotspots = TRUE,
  genes_of_interest = c("FOXO1", "MYD88", "CREBBP"),
  genome_build = "hg19"
)
}
\arguments{
\item{gene_symbols}{List of gene symbols for which the mutation status will be tabulated. If not provided, lymphoma genes will be returned by default.}

\item{these_samples_metadata}{The matedata for samples of interest to be included in the returned matrix. Only the column "sample_id" is required. If not provided, the matrix is tabulated for all available samples as default.}

\item{from_flatfile}{Integer value indicating minimal recurrence level}

\item{maf_path}{If the status of coding SSM should be tabulated from a custom maf file, provide path to the maf in this argument. The default is set to NULL.}

\item{include_hotspots}{Logical parameter indicating whether hotspots object should also be tabulated. Default is TRUE.}

\item{review_hotspots}{Logical parameter indicating whether hotspots object should be reviewed to include functionally relevant mutations or rare lymphoma-related genes. Default is TRUE.}

\item{drop_unreviewed_hotspots}{Logical argumant indicating whether the hotspots for all unsupported genes should be tabulated separately (TRUE) or disregarded to avoid signal dilution between coding cnage and hotspot (FALSE). Default is FALSE.}

\item{...}{Other parameters accepted by the review_hotspots() function}
}
\value{

}
\description{
Tabulate mutation status for non-silent SSMs for a set of genes
}
\examples{
coding_tabulated_df = get_coding_ssm_status(gene_symbols=c("MYC","KMT2D"))
coding_tabulated_df = get_coding_ssm_status() #all lymphoma genes from bundled NHL gene list
}
