% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_excluded_samples}
\alias{get_excluded_samples}
\title{Exclude samples that have been excluded from certain analyses and drop from merges}
\usage{
get_excluded_samples(tool_name = "slms-3")
}
\arguments{
\item{tool_name}{The tool or pipeline that generated the files (should be the same for all).}
}
\description{
Exclude samples that have been excluded from certain analyses and drop from merges
}
\examples{
excluded_samp = get_excluded_samples()

}
