% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_gambl_colours}
\alias{get_gambl_colours}
\title{Title}
\usage{
get_gambl_colours(classification = "lymphgen")
}
\arguments{
\item{classification}{(optionally request only colours for pathology, lymphgen or copy_number)}
}
\value{
A named vector of colour codes for lymphgen classes and pathology
}
\description{
Title
}
