% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_gambl_metadata}
\alias{get_gambl_metadata}
\title{Get GAMBL metadata}
\usage{
get_gambl_metadata(
  seq_type_filter = "genome",
  tissue_status_filter = c("tumour"),
  case_set,
  remove_benchmarking = TRUE,
  with_outcomes = TRUE,
  from_flatfile = TRUE
)
}
\arguments{
\item{seq_type_filter}{Filtering criteria (default: all genomes)}

\item{tissue_status_filter}{Filtering criteria (default: only tumour genomes)}

\item{case_set}{optional short name for a pre-defined set of cases avoiding any}

\item{remove_benchmarking}{By default the FFPE benchmarking duplicate samples will be dropped}

\item{with_outcomes}{Optionally join to gambl outcome data}

\item{from_flatfile}{New default is to use the metadata in the flatfiles from your clone of the repo. Can be over-ridden to use the database
embargoed cases (current options: 'BLGSP-study', 'FL-study', 'DLBCL-study', 'FL-DLBCL-study', 'DLBCL-unembargoed', 'BL-DLBCL-manuscript')}
}
\value{
A data frame with metadata for each biopsy in GAMBL
}
\description{
Get GAMBL metadata
}
\examples{
# basic usage
my_metadata = get_gambl_metadata()
# use pre-defined custom sample sets
only_blgsp_metadata = get_gambl_metadata(case_set="BLGSP-study")
# override default filters and request metadata for samples other than tumour genomes, e.g. also get the normals
only_normal_metadata = get_gambl_metadata(tissue_status_filter = c('tumour','normal'))
}
