% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_gene_cn_and_expression}
\alias{get_gene_cn_and_expression}
\title{Get Gene CN and Expression.}
\usage{
get_gene_cn_and_expression(gene_symbol, ensembl_id)
}
\arguments{
\item{hugo_symbol}{One or more gene symbols. Should match the values in a maf file.}

\item{ensembl_gene_id}{One or more ensembl gene IDs. Only one of hugo_symbols or ensembl_gene_ids may be used.}
}
\value{
A list.
}
\description{
Get the copy number and expression for a single gene.
}
\details{
This function works well with both Hugo Symbols and Ensembl Gene IDs.
It's also possible to specify more than one gene.
}
\examples{
MYC_cn_expression = get_gene_cn_and_expression("MYC")

}
