% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_gene_expression}
\alias{get_gene_expression}
\title{Get the expression for one or more genes for all GAMBL samples}
\usage{
get_gene_expression(
  hugo_symbols,
  tidy_expression_data,
  metadata,
  join_with = "mrna"
)
}
\arguments{
\item{join_with}{}
}
\value{

}
\description{
Get the expression for one or more genes for all GAMBL samples
}
