% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_gene_expression}
\alias{get_gene_expression}
\title{Title}
\usage{
get_gene_expression(
  database_name,
  hugo_symbols,
  tidy_expression_data,
  metadata,
  join_with = "mrna"
)
}
\arguments{
\item{join_with}{}
}
\value{

}
\description{
Title
}
