% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_lymphgen}
\alias{get_lymphgen}
\title{Get a specific flavour of LymphGen from the main GAMBL outputs and tidy the composites.
Optionally return a matrix of features instead}
\usage{
get_lymphgen(
  these_samples_metadata,
  flavour,
  return_feature_matrix = FALSE,
  return_feature_annotation = FALSE,
  lymphgen_file,
  keep_all_rows = FALSE,
  keep_original_columns = FALSE
)
}
\arguments{
\item{keep_original_columns}{}
}
\description{
Get a specific flavour of LymphGen from the main GAMBL outputs and tidy the composites.
Optionally return a matrix of features instead
}
