% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_manta_sv}
\alias{get_manta_sv}
\title{Retrieve Manta SVs from the database and filter}
\usage{
get_manta_sv(
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  pairing_status,
  sample_id,
  chromosome,
  qstart,
  qend,
  region,
  with_chr_prefix = FALSE,
  from_flatfile = FALSE,
  projection = "grch37"
)
}
\arguments{
\item{min_vaf}{The minimum tumour VAF for a SV to be returned}

\item{min_score}{The lowest Manta somatic score for a SV to be returned}

\item{chromosome}{The chromosome you are restricting to}

\item{qstart}{Query start coordinate of the range you are restricting to}

\item{qend}{Query end coordinate of the range you are restricting to}

\item{region}{Region formatted like chrX:1234-5678 instead of specifying chromosome, start and end separately}

\item{with_chr_prefix}{Prepend all chromosome names with chr (required by some downstream analyses)}

\item{table_name}{The table we are querying}

\item{pair_status}{Use to restrict results (if desired) to matched or unmatched results (default is to return all)}
}
\value{
A data frame in a bedpe-like format with additional columns that allow filtering of high-confidence SVs
}
\description{
Retrieve Manta SVs from the database and filter
}
\examples{
#lazily get every SV in the table with default quality filters
all_sv = get_manta_sv()
#get all SVs for a single sample
some_sv = get_manta_sv(sample_id="94-15772_tumorA")
#get the SVs in a region around MYC
myc_locus_sv = get_manta_sv(region="8:128723128-128774067")
}
