% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_manta_sv_by_sample}
\alias{get_manta_sv_by_sample}
\title{Load the manta output (from individual flat file) for 1 sample.}
\usage{
get_manta_sv_by_sample(
  this_sample_id,
  these_samples_metadata,
  force_lift = FALSE,
  return_anyway = FALSE,
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  projection = "grch37"
)
}
\arguments{
\item{this_sample_id}{The single sample ID you want to obtain the result from.}

\item{these_samples_metadata}{A metadata table containing metadata for this_sample_id, or sample of interest. This parameter is required.}

\item{force_lift}{If TRUE, coordinates will be lifted (if needed) to the selected projection. Default is FALSE. WARNING: if your code calls this function directly, set this parameter to TRUE to ensure that the returned calls are in respect to the requested projection.}

\item{return_anyway}{Set to TRUE to force variant calls to be returned, even if they're not lifted, i.e you either know what you're doing and plan to use liftover_bedpe yourself, or you plan to run liftover_bedpe for mutiple samples in on one df, (as in get_manta_sv_by_samples). Default is FALSE.}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Default value is 0.1.}

\item{min_score}{The lowest Manta somatic score for a SV to be returned. Default value is 40.}

\item{pass}{If set to TRUE, only return SVs that are annotated with PASS in the FILTER column. Set to FALSE to keep all variants, regardless if they PASS the filters. Default is TRUE.}

\item{projection}{The projection of returned calls. Default is grch37.}
}
\value{
a data frame containing the Manta outputs from this_sample_id in a bedpe-like format with additional columns extracted from the VCF column.
}
\description{
This function is used for retrieving Manta results (structural variants) from individual flat-files (one sample).
For multiple samples, please see get_manta_sv_by_samples (a convenience wrapper function for get_manta_by_sample).
Additional columns are extracted from the VCF column and standard filtering options are available.
This function also performs a lift-over to selected projection, if needed.
Please note, if force_lift is set to FALSE, an extra column will be added that states if the returned variant calls need to be lifted.
The value for this column is returned TRUE (for all rows) if the available genome projection for the selected sample does not match the selected projection (i.e requiring the user to manually lift the calls).
}
\examples{
#example 1
#get manta calls for a sample that needs to be lifted to "hg38" and let this function take care of the liftover step for you. 
my_sv_hg38_lifted = get_manta_sv_by_sample(this_sample_id = "99-27783_tumorA", these_samples_metadata = metadata, projection = "hg38", force_lift = TRUE)

#example 2
#get manta calls for a sample and do the liftover on your own, based on the information in the extra column in the returned bedpe data frame (need_lift)
my_sv = get_manta_sv_by_sample(this_sample_id = "99-27783_tumorA", these_samples_metadata = metadata, projection = "hg38", force_lift = FALSE, return_anyway = TRUE)

#check if returned variant calls need to be lifted
my_sv$need_lift[1]

#yes, run liftover on returned variant calls

}
