% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_manta_sv_by_sample}
\alias{get_manta_sv_by_sample}
\title{Load the manta output (from individual flat file) for 1 sample.}
\usage{
get_manta_sv_by_sample(
  this_sample_id,
  these_samples_metadata,
  force_lift = TRUE,
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  with_chr_prefix = FALSE,
  projection = "grch37"
)
}
\arguments{
\item{this_sample_id}{The single sample ID you want to obtain the result from.}

\item{these_samples_metadata}{A metadata table containing metadata for this_sample_id, or sample of interest. This parameeter is required.}

\item{force_lift}{If TRUE, coordinates will be lifted (if needed) to the selected projection. WARNING! Should always be TRUE if this function is running in stand-alone mode.}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Default value is 0.1.}

\item{min_score}{The lowest Manta somatic score for a SV to be returned. Default value is 40.}

\item{pass}{If set to TRUE, include SVs that are annotated with PASS in FILTER column. Default is TRUE.}

\item{with_chr_prefix}{Toggle if chromosome names should be prefixed or not, default is FALSE.}

\item{projection}{The projection of returned calls. Default is grch37.}
}
\value{
a data frame containing the Manta outputs from this_sample_id in a bedpe-like format with additional columns extracted from the VCF column.
}
\description{
This function is used for retrieving Manta results (structural variants) from individual flat-files (one sample).
For multiple samples, please see get_manta_sv_by_samples (a convenience wrapper function for get_manta_by_sample).
Additional columns are extracted from the VCF column and standard filtering options are available.
This function also performs a lift-over to selected projection, if needed.
}
\examples{
my_svs = get_manta_sv_by_sample(this_sample_id = "00-14595_tumorD",
                                these_samples_metadata = get_gambl_metadata())

}
