% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_manta_sv_by_samples}
\alias{get_manta_sv_by_samples}
\title{Get Manta SV By Samples.}
\usage{
get_manta_sv_by_samples(
  these_samples_metadata,
  min_vaf = 0.1,
  min_score = 40,
  pass = TRUE,
  projection = "grch37"
)
}
\arguments{
\item{these_samples_metadata}{The only required parameter is a metadata table (data frame) that must contain a row for each sample you want the data from. The additional columns the data frame needs to contain, besides sample_id, are: unix_group, genome_build, seq_type, pairing_status.}

\item{min_vaf}{The minimum tumour VAF for a SV to be returned. Default value is 0.1.}

\item{min_score}{The lowest Manta somatic score for a SV to be returned. Default value is 40.}

\item{pass}{If set to TRUE, only return SVs that are annotated with PASS in the FILTER column. Set to FALSE to keep all variants, regardless if they PASS the filters. Default is TRUE.}

\item{projection}{The projection of returned calls. Default is grch37.}
}
\value{
A data frame containing the Manta outputs from all sample_id in these_samples_metadata in a bedpe-like format with additional columns extracted from the VCF column.
}
\description{
Load the manta output for a set of samples.
}
\details{
This is a convenience wrapper function for get_manta_sv_by_sample (and called by get_manta_sv).
}
\examples{
all_sv = get_manta_sv()
missing_samples = get_gambl_metadata() \%>\%
 anti_join(all_sv, by = c("sample_id" = "tumour_sample_id"))

missing_from_merge = get_manta_sv_by_samples(these_samples_metadata = missing_samples)

}
