% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_mutation_frequency_bin_matrix}
\alias{get_mutation_frequency_bin_matrix}
\title{Count hypermutated bins and generate heatmaps/cluster the data}
\usage{
get_mutation_frequency_bin_matrix(
  regions,
  regions_df,
  these_samples_metadata,
  region_padding = 1000,
  metadataColumns = c("pathology"),
  sortByColumns = c("pathology"),
  expressionColumns = c(),
  orientation = "sample_rows",
  customColour = NULL,
  slide_by = 100,
  window_size = 500,
  min_count_per_bin = 3,
  min_bin_recurrence = 5,
  min_bin_patient = 0,
  region_fontsize = 8,
  cluster_rows_heatmap = FALSE,
  cluster_cols_heatmap = FALSE,
  show_gene_colours = FALSE,
  legend_row = 3,
  legend_col = 3
)
}
\arguments{
\item{regions}{Vector of regions in the format "chr:start-end"}

\item{these_samples_metadata}{GAMBL metadata subset to the cases you want to process (or full metadata)}

\item{region_padding}{How many bases will be added on the left and right of the regions to ensure any small regions are sufficiently covered by bins}

\item{metadataColumns}{What metadata will be shown in the visualization}

\item{sortByColumns}{Which of the metadata to sort on for the heatmap}

\item{customColour}{Optional named list of named vectors for specifying all colours for metadata. Can be generated with map_metadata_to_colours}

\item{slide_by}{How far to shift before starting the next window}

\item{window_size}{The width of your sliding window}

\item{min_bin_recurrence}{How many samples a bin must be mutated in to retain in the visualization}

\item{min_bin_patient}{How many bins must a patient mutated in to retain in the visualization}

\item{cluster_rows_heatmap}{Optional parameter to enable/disable clustering of each dimension of the heatmap}

\item{show_gene_colours}{Optional logical argument indicating whether regions should have associated colours plotted as annotation track of heatmap}

\item{legend_row}{Fiddle with these to widen or narrow your legend}

\item{legend_col}{Fiddle with these to widen or narrow your legend}

\item{region_df}{Data frame of regions with four columns (chrom,start,end,gene_name)}
}
\value{

}
\description{
Count hypermutated bins and generate heatmaps/cluster the data
}
