% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sample_cn_segments}
\alias{get_sample_cn_segments}
\title{Get all segments for a single sample_id}
\usage{
get_sample_cn_segments(sample_id, with_chr_prefix = FALSE, streamlined = FALSE)
}
\arguments{
\item{sample_id}{The sample_id for the sample to retrieve segments for}

\item{with_chr_prefix}{Set to TRUE to add a chr prefix to chromosome names}

\item{streamlined}{Return a minimal output rather than full details}
}
\value{

}
\description{
Get all segments for a single sample_id
}
