% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sample_cn_segments}
\alias{get_sample_cn_segments}
\title{Get Sample CN Segments.}
\usage{
get_sample_cn_segments(
  these_sample_ids,
  these_samples_metadata,
  from_flatfile = TRUE,
  projection = "grch37",
  this_seq_type = "genome",
  with_chr_prefix = FALSE,
  streamlined = FALSE
)
}
\arguments{
\item{these_sample_ids}{Optional argument, sample_id (vector of characters) for the sample(s) to retrieve segments for. If not provided, the function will return CN segments for all available sample IDs present in the current metadata.}

\item{these_samples_metadata}{Optional, provide a metadata (data frame) subset to the sample IDs of interest.}

\item{from_flatfile}{Set to TRUE by default.}

\item{projection}{Selected genome projection for returned CN segments. Default is "grch37".}

\item{this_seq_type}{Seq type for returned CN segments. One of "genome" (default) or "capture".}

\item{with_chr_prefix}{Set to TRUE to add a chr prefix to chromosome names. Default is FALSE.}

\item{streamlined}{Return a minimal output rather than full details. Default is FALSE.}
}
\value{
A data frame of segments for a specific or multiple sample ID(s).
}
\description{
Get all segments for a single (or multiple) sample_id(s).
}
\details{
This function returns CN segments for samples. This works for single sample or multiple samples.
Specify the sample IDs you are interested in with \code{these_sample_ids} (as a vector of characters),
Or call this function with \code{these_samples_metadata} if you already ahve a metadata table subset to the sample IDs of interest.
If none off the above parameters are specified, the function will return CN segments for available samples.
Note, this. function internally calls \link{id_ease} for dealing with sample IDs and metadata tables.
Is this function not what you are looking for? Try one of the following, similar, functions; \link{assign_cn_to_ssm}, \link{get_cn_segments}, \link{get_cn_states},
}
\examples{
#return everything (default: genome, grch37)
all_segs = get_sample_cn_segments()

#using sample ID parameter
one_sample = get_sample_cn_segments(these_sample_ids = "00-22011_tumorB", 
                                    this_seq_type = "capture")

two_samples = get_sample_cn_segments(these_sample_ids = c("00-14595_tumorA",
                                                          "00-16220_tumorB"), 
                                     projection = "hg38")

#get some metadata
this_meta = get_gambl_metadata() \%>\% 
 dplyr::filter(sample_id \%in\% c("BLGSP-71-23-00440-01A-01E", 
                                "FL1014T2",
                                "HTMCP-01-06-00563-01A-01D"))

capture_meta = get_gambl_metadata(seq_type_filter = "capture")

#return sample segments by using the metadata parameter
segs_meta = get_sample_cn_segments(these_samples_metadata = this_meta)

all_segs_cap = get_sample_cn_segments(these_samples_metadata = capture_meta, 
                                      projection = "hg38")

}
