% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sample_wildcards}
\alias{get_sample_wildcards}
\title{Get wildcards for a sample_id/seq_type combination.}
\usage{
get_sample_wildcards(this_sample_id, seq_type)
}
\arguments{
\item{this_sample_id}{}

\item{seq_type}{}
}
\value{

}
\description{
Get wildcards for a sample_id/seq_type combination.
}
