% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_ssh_session}
\alias{get_ssh_session}
\title{Create an ssh session to the GSC (requires active VPN connection)}
\usage{
get_ssh_session(host = "gphost01.bcgsc.ca")
}
\arguments{
\item{host}{(default is gphost01.bcgsc.ca)}
}
\description{
Create an ssh session to the GSC (requires active VPN connection)
}
