% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_patients}
\alias{get_ssm_by_patients}
\title{Get MAF-format data frame for more than one patient using at most one augmented_maf  per patient(i.e. unique superset of variants)
and combine or subset from a merged MAF (wraps get_ssm_by_samples)
See get_ssm_by_samples for more information}
\usage{
get_ssm_by_patients(
  these_patient_ids,
  these_samples_metadata,
  tool_name = "slms-3",
  projection = "grch37",
  seq_type = "genome",
  flavour = "clustered",
  min_read_support = 3,
  subset_from_merge = TRUE
)
}
\arguments{
\item{these_patient_ids}{A vector of sample_id that you want results for. This is the only required argument.}

\item{these_samples_metadata}{Optional metadata table}

\item{tool_name}{Only supports slms-3 currently}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38)}

\item{flavour}{Currently this function only supports one flavour option but this feature is meant for eventual compatability with additional variant calling parameters/versions}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs)}

\item{subset_from_merge}{Instead of merging individual MAFs, the data will be subset from a pre-merged MAF of samples with the specified seq_type}
}
\value{

}
\description{
Get MAF-format data frame for more than one patient using at most one augmented_maf  per patient(i.e. unique superset of variants)
and combine or subset from a merged MAF (wraps get_ssm_by_samples)
See get_ssm_by_samples for more information
}
\examples{
merged_maf_force_unmatched = get_ssm_by_samples(these_sample_ids=c("HTMCP-01-06-00485-01A-01D","14-35472_tumorA","14-35472_tumorB"))
}
