% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_region}
\alias{get_ssm_by_region}
\title{Retrieve all SSMs from the GAMBL database within a single genomic coordinate range}
\usage{
get_ssm_by_region(
  chromosome,
  qstart,
  qend,
  region = "",
  basic_columns = TRUE,
  streamlined = FALSE,
  maf_data
)
}
\arguments{
\item{chromosome}{The chromosome you are restricting to (with or without a chr prefix)}

\item{qstart}{Query start coordinate of the range you are restricting to}

\item{qend}{Query end coordinate of the range you are restricting to}

\item{region}{Region formatted like chrX:1234-5678 instead of specifying chromosome, start and end separately}

\item{basic_columns}{Set to TRUE to override the default behaviour of returning only the first 45 columns of MAF data}
}
\value{
A data frame containing all the MAF data columns (one row per mutation)
}
\description{
Retrieve all SSMs from the GAMBL database within a single genomic coordinate range
}
\examples{
#basic usage
my_mutations=get_ssm_by_region(region="chr8:128,723,128-128,774,067")
#specifying chromosome, start and end individually
my_mutations=get_ssm_by_region(chromosome="8",qstart=128723128,qend=128774067)
}
