% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_region}
\alias{get_ssm_by_region}
\title{Retrieve all SSMs from the GAMBL database within a single genomic coordinate range}
\usage{
get_ssm_by_region(
  db = "gambl_test",
  table_name = "maf_slms3_hg19_icgc",
  chromosome = NULL,
  start = NULL,
  end = NULL,
  region = ""
)
}
\arguments{
\item{db}{The GAMBL database name}

\item{table_name}{The table we are querying}

\item{chromosome}{The chromosome you are restricting to}

\item{start}{Start coordinate of the range you are restricting to}

\item{end}{End coordinate of the range you are restricting to}

\item{region}{Region formatted like chrX:1234-5678 instead of specifying chromosome, start and end separately}
}
\value{
A data frame containing all the MAF data columns (one row per mutation)
}
\description{
Retrieve all SSMs from the GAMBL database within a single genomic coordinate range
}
\examples{
#basic usage
my_mutations=get_ssm_by_region(region="chr8:128,723,128-128,774,067")
#specifying chromosome, start and end individually
my_mutations=get_ssm_by_region(chromosome="8",start=128723128,end=128774067)
}
