% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_regions}
\alias{get_ssm_by_regions}
\title{Efficiently retrieve all mutations across a range of genomic regions.}
\usage{
get_ssm_by_regions(
  regions_list,
  regions_bed,
  streamlined = FALSE,
  basic_columns = TRUE,
  maf_cols = NULL,
  return_cols = FALSE,
  use_name_column = FALSE,
  maf_data = maf_data,
  from_indexed_flatfile = TRUE,
  mode = "slms-3",
  augmented = TRUE,
  seq_type = "genome",
  projection = "grch37",
  min_read_support = 4,
  allow_clustered = TRUE
)
}
\arguments{
\item{regions_list}{Either provide a vector of regions in the chr:start-end format OR.}

\item{regions_bed}{Better yet, provide a bed file with the coordinates you want to retrieve.}

\item{streamlined}{Set to TRUE to return only the following columns; start, sample_id and region_name.}

\item{basic_columns}{Set to FALSE to override the default behavior of returning only the following columns; chromosome, start, end and sample_id.}

\item{maf_cols}{if basic_columns is set to FALSE, the user can specify what columns to be returned within the MAF. This parameter can either be a list of indexes (integer) or a list of characters.}

\item{return_cols}{If set to TRUE, a vector with all available column names will be returned. Default is FALSE.}

\item{from_indexed_flatfile}{Set to TRUE to avoid using the database and instead rely on flatfiles (only works for streamlined data, not full MAF details).}

\item{mode}{Only works with indexed flatfiles. Accepts 2 options of "slms-3" and "strelka2" to indicate which variant caller to use. Default is "slms-3".}

\item{augmented}{default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF}

\item{seq_type}{The seq_type you want back, default is genome.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38).}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).}

\item{allow_clustered}{Logical parameter indicating whether to use SLMS-3 results with clustered events. Default is FALSE.}
}
\value{
Returns a data frame of variants in MAF-like format.
}
\description{
Efficiently retrieve all mutations across a range of genomic regions.
}
\examples{
#basic usage, adding custom names from bundled ashm data frame
regions_bed = grch37_ashm_regions \%>\% mutate(name = paste(gene, region, sep = "_"))
ashm_maf = get_ssm_by_regions(regions_bed = regions_bed)

}
