% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_regions}
\alias{get_ssm_by_regions}
\title{Efficiently retrieve all mutations across a range of genomic regions.}
\usage{
get_ssm_by_regions(
  regions_list,
  regions_bed,
  streamlined = TRUE,
  maf_data = maf_data,
  use_name_column = FALSE,
  from_indexed_flatfile = TRUE,
  mode = "slms-3",
  augmented = TRUE,
  seq_type = "genome",
  projection = "grch37",
  min_read_support = 4,
  basic_columns = F
)
}
\arguments{
\item{regions_list}{Either provide a vector of regions in the chr:start-end format OR.}

\item{regions_bed}{Better yet, provide a bed file with the coordinates you want to retrieve.}

\item{streamlined}{Return a basic rather than full MAF format, default is FALSE.}

\item{use_name_column}{If your bed-format data frame has a name column (must be named "name") these can be used to name your regions.}

\item{from_indexed_flatfile}{Set to TRUE to avoid using the database and instead rely on flatfiles (only works for streamlined data, not full MAF details).}

\item{mode}{Only works with indexed flatfiles. Accepts 2 options of "slms-3" and "strelka2" to indicate which variant caller to use. Default is "slms-3".}

\item{augmented}{default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF}

\item{seq_type}{The seq_type you want back, default is genome.}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38).}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs).}
}
\value{
Returns a data frame of variants in MAF-like format.
}
\description{
Efficiently retrieve all mutations across a range of genomic regions.
}
\examples{
#basic usage, adding custom names from bundled ashm data frame
regions_bed = grch37_ashm_regions \%>\% mutate(name = paste(gene, region, sep = "_"))
ashm_basic_details = get_ssm_by_regions(regions_bed = regions_bed)
full_details_maf = get_ssm_by_regions(regions_bed = regions_bed,basic_columns=T)
}
