% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_sample}
\alias{get_ssm_by_sample}
\title{Get the ssms (i.e. load MAF) for a single sample. This was implemented to allow flexibility because
there are some samples that we may want to use a different set of variants than those in the main GAMBL merge.
The current use case is to allow a force_unmatched output to be used to replace the SSMs from the merge for samples
with known contamination in the normal. This will also be useful to apply a blacklist to individual MAFs when coupled with
annotate_ssm_blacklist}
\usage{
get_ssm_by_sample(
  this_sample_id,
  these_samples_metadata,
  tool_name = "slms-3",
  projection = "grch37",
  seq_type = "genome",
  flavour = "",
  verbose = FALSE
)
}
\arguments{
\item{this_sample_id}{Required. The sample_id you want the data from.}

\item{tool_name}{The name of the variant calling pipeline (currently only slms-3 is supported)}

\item{projection}{The projection genome build. Currently only grch37 is supported but hg38 should be easy to add.}

\item{seq_type}{What type of sequencing data you want mutations from (e.g. genome, exome, mrna)}

\item{force_unmatched}{Defaults to TRUE and only supports this option curently (added option for anticipated future flexibility)}
}
\value{
data frame in MAF format
}
\description{
Get the ssms (i.e. load MAF) for a single sample. This was implemented to allow flexibility because
there are some samples that we may want to use a different set of variants than those in the main GAMBL merge.
The current use case is to allow a force_unmatched output to be used to replace the SSMs from the merge for samples
with known contamination in the normal. This will also be useful to apply a blacklist to individual MAFs when coupled with
annotate_ssm_blacklist
}
\examples{

}
