% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_sample}
\alias{get_ssm_by_sample}
\title{Get the ssms (i.e. load MAF) for a single sample. This was implemented to allow flexibility because
there are some samples that we may want to use a different set of variants than those in the main GAMBL merge.
The current use case is to allow a force_unmatched output to be used to replace the SSMs from the merge for samples
with known contamination in the normal. This will also be useful to apply a blacklist to individual MAFs when coupled with
annotate_ssm_blacklist.}
\usage{
get_ssm_by_sample(
  this_sample_id,
  these_samples_metadata,
  tool_name = "slms-3",
  projection = "grch37",
  these_genes,
  augmented = TRUE,
  flavour = "clustered",
  min_read_support = 3,
  basic_columns = TRUE,
  maf_cols = NULL,
  return_cols = FALSE,
  verbose = FALSE,
  ssh_session
)
}
\arguments{
\item{this_sample_id}{Required. The sample_id you want the data from.}

\item{these_samples_metadata}{Either a single row or entire metadata table containing your sample_id}

\item{tool_name}{The name of the variant calling pipeline (currently only slms-3 is supported)}

\item{projection}{The projection genome build. Supports hg38 and grch37.}

\item{these_genes}{A vector of genes to subset ssm to.}

\item{augmented}{default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead of the augmented MAF}

\item{flavour}{Currently this function only supports one flavour option but this feature is meant for eventual compatability with additional variant calling parameters/versions}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs)}

\item{basic_columns}{Return first 43 columns of MAF rather than full details. Default is TRUE.}

\item{maf_cols}{if basic_columns is set to FALSE, the suer can specify what columns to be returned within the MAF. This parameter can either be a list of indexes (integer) or a list of characters.}

\item{return_cols}{If set to TRUE, a vector with all avaialble column names will be returned. Default is FALSE.}

\item{verbose}{Enable for debugging/noisier output}

\item{ssh_session}{BETA feature! pass active ssh session object.
If specified, the function will assume the user is not on the network and will temporarily copy the file locally.}
}
\value{
data frame in MAF format.
}
\description{
Get the ssms (i.e. load MAF) for a single sample. This was implemented to allow flexibility because
there are some samples that we may want to use a different set of variants than those in the main GAMBL merge.
The current use case is to allow a force_unmatched output to be used to replace the SSMs from the merge for samples
with known contamination in the normal. This will also be useful to apply a blacklist to individual MAFs when coupled with
annotate_ssm_blacklist.
}
\examples{
ssm_sample = get_ssm_by_sample(this_sample_id = "HTMCP-01-06-00485-01A-01D", tool_name = "slims-3", projection = "grch37")

}
