% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_samples}
\alias{get_ssm_by_samples}
\title{Get MAF-format data frame for more than one sample and combine together (wraps get_ssm_by_sample)
See get_ssm_by_sample for more information}
\usage{
get_ssm_by_samples(
  these_sample_ids,
  these_samples_metadata,
  tool_name = "slms-3",
  projection = "grch37",
  seq_type = "genome",
  flavour = ""
)
}
\arguments{
\item{these_sample_ids}{A vector of sample_id that you want results for. This is the only required argument.}

\item{flavour}{Specify this as either "augmented" or "force_unmatched" or leave out to get the default (original MAF)}
}
\value{

}
\description{
Get MAF-format data frame for more than one sample and combine together (wraps get_ssm_by_sample)
See get_ssm_by_sample for more information
}
\examples{
merged_maf_force_unmatched = get_ssm_by_samples(these_sample_ids=c("HTMCP-01-06-00485-01A-01D","14-35472_tumorA","14-35472_tumorB"))
}
