% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_ssm_by_samples}
\alias{get_ssm_by_samples}
\title{Get MAF-format data frame for more than one sample and combine together (wraps get_ssm_by_sample)
See get_ssm_by_sample for more information}
\usage{
get_ssm_by_samples(
  these_sample_ids,
  these_samples_metadata,
  tool_name = "slms-3",
  projection = "grch37",
  seq_type = "genome",
  flavour = "clustered",
  these_genes,
  min_read_support = 3,
  basic_columns = TRUE,
  maf_cols = NULL,
  return_cols = FALSE,
  subset_from_merge = TRUE,
  augmented = TRUE,
  ssh_session
)
}
\arguments{
\item{these_sample_ids}{A vector of sample_id that you want results for. This is the only required argument.}

\item{these_samples_metadata}{Optional metadata table}

\item{tool_name}{Only supports slms-3 currently}

\item{projection}{Obtain variants projected to this reference (one of grch37 or hg38)}

\item{flavour}{Currently this function only supports one flavour option but this feature is meant for eventual compatability with additional variant calling parameters/versions}

\item{these_genes}{A vector of genes to subset ssm to.}

\item{min_read_support}{Only returns variants with at least this many reads in t_alt_count (for cleaning up augmented MAFs)}

\item{basic_columns}{Return first 45 columns of MAF rather than full details. Default is TRUE.}

\item{maf_cols}{if basic_columns is set to FALSE, the suer can specify what columns to be returned within the MAF. This parameter can either be a list of indexes (integer) or a list of characters.}

\item{return_cols}{If set to TRUE, a vector with all avaialble column names will be returned. Default is FALSE.}

\item{subset_from_merge}{Instead of merging individual MAFs, the data will be subset from a pre-merged MAF of samples with the specified seq_type}

\item{augmented}{default: TRUE. Set to FALSE if you instead want the original MAF from each sample for multi-sample patients instead}

\item{BETA}{optional argument to supply active ssh session connection for remote transfers}
}
\value{
data frame in MAF format.
}
\description{
Get MAF-format data frame for more than one sample and combine together (wraps get_ssm_by_sample)
See get_ssm_by_sample for more information
}
\examples{
merged_maf_force_unmatched = get_ssm_by_samples(these_sample_ids=c("HTMCP-01-06-00485-01A-01D","14-35472_tumorA","14-35472_tumorB"))

}
