% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{maf_to_custom_track}
\alias{maf_to_custom_track}
\title{Convert a maf-formatted data frame into a bed custom track file for UCSC.}
\usage{
maf_to_custom_track(
  maf_data,
  these_samples_metadata,
  seq_type = "genome",
  output_file,
  as_bigbed = FALSE,
  colour_column = "lymphgen",
  as_biglolly = FALSE,
  track_name = "GAMBL mutations",
  track_description = "mutations from GAMBL",
  verbose = FALSE,
  padding_size = 0
)
}
\arguments{
\item{maf_data}{Either a maf loaded from disk or from the database using a get_ssm function.}

\item{output_file}{Name for your new bed file that can be uploaded as a custom track to UCSC.}
}
\value{
Nothing.
}
\description{
Convert a maf-formatted data frame into a bed custom track file for UCSC.
}
\examples{
maf_to_custom_track(my_maf_data, "/home/rmorin/private/some_mutations.bed")

}
