% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_igv_snapshot}
\alias{make_igv_snapshot}
\title{Load bams and generate an IGV screenshot for one or more regions}
\usage{
make_igv_snapshot(
  bams,
  genome_build,
  region,
  padding = 200,
  chrom,
  start,
  end,
  sample_id,
  out_path = "/tmp/",
  igv_port = 60506
)
}
\arguments{
\item{bams}{Character vector containing the full path to one or more bam files}

\item{genome_build}{String specifying the genome build for the bam files provided}

\item{region}{Optionally specify the region as a single string (e.g. "chr1:1234-1235")}

\item{padding}{Optionally specify a positive value to broaden the region around the specified position}

\item{chrom}{Optionally specify the region by specifying the chromosome, start and end (see below)}

\item{start}{Optionally specify the region by specifying the start}

\item{end}{Optionally specify the region by specifying the end}

\item{sample_id}{Specify the sample_id or any other string you want embedded in the file name}

\item{out_path}{Specify the output directory where the snapshot will be written}

\item{igv_port}{Specify the port IGV is listening on}

\item{region_bed}{Optionally specify regions in bed format (column order is assumed)}
}
\value{

}
\description{
Load bams and generate an IGV screenshot for one or more regions
}
\examples{
#IMPORTANT: you must be running IGV on the host that is running R and you need to have it listening on a port
# The simplest scenario is to run this command on a terminal (if using a Mac), assuming you are using R on gphost10 and you have a ssh config that routes gp10 to that host
# ssh -X gp10
# then launch IGV (e.e. from a conda installation):
# conda activate igv; igv &
this_sv = annotated_sv \%>\% filter(gene=="ETV6")
tumour_bam = get_bam_path(sample=this_sv$tumour_sample_id)
make_igv_snapshot(chrom=this_sv$chrom2, start=this_sv$start2, end=this_sv$end2, sample_id=this_sv$tumour_sample_id,out_path="~/IGV_snapshots/")
}
