% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{map_metadata_to_colours}
\alias{map_metadata_to_colours}
\title{Assign a colour palette to metadata columns automatically and consistently}
\usage{
map_metadata_to_colours(
  metadataColumns,
  these_samples_metadata,
  as_vector = TRUE,
  verbose = F,
  annoAlpha = 1
)
}
\arguments{
\item{metadataColumns}{Names of the metadata columns to assign colours for}

\item{these_samples_metadata}{Metadata for just the samples you need colours for}

\item{annoAlpha}{Optional alpha to apply to annotation colours}
}
\value{
Either a vector or list of colours
}
\description{
Assign a colour palette to metadata columns automatically and consistently
}
\examples{
all_cols=map_metadata_to_colours(legend_metadata_columns,these_meta,verbose=T)
}
