% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{plot_multi_timepoint}
\alias{plot_multi_timepoint}
\title{Title}
\usage{
plot_multi_timepoint(mafs, sample_id, genes, show_noncoding = FALSE, detail)
}
\arguments{
\item{detail}{}
}
\value{

}
\description{
Title
}
