% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_multi_timepoint}
\alias{plot_multi_timepoint}
\title{This function doesn't do anything yet, thus, I ahve removed it from being exported to NAMESPACE (for now).}
\usage{
plot_multi_timepoint(
  mafs,
  this_sample_id,
  genes,
  show_noncoding = FALSE,
  detail
)
}
\arguments{
\item{mafs}{TODO}

\item{this_sample_id}{TODO}

\item{genes}{TODO}

\item{show_noncoding}{TODO}

\item{detail}{TODO}
}
\description{
This function doesn't do anything yet, thus, I ahve removed it from being exported to NAMESPACE (for now).
}
