% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_multi_timepoint}
\alias{plot_multi_timepoint}
\title{This function doesn't do anything yet}
\usage{
plot_multi_timepoint(mafs, sample_id, genes, show_noncoding = FALSE, detail)
}
\arguments{
\item{detail}{}
}
\value{

}
\description{
This function doesn't do anything yet
}
