% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{populate_each_tool_result}
\alias{populate_each_tool_result}
\title{Populate Each Tool Results.}
\usage{
populate_each_tool_result(
  tool,
  genome_builds,
  unix_groups,
  include_silent = FALSE
)
}
\arguments{
\item{tool}{Name of tool to get results from.}

\item{include_silent}{Logical parameter indicating whether to include silent mutations into coding mutations. Default is FALSE.}

\item{genome_build}{A single genome build or a vector of all genome builds to process.}

\item{unix_group}{A single unix group or a vector of all unix groups to process.}
}
\value{
Nothing.
}
\description{
Convenience function for returning results from a specified tool.
}
\details{
This function takes a tool name \code{tool} as well as other parameters for specifying the requested result.
Other parameters include \code{genome_build}, this can be just one parameter or a vector with different genome builds to return results for.
Similarly, \code{unix_group} can take either one value or a vector with all the different unix groups to return results for.
Lastly, the user can subset the returned results to only silent mutations.
This is done with setting \code{include_silent = TRUE} (default is FALSE).
}
\examples{
tool_results = populate_each_tool_result("smlims_3", "grch37", "gambl", FALSE)

}
