% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{populate_each_tool_result}
\alias{populate_each_tool_result}
\title{Title}
\usage{
populate_each_tool_result(
  tool,
  genome_builds,
  unix_groups,
  include_silent = FALSE
)
}
\arguments{
\item{tool}{Name of tool to get results from}

\item{include_silent}{Logical parameter indicating whether to include siment mutations into coding mutations. Default is FALSE}

\item{genome_build}{A list of all genome builds to process}

\item{unix_group}{A list of all unix groups to process}
}
\description{
Title
}
\examples{
tool_results = populate_each_tool_result("smlims_3", "grch37", "gambl", FALSE)

}
