% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{populate_tool_results}
\alias{populate_tool_results}
\title{Title}
\usage{
populate_tool_results(
  database_name = "gambl_test",
  sample_table,
  tool = "battenberg",
  base_directory_gambl,
  base_directory_other
)
}
\arguments{
\item{base_directory_other}{}
}
\value{

}
\description{
Title
}
