% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{populate_tool_results}
\alias{populate_tool_results}
\title{Populate Tool Results.}
\usage{
populate_tool_results(tool_name)
}
\arguments{
\item{tool_name}{Name of the tool to get the results for.}
}
\value{
Nothing.
}
\description{
Populate the database with the per-sample summarized results of various tools.
}
\details{
this function is still in draft mode, export to NAMESPACE has been removed for now.
}
\examples{
\dontrun{
results = populate_tool_results(tool_name = "slims_3")
}

}
