% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{pretty_lollipop_plot}
\alias{pretty_lollipop_plot}
\title{Generate a more visually appealing and flexible lollipop plot.}
\usage{
pretty_lollipop_plot(maf_df, gene, plot_title, plot_theme = "cbioportal")
}
\arguments{
\item{maf_df}{A data frame containing the mutation data (from a MAF).}

\item{gene}{The gene symbol to plot.}

\item{plot_title}{Optional (defaults to gene name).}

\item{plot_theme}{Options: cbioportal(default), blue, simple, nature, nature2, ggplot2, and dark.}
}
\value{
Nothing
}
\description{
Generate a more visually appealing and flexible lollipop plot.
}
\examples{
pretty_lollipop_plot = (mutation_df, "MYC", "Mutation data for MYC", "blue")
pretty_lollipop_plot = (mutation_df, "BCL2")

}
