% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{pretty_lollipop_plot}
\alias{pretty_lollipop_plot}
\title{Lollipop Plot}
\usage{
pretty_lollipop_plot(
  maf_df,
  gene,
  plot_title,
  plot_theme = "cbioportal",
  out_name = paste0("my_lollipop_plot_", gene)
)
}
\arguments{
\item{maf_df}{A data frame containing the mutation data (from a MAF).}

\item{gene}{The gene symbol to plot.}

\item{plot_title}{Optional (defaults to gene name).}

\item{plot_theme}{Options: cbioportal(default), blue, simple, nature, nature2, ggplot2, and dark.}

\item{out_name}{Optional, set the file name of the plot, if you export it to disk. Default name is my_lollipop_plot_{gene}.}
}
\value{
Nothing.
}
\description{
Generates a visually appealing and interactive lollipop plot.
}
\details{
This function is depending on a modified version of \code{readMAF} from the \code{g3viz} package.
Returned plot is interactive, meaning the user can hover over individual points in the plot to reveal more information.
The plot can also be exported in a variety of file formats inside the interactive view of the lollipop plot.
}
\examples{
#get metadata (Fl and DLBCL)
metadata = get_gambl_metadata()
this_metadata = dplyr::filter(metadata, consensus_pathology \%in\% c("FL", "DLBCL"))

#get maf data for returned samples
maf = get_coding_ssm(limit_samples = this_metadata$sample_id,
                     seq_type = "genome")

#construct pretty_lollipop_plot.
pretty_lollipop_plot(maf_df = maf,
                     gene = "MYC",
                     plot_title = "Mutation data for MYC",
                     plot_theme = "nature2")

}
