% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyChromoplot}
\alias{prettyChromoplot}
\title{Use GISTIC2.0 scores output to reproduce maftools::chromoplot with more flexibility}
\usage{
prettyChromoplot(scores, genes_to_label, cutoff = 0.5)
}
\arguments{
\item{genes_to_label}{optional. Provide a data frame of genes to label (if mutated). The first 3 columns must contain chromosome, start, and end coordinates. Another required column must contain gene names and be named \code{gene}. All other columns are ignored. If no data frame provided, oncogenes from GAMBLR packages are used by default to annotate on the plot.}

\item{cutoff}{optional. Used to determine which regions to color as aberrant. Must be float in the range \link{0-1}. The higher the number, the less regions will be considered as aberrant. The default is 0.5.}
}
\value{
nothing
}
\description{
Use GISTIC2.0 scores output to reproduce maftools::chromoplot with more flexibility
}
\examples{
# basic usage
prettyChromoplot("path_to_gistic_results/scores.gistic")
# advanced usages
prettyChromoplot("path_to_gistic_results/scores.gistic", genes_to_label="path_to_gene_coordinates_table.tsv", cutoff=0.75) +
...(any ggplot options to customize plot appearance)
}
