% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyOncoplot}
\alias{prettyOncoplot}
\title{Make an oncoplot that is pretty using ComplexHeatmap. The metadata is expected to follow the structure and column naming used in GAMBL.
If you provide your own non-GAMBL samples and metadata, you must include at least the following columns with these names.
The first one should match the Tumor_Sample_Barcode in the MAF object or onco_matrix you provide.
sample_id, pathology}
\usage{
prettyOncoplot(
  maftools_obj,
  onco_matrix_path,
  genes,
  keepGeneOrder = TRUE,
  keepSampleOrder = TRUE,
  highlightHotspots = FALSE,
  these_samples_metadata,
  metadataColumns,
  numericMetadataColumns,
  expressionColumns = c(),
  numericMetadataMax,
  sortByColumns,
  removeNonMutated = FALSE,
  minMutationPercent,
  fontSizeGene = 6,
  annoAlpha = 1,
  mutAlpha = 1,
  recycleOncomatrix = FALSE,
  box_col = NA,
  metadataBarHeight = 1.5,
  metadataBarFontsize = 5,
  hideTopBarplot = FALSE,
  hideSideBarplot = FALSE,
  legend_row = 3,
  legend_col = 3
)
}
\arguments{
\item{maftools_obj}{A maftools object containing the mutations you want to plot}

\item{onco_matrix_path}{Provide a path to an onco_matrix file instead of a MAF object if the former is unavailable (this limits functionality a bit)}

\item{genes}{An optional list of genes to restrict your plot to}

\item{keepGeneOrder}{Set to TRUE if you want to preserve the gene order specified}

\item{keepSampleOrder}{Set to TRUE if you want to preserve the sample order specified}

\item{these_samples_metadata}{Data frame containing metadata for your samples}

\item{metadataColumns}{A vector containing the categorical column names you want to plot below}

\item{numericMetadataColumns}{A vector containing the numeric columns you want to plot below}

\item{numericMetadataMax}{A numeric vector of cutoffs to apply to numeric columns above}

\item{sortByColumns}{A vector containing the column names you want to sort columns (patients) on}

\item{removeNonMutated}{Set to TRUE to drop unmutated cases}

\item{minMutationPercent}{Only genes mutated in more than minMutationPercent \% patients will be included}

\item{recycleOncomatrix}{Set to TRUE most of the time to reuse the oncomatrix saved by maftools}

\item{legend_row}{Fiddle with these to widen or narrow your legend}

\item{legend_col}{Fiddle with these to widen or narrow your legend}
}
\value{

}
\description{
Make an oncoplot that is pretty using ComplexHeatmap. The metadata is expected to follow the structure and column naming used in GAMBL.
If you provide your own non-GAMBL samples and metadata, you must include at least the following columns with these names.
The first one should match the Tumor_Sample_Barcode in the MAF object or onco_matrix you provide.
sample_id, pathology
}
