% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyOncoplot}
\alias{prettyOncoplot}
\title{Make an oncoplot that is pretty using ComplexHeatmap. The metadata is expected to follow the structure and column naming used in GAMBL.
If you provide your own non-GAMBL samples and metadata, you must include at least the following columns with these names.
The first one should match the Tumor_Sample_Barcode in the MAF object or onco_matrix you provide.
sample_id, pathology}
\usage{
prettyOncoplot(
  maftools_obj,
  onco_matrix_path,
  genes,
  include_noncoding = list(NFKBIZ = c("3'UTR"), HNRNPH1 = "Splice_Region"),
  keepGeneOrder = TRUE,
  keepSampleOrder = TRUE,
  highlightHotspots = FALSE,
  these_samples_metadata,
  metadataColumns,
  numericMetadataColumns,
  expressionColumns = c(),
  numericMetadataMax,
  sortByColumns,
  removeNonMutated = FALSE,
  minMutationPercent,
  fontSizeGene = 6,
  annoAlpha = 1,
  mutAlpha = 1,
  recycleOncomatrix = FALSE,
  box_col = NA,
  metadataBarHeight = 1.5,
  metadataBarFontsize = 5,
  hideTopBarplot = FALSE,
  hideSideBarplot = FALSE,
  splitColumnName,
  splitGeneGroups,
  legend_row = 3,
  legend_col = 3,
  showTumorSampleBarcode = FALSE,
  groupNames,
  verbose = FALSE,
  hide_annotations,
  custom_colours = NULL
)
}
\arguments{
\item{maftools_obj}{A maftools object containing the mutations you want to plot}

\item{onco_matrix_path}{Provide a path to an onco_matrix file instead of a MAF object if the former is unavailable (this limits functionality a bit)}

\item{genes}{An optional list of genes to restrict your plot to}

\item{keepGeneOrder}{Set to TRUE if you want to preserve the gene order specified}

\item{keepSampleOrder}{Set to TRUE if you want to preserve the sample order specified}

\item{these_samples_metadata}{Data frame containing metadata for your samples}

\item{metadataColumns}{A vector containing the categorical column names you want to plot below}

\item{numericMetadataColumns}{A vector containing the numeric columns you want to plot below}

\item{numericMetadataMax}{A numeric vector of cutoffs to apply to numeric columns above}

\item{sortByColumns}{A vector containing the column names you want to sort columns (patients) on}

\item{removeNonMutated}{Set to TRUE to drop unmutated cases}

\item{minMutationPercent}{Only genes mutated in more than minMutationPercent \% patients will be included}

\item{fontSizeGene}{Font size for gene labels (default 6)}

\item{annoAlpha}{Optional alpha to apply to annotation colours}

\item{mutAlpha}{Optional alpha to apply to mutation colours}

\item{recycleOncomatrix}{Set to TRUE most of the time to reuse the oncomatrix saved by maftools}

\item{box_col}{Colour of boxes for outlining mutations (can be problematic with larger oncoprints)}

\item{legend_row}{Fiddle with these to widen or narrow your legend}

\item{legend_col}{Fiddle with these to widen or narrow your legend}

\item{custom_colours}{Provide named vector (or named list of vectors) containing custom annotation colours if you do not want to use standartized pallette}
}
\value{

}
\description{
Make an oncoplot that is pretty using ComplexHeatmap. The metadata is expected to follow the structure and column naming used in GAMBL.
If you provide your own non-GAMBL samples and metadata, you must include at least the following columns with these names.
The first one should match the Tumor_Sample_Barcode in the MAF object or onco_matrix you provide.
sample_id, pathology
}
\examples{
prettyOncoplot(maftools_obj = maf_obj,genes = bl_genes,
these_samples_metadata = extra_meta,
metadataColumns = c("pathology","COO_consensus",
                    "cluster_name", "lymphgen","EBV_status_inf",
                    "manta_BCL6_sv"),
hide_annotations = c(some_ashm,"lymphgen","COO_consensus",
                     "pathology","manta_BCL6_sv"),
expressionColumns = c("IRF4",some_ashm),
sortByColumns = c("IRF4"),
keepGeneOrder = FALSE,splitGeneGroups = groups,
splitColumnName = "cluster_name",
metadataBarHeight = 2.5,metadataBarFontsize = 6,fontSizeGene = 8,
recycleOncomatrix = TRUE,removeNonMutated = FALSE)
}
