% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{prettyRainfallPlot}
\alias{prettyRainfallPlot}
\title{Plot a rainfall plot for one sample. This function takes in MAF data frame, or path to custom MAF file.
If non are specified, the SSM will be obtained though GAMBLR directly.}
\usage{
prettyRainfallPlot(
  this_sample_id,
  label_ashm_genes = TRUE,
  projection = "grch37",
  chromosome,
  this_maf,
  maf_path,
  zoom_in_region,
  label_sv = FALSE
)
}
\arguments{
\item{this_sample_id}{Sample id for the sample to display. This is the only required argument.}

\item{label_ashm_genes}{Boolean argument indicating whether the aSHM regions will be labeled or not.}

\item{projection}{Specify projection (grch37 or hg38) of mutations. Default is grch37.}

\item{chromosome}{Provide one or more chromosomes to plot. The chr prefix can be inconsistent with projection and will be handled.}

\item{this_maf}{Specify custom MAF data frame of mutations.}

\item{maf_path}{Specify path to MAF file if it is not already loaded into data frame.}

\item{zoom_in_region}{Provide a specific region in the format "chromosome:start-end" to zoom in to a specific region.}

\item{label_sv}{Boolean argument to specify whether label SVs or not. Only supported if a specific chromosome or zoom in region are specified.}
}
\value{
a ggplot2 plot. Print it using print() or save it using ggsave()
}
\description{
Plot a rainfall plot for one sample. This function takes in MAF data frame, or path to custom MAF file.
If non are specified, the SSM will be obtained though GAMBLR directly.
}
\examples{
prettyRainfallPlot("Raji")
prettyRainfallPlot("Raji", chromosome = c(3,9,"chr14",22,"X"))
prettyRainfallPlot("Raji", chromosome = c(3,9), projection = "hg38", label_ashm_genes = FALSE)
prettyRainfallPlot("Raji", zoom_in_region = "8:125252796-135253201", label_sv = TRUE)
prettyRainfallPlot("Raji", chromosome = 6, label_sv = TRUE)

}
