% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{process_all_manta_bedpe}
\alias{process_all_manta_bedpe}
\title{Process All Manta Bedpe.}
\usage{
process_all_manta_bedpe(
  file_df,
  out_dir,
  group,
  genome_build,
  projection_build = "grch37"
)
}
\arguments{
\item{file_df}{Paths to bedpe.}

\item{out_dir}{output directory.}

\item{group}{The unix group.}

\item{genome_build}{Genome build.}

\item{projection_build}{The genome we want all results to be relative to (lifted if necessary).}
}
\description{
This function is in draft mode.
}
\details{
This is a helper function that is not meant to be used routinely.
}
