% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{read_merge_manta_with_liftover}
\alias{read_merge_manta_with_liftover}
\title{This is a helper function that is not meant to be used routinely}
\usage{
read_merge_manta_with_liftover(
  bedpe_paths = c(),
  pattern = "--matched",
  out_dir
)
}
\arguments{
\item{bedpe_paths}{path to bedpe}

\item{pattern}{pattern}

\item{out_dir}{output directory}
}
\description{
This is a helper function that is not meant to be used routinely
}
