% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_merge_manta_with_liftover}
\alias{read_merge_manta_with_liftover}
\title{This is a helper function that is not meant to be used routinely}
\usage{
read_merge_manta_with_liftover(
  bedpe_paths = c(),
  pattern = "--matched",
  out_dir
)
}
\arguments{
\item{out_dir}{}
}
\value{

}
\description{
This is a helper function that is not meant to be used routinely
}
