% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_io.R
\name{read_merge_manta_with_liftover}
\alias{read_merge_manta_with_liftover}
\title{Read Merge Manta With Liftover.}
\usage{
read_merge_manta_with_liftover(
  bedpe_paths = c(),
  pattern = "--matched",
  out_dir
)
}
\arguments{
\item{bedpe_paths}{path to bedpe}

\item{pattern}{pattern}

\item{out_dir}{output directory}
}
\description{
Takes a path to bedpe and runs liftover (\code{liftover_bedpe}) based on the original genome build of the bedpe.
}
\details{
This is a helper function that is not meant to be used routinely.
}
\examples{
manta_bedpe = read_merge_manta_with_liftover(bedpe_paths = "some_path.bedpe",
                                             out_dir = "../")

}
