% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{referesh_metadata_tables}
\alias{referesh_metadata_tables}
\title{INTERNAL FUNCTION, not meant for out-of-package usage.
Refresh the contents of meta data table.}
\usage{
referesh_metadata_tables()
}
\value{
Table.
}
\description{
INTERNAL FUNCTION, not meant for out-of-package usage.
Refresh the contents of meta data table.
}
\examples{
ref_meta = referesh_metadata_tables()

}
