% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{region_to_gene}
\alias{region_to_gene}
\title{Region To Gene.}
\usage{
region_to_gene(region, gene_format = "hugo", genome_build = "grch37")
}
\arguments{
\item{region}{Regions to intersect genes with, this can be either a data frame with regions sorted in the following columns; chromosome, start, end. Or it can be a character vector in "region" format, i.e chromosome:start-end.}

\item{gene_format}{Parameter for specifying the format of returned genes, default is "hugo", other acceptable inputs are "ensembl".}

\item{genome_build}{Reference genome build.}
}
\value{
A data frame with gene(s) that are residing in the specified region(s).
}
\description{
Return genes residing in defined region(s).
}
\details{
This function takes a region as a vector of characters, or a data frame with of regions (e.g output from \code{gene_to_region(return_as="df")}).
and returns the genes residing withing the specified region. For the other way around (i.e gene to regions, refer to \code{gene_to_regions}).
}
\examples{
myc_region = gene_to_region(gene_symbol = "MYC", genome_build = "grch37", return_as = "df")
region = region_to_gene(region = myc_region, gene_format = "hugo", genome_build = "grch37")

}
