% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{review_hotspots}
\alias{review_hotspots}
\title{Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations.}
\usage{
review_hotspots(
  annotated_maf,
  genes_of_interest = c("FOXO1", "MYD88", "CREBBP", "NOTCH1", "NOTCH2", "CD79B",
    "EZH2"),
  genome_build = "hg19"
)
}
\arguments{
\item{annotated_maf}{A data frame in MAF format that has hotspots annotated using function annotate_hotspots().}

\item{genes_of_interest}{List of genes for hotspot review. Currently only FOXO1, MYD88, and CREBBP are supported.}

\item{genome_build}{Reference genome build for the coordinates in the MAF file. The default is hg19 genome build.}
}
\value{
The same data frame with reviewed column "hot_spot".
}
\description{
Annotate MAF-like data frome with a hot_spot column indicating recurrent mutations.
}
\examples{
hot_ssms = review_hotspots(annotate_hotspots(get_coding_ssm()), genes_of_interest = c("CREBBP"))
}
