% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sanity_check_metadata}
\alias{sanity_check_metadata}
\title{Function that performes sanity checks on meta data.}
\usage{
sanity_check_metadata()
}
\value{
Table.
}
\description{
Function that performes sanity checks on meta data.
}
\examples{
sane_meta_data = sanity_check_metadata()

}
