% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{setup_fusions}
\alias{setup_fusions}
\title{Setup Fusions (cBioPortal).}
\usage{
setup_fusions(
  short_name = "GAMBL",
  human_friendly_name = "GAMBL data",
  project_name = "gambl_genome",
  description = "GAMBL data from genome",
  gambl_maf = "maf_slms3_hg19",
  gambl_icgc_maf = "maf_slms3_hg19_icgc",
  out_dir
)
}
\arguments{
\item{short_name}{A concise name for your portal project.}

\item{human_friendly_name}{A slightly more verbose name for your project.}

\item{project_name}{Unique ID for your project.}

\item{description}{A verbose description of your data set.}

\item{gambl_maf}{maf origin.}

\item{gambl_icgc_maf}{ICGC maf origin.}

\item{out_dir}{The full path to the base directory where the files are being created.}
}
\value{
A vector of sample_id for the patients that have been included.
}
\description{
Annotate SVs and create the input for fusions to be displayed in cBioPortal instance.
}
\details{
This function calls \code{get_combined_sv} and runs \code{annotate_sv} on the returned data frame.
Should be run as the next step after running \code{setup_study}. Note that the parameters called with this function
has to match the parameter arguments of \code{setup_study}, i.e if \code{short_name} is for \code{setup_study} is "GAMBL",
then the \code{short_name} in \code{setup_fusions} also has to be "GAMBL", etc.
}
\examples{
\dontrun{
fusion_ids = setup_fusions(out_dir = "GAMBLR/cBioPortal/instance01/")
}
}
