% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{setup_study}
\alias{setup_study}
\title{Initialize a new cBioPortal instance or update existing portal data set}
\usage{
setup_study(
  short_name = "GAMBL",
  include_icgc_data = FALSE,
  human_friendly_name = "GAMBL data",
  project_name = "gambl_minus_icgc",
  description = "GAMBL data without ICGC",
  cancer_type = "mixed",
  out_dir,
  overwrite = FALSE,
  exclude_cohorts = c("FFPE_Benchmarking")
)
}
\arguments{
\item{short_name}{a concise name for your portal project}

\item{include_icgc_data}{whether or not you want ICGC and other external data included}

\item{human_friendly_name}{a slightly more verbose name for your project}

\item{project_name}{unique ID for your project}

\item{description}{A verbose description of your data set}

\item{out_dir}{The full path to the base directory where the files are being created}

\item{overwrite}{Flag to specify that files should be overwritten if they exist}
}
\value{
a vector of sample_id for the patients that have been included
}
\description{
Initialize a new cBioPortal instance or update existing portal data set
}
