% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal.R
\name{study_check}
\alias{study_check}
\title{Helper function for checking integrity of study files.}
\usage{
study_check(
  data_clinical_samples_path = "data_clinical_samples.txt",
  data_fusions_path = "data_fusions.txt",
  cases_fusions_path = "case_lists/cases_fusion.txt",
  cases_all_path = "case_lists/cases_all.txt",
  cases_sequenced_path = "case_lists/cases_sequenced.txt",
  project_name = "gambl_minus_icgc",
  out_dir
)
}
\arguments{
\item{project_name}{Project name, should match what is specified under setup_study/setup_fusions.}

\item{out_dir}{Directory with all study related files, the only argument that needs to be specified, given that path to all generated study files are not changed from default.}

\item{data_clinical_samples}{Path to clinical file.}

\item{data_fusions}{Path to data_fusion file from setup_fusions.}

\item{cases_fusions}{Path to cases_fusion from setup_fusions.}

\item{cases_all}{Path to cases_all from setup_study.}

\item{cases_sequenced}{Path to cases_sequenced from setup_study.}
}
\value{
Nothing.
}
\description{
Helper function for checking integrity of study files.
}
\examples{
samples_not_in_clinical = study_check(out_dir = "GAMBLR/cBioPortal/instance01/")

}
