% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trim_scale_expression}
\alias{trim_scale_expression}
\title{INTERNAL FUNCTION called by prettyOncoplot, not meant for out-of-package usage.
Unsure what this function actually does, old helper function that is outdated?}
\usage{
trim_scale_expression(x)
}
\arguments{
\item{x}{}
}
\value{
Numeric value.
}
\description{
INTERNAL FUNCTION called by prettyOncoplot, not meant for out-of-package usage.
Unsure what this function actually does, old helper function that is outdated?
}
\examples{
trimmed = trim_scale_expression(2)

}
