% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web.R
\name{web_initialize_gambl_site}
\alias{web_initialize_gambl_site}
\title{Set up a fresh instance of a website to host on gitlab}
\usage{
web_initialize_gambl_site(
  site_base_name,
  base_directory = "/home/rmorin/",
  my_name = "Ryan Morin",
  my_gitlab_email = "rdmorin@sfu.ca"
)
}
\arguments{
\item{base_directory}{}
}
\value{

}
\description{
Set up a fresh instance of a website to host on gitlab
}
