% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_sample_set_hash}
\alias{write_sample_set_hash}
\title{Write Sample Set Hash}
\usage{
write_sample_set_hash(update = TRUE, new_sample_sets_df)
}
\arguments{
\item{update}{Leave as TRUE for default functionality (i.e. updating the existing table). If the table doesn't exist you probably need to pull from Master.}

\item{new_sample_sets_df}{Data frame of all existing and new sample sets. Required when running in default update mode.}
}
\value{

}
\description{
Update or create a file to track unique identifiers for sample sets in GAMBL
}
\details{
Run this function with \code{update = TRUE} (default) to use an existing sample table.
If this table does not exist, perhaps you need to pull from the master branch.
If this function is run with the default for \code{update}, the user must also provide the new sample sets with the \code{new_sample_sets_df}.
}
